### 2024-05-03 #6719 @editorialbot
[PRE REVIEW]: A Module for Calibrating Impact Functions in the Climate Risk Modeling Platform CLIMADA
**Submitting author:** @peanutfun (<a href="http://orcid.org/0000-0002-4667-3652">Lukas Riedel</a>)
**Repository:** https://github.com/CLIMADA-project/climada_python
**Branch with paper.md** (empty if default branch): calibrate-impact-functions
**Version:** v5
