### 2024-05-03 40779e9
Rename variables, Add exception, Improve Docstring
