### 2024-05-03 c2664c0
Improve Test style and Add test for error
