### 2024-05-03 c3902f8
add docstring for new variable
