### 2024-05-03 c754345
update license statement and better error raise
