### 2024-05-03 a6707f6
update gmst info dictionary docstring
