+++ doc/joss/calibration-module/paper.bib
@@ -80,7 +80,7 @@
 	volume = {21},
 	issn = {1561-8633},
 	url = {https://nhess.copernicus.org/articles/21/393/2021/},
-	doi = {https://doi.org/10.5194/nhess-21-393-2021},
+	doi = {10.5194/nhess-21-393-2021},
 	abstract = {Assessing the adverse impacts caused by tropical cyclones has become increasingly important as both climate change and human coastal development increase the damage potential. In order to assess tropical cyclone risk, direct economic damage is frequently modeled based on hazard intensity, asset exposure, and vulnerability, the latter represented by impact functions. In this study, we show that assessing tropical cyclone risk on a global level with one single impact function calibrated for the USA {\textendash} which is a typical approach in many recent studies {\textendash} is problematic, biasing the simulated damage by as much as a factor of 36 in the north West Pacific. Thus, tropical cyclone risk assessments should always consider regional differences in vulnerability, too. This study proposes a calibrated model to adequately assess tropical cyclone risk in different regions by fitting regional impact functions based on reported damage data. Applying regional calibrated impact functions within the risk modeling framework CLIMADA (CLIMate ADAptation) at a resolution of 10 km worldwide, we find global annual average direct damage caused by tropical cyclones to range from USD 51 up to USD 121 billion (value in 2014, 1980{\textendash}2017) with the largest uncertainties in the West Pacific basin where the calibration results are the least robust. To better understand the challenges in the West Pacific and to complement the global perspective of this study, we explore uncertainties and limitations entailed in the modeling setup for the case of the Philippines. While using wind as a proxy for tropical cyclone hazard proves to be a valid approach in general, the case of the Philippines reveals limitations of the model and calibration due to the lack of an explicit representation of sub-perils such as storm surges, torrential rainfall, and landslides. The globally consistent methodology and calibrated regional impact functions are available online as a Python package ready for application in practical contexts like physical risk disclosure and providing more credible information for climate adaptation studies.},
 	language = {English},
 	number = {1},
@@ -230,6 +230,7 @@
 	address = {Cambridge, United Kingdom and New York, NY, USA},
 	title = {Human {Health}: {Impacts}, {Adaptation}, and {Co}-{Benefits}},
 	url = {https://www.ipcc.ch/report/ar5/wg2/},
+	doi = {10.1017/CBO9781107415379.016},
 	booktitle = {Climate {Change} 2014: {Impacts}, {Adaptation}, and {Vulnerability}. {Part} {A}: {Global} and {Sectoral} {Aspects}. {Contribution} of {Working} {Group} {II} to the {Fifth} {Assessment} {Report} of the {Intergovernmental} {Panel} on {Climate} {Change}},
 	publisher = {Cambridge University Press},
 	author = {Smith, Kirk R. and Woodward, Alistair and Campbell-Lendrum, Diarmid and Chadee, Dave D. and Honda, Yasushi and Liu, Qiyong and Olwoch, Jane M. and Revich, Boris and Sauerborn, Rainer},
@@ -305,7 +306,7 @@ section: Bulletin of the American Meteorological Society},
 	volume = {12},
 	issn = {1866-3508},
 	url = {https://www.earth-syst-sci-data.net/12/817/2020/},
-	doi = {https://doi.org/10.5194/essd-12-817-2020},
+	doi = {10.5194/essd-12-817-2020},
[…]
