### 2024-05-03 e7e441a
JOSS Paper: Do not call it 'natural disaster'.
