### 2024-05-03 #6719 @crvernon
:wave: @peanutfun - thanks for your submission to JOSS. We are currently managing a large backlog of submissions and the editor most appropriate for your area is already rather busy.
