### 2024-05-03 #6719 @peanutfun
@crvernon Thanks for letting me know. We will be patient :)

The newer code is located in the [/climada/util/calibrate/](https://github.com/CLIMADA-project/climada_python/tree/calibrate-impact-functions/climada/util/calibrate) directory in the `calibrate-impact-funcions` branch. All additions we submit via this publication, including the paper, a tutorial Jupyter script, and the documentation can be reviewed from the following pull request: https://github.com/CLIMADA-project/climada_python/pull/692
