+++ .github/workflows/ci.yml
@@ -69,7 +69,7 @@ jobs:
 
   test-petals:
     name: Petals Compatibility
-    uses: CLIMADA-project/climada_petals/.github/workflows/testing.yml@github-actions
+    uses: CLIMADA-project/climada_petals/.github/workflows/testing.yml@develop
     needs: build-and-test
     with:
       core_branch: ${{ github.ref }}
