### 2024-05-03 c762325
Improve docstring latex
