### 2024-05-03 1919ef8
Start updating unsequa.calc_base to include SAlib's 'ff method'
