### 2024-05-03 24238e7
Update latex docstring
