+++ climada/entity/impact_funcs/base.py
@@ -291,8 +291,8 @@ class ImpactFunc():
 
         .. math::
 
-            f(I) = \frac{luk(I)^{exponent}}{1 + luk(I)^{exponent}}\cdot scale \\
-            luk(I) = \frac{\max[I - threshold, 0]}{half\_point - threshold}
+            f(I) = \frac{\textrm{luk}(I)^{\textrm{exponent}}}{1 + \textrm{luk}(I)^{\textrm{exponent}}}\cdot \textrm{scale} \\
+            \textrm{luk}(I) = \frac{\max[I - \textrm{threshold}, 0]}{\textrm{half_point} - \textrm{threshold}}
 
         This function is inspired by Emanuel et al. (2011)
         https://doi.org/10.1175/WCAS-D-11-00007.1
