### 2024-05-06 3fa55db
Update docstrings for clarity and typos
