+++ climada/entity/impact_funcs/base.py
@@ -173,11 +173,8 @@ class ImpactFunc():
 
         """ Step function type impact function.
 
-        By default, everything is destroyed above the step.
+        By default, the impact is 100% above the step.
         Useful for high resolution modelling.
-
-        This method modifies self (climada.entity.impact_funcs instance)
-        by assigning an id, intensity, mdd and paa to the impact function.
 
         Parameters
         ----------
@@ -242,7 +239,7 @@ class ImpactFunc():
 
         Parameters
         ----------
-        intensity: tuple(float, float, float)
+        intensity : tuple(float, float, float)
             tuple of 3 intensity numbers along np.arange(min, max, step)
         L : float
             "top" of sigmoid
@@ -297,22 +294,26 @@ class ImpactFunc():
         This function is inspired by Emanuel et al. (2011)
         https://doi.org/10.1175/WCAS-D-11-00007.1
 
+        This method only specifies mdd, and paa = 1 for all intensities.
+
         Parameters
         ----------
-        intensity: tuple(float, float, float)
+        intensity : tuple(float, float, float)
             tuple of 3 intensity numbers along np.linsapce(min, max, num)
         threshold : float
             Intensity threshold below which there is no impact.
-            Should in general be larger than 0 for computational efficiency
+            In general choose threshold > 0 for computational efficiency
             of impacts.
         half_point : float
-            Intensity at which 50% of maxixmum impact is expected.
-            If smaller than threshold, mdd = 0 for all intensities.
+            Intensity at which 50% of maximum impact is expected.
+            If smaller than threshold, mdd = 0 (and f(I)=0) for all
+            intensities.
         scale : float
             Multiplicative factor for the whole function. Typically,
             this sets the maximum value at large intensities.
         exponent: float
             Exponent of the polynomial. Must be larger than 0.
+            Emanuel et al. (2011) uses the value 3.
         haz_type: str
             Reference string for the hazard (e.g., 'TC', 'RF', 'WS', ...)
         impf_id : int, optional, default=1
@@ -320,10 +321,14 @@ class ImpactFunc():
         kwargs :
             keyword arguments passed to ImpactFunc()
 
+        Raises
+        ------
+        ValueError : if exponent <= 0
+
         Returns
         -------
         impf : climada.entity.impact_funcs.ImpactFunc
-            s-shapep polynomial impact function
+            s-shaped polynomial impact function
         """
         if exponent <= 0:
             raise ValueError('Exponent value must larger than 0')
