### 2024-05-06 133c834
Update climada/entity/impact_funcs/base.py
Co-authored-by: Lukas Riedel <34276446+peanutfun@users.noreply.github.com>
