### 2024-05-06 b355c0a
Allow exponent 0, Make mdd=0 for threshold = half_point
