### 2024-05-06 7738db3
Update varnames for EMDAT data for 2024 and change formatting for reading csv
