### 2024-05-06 45c2ed9
Fix error when no second_order argument for handling ff method
