### 2024-05-09 123b61b
Improve a check in tc_tracks_synth slightly
