### 2024-05-10 e18e824
create an empty centroids before  with CRS from the first haz.
