+++ climada/hazard/base.py
@@ -818,6 +818,7 @@ class Hazard(HazardIO, HazardPlot):
             return cls()
         haz_concat = haz_list[0].__class__()
         haz_concat.haz_type = haz_list[0].haz_type
+        haz_concat.centroids = Centroids(lat=[], lon=[], crs=haz_list[0].centroids.crs)
         for attr_name, attr_val in vars(haz_list[0]).items():
             # to save memory, only copy simple attributes like
             # "units" that are not explicitly handled by Hazard.append
