### 2024-05-13 b101fdb
update haz.concat
