### 2024-05-13 8dc29ef
refactor gmst_info dictionary
