+++ climada/hazard/tc_clim_change.py
@@ -205,19 +205,8 @@ def get_gmst_info():
         - gmst_data: array with GMST data across RCPs (first dim) and years (second dim)
     """
 
-    gmst_info = {}
-
-    gmst_info.update({'rcps' : ['2.6', '4.5', '6.0', '8.5']})
-    gmst_info.update({'gmst_start_year' : 1880})
-    gmst_info.update({'gmst_end_year' : 2100})
-
-    gmst_info.update({'gmst_data':
-        np.empty((len(gmst_info['rcps']),
-                  gmst_info['gmst_end_year']-gmst_info['gmst_start_year']+1))
-    })
-
-    gmst_info['gmst_data'][0] = [
-        -0.16,-0.08,-0.1,-0.16,-0.28,-0.32,-0.3,-0.35,-0.16,-0.1,
+    gmst_data = np.array([
+        [-0.16,-0.08,-0.1,-0.16,-0.28,-0.32,-0.3,-0.35,-0.16,-0.1,
         -0.35,-0.22,-0.27,-0.31,-0.3,-0.22,-0.11,-0.11,-0.26,-0.17,
         -0.08,-0.15,-0.28,-0.37, -0.47,-0.26,-0.22,-0.39,-0.43,-0.48,
         -0.43,-0.44,-0.36,-0.34,-0.15,-0.14,-0.36,-0.46,-0.29,-0.27,
@@ -245,85 +234,94 @@ def get_gmst_info():
         1.262514286,1.290514286,1.285114286,1.267214286,1.267414286,
         1.294314286,1.315614286,1.310314286,1.283914286,1.296614286,
         1.281214286,1.301014286,1.300114286,1.303114286,1.286714286,
-        1.297514286,1.312114286,1.276714286,1.281414286,1.276414286
-    ]
-
-    gmst_info['gmst_data'][1] = [
-        -0.16,-0.08,-0.1,-0.16,-0.28,-0.32,-0.3,-0.35,-0.16,-0.1,
+        1.297514286,1.312114286,1.276714286,1.281414286,1.276414286],
+        [-0.16,-0.08,-0.1,-0.16,-0.28,-0.32,-0.3,-0.35,-0.16,-0.1,
         -0.35, -0.22,-0.27,-0.31,-0.3,-0.22,-0.11,-0.11,-0.26,-0.17,
         -0.08,-0.15,-0.28,-0.37,-0.47,-0.26,-0.22,-0.39,-0.43,-0.48,
         -0.43,-0.44,-0.36,-0.34,-0.15,-0.14,-0.36,-0.46, -0.29,-0.27,
-        -0.27,-0.19,-0.28,-0.26,-0.27,-0.22,-0.1,-0.22,-0.2,-0.36,-0.16,
-        -0.1,-0.16,-0.29,-0.13,-0.2,-0.15,-0.03,-0.01,-0.02,0.13,0.19,
-        0.07,0.09,0.2,0.09,-0.07,-0.03,-0.11,-0.11,-0.17,-0.07,0.01,0.08,
-        -0.13,-0.14,-0.19,0.05,0.06,0.03,-0.02,0.06,0.04,0.05,-0.2,-0.11,
-        -0.06,-0.02,-0.08,0.05,0.02,-0.08,0.01,0.16,-0.07,-0.01,-0.1,0.18,
-        0.07,0.16,0.26,0.32,0.14,0.31,0.15,0.11,0.18,0.32,0.38,0.27,0.45,
-        0.4,0.22,0.23,0.32,0.45,0.33,0.47,0.61,0.39,0.39,0.54,0.63,0.62,
-        0.54,0.68,0.64,0.66,0.54,0.66,0.72,0.61,0.64,0.68,0.75,0.9,1.02,
[…]
