### 2024-05-14 954a31c
Logging tc_tracks_synth small fix
