+++ climada/entity/exposures/base.py
@@ -77,6 +77,70 @@ DEF_VAR_MAT = {'sup_field_name': 'entity',
               }
 """MATLAB variable names"""
 
+class _InitArguments():
+    """helper class for sorting out `Exposrues.__init__` arguments"""
+
+    def _consolidate(self, meta, nam, val, default=None):
+        if not (val is None or meta.get(nam) is None or val == meta.get(nam)):
+            raise ValueError(f"conflicting arguments: the given value for {nam}"
+                             " is different from its value in meta")
+        return meta.get(nam, val) or default
+
+    def __init__(self,
+            data=None,
+            index=None,
+            columns=None,
+            dtype=None,
+            copy=False,
+            geometry=None,
+            crs=None,
+            meta=None,
+            description=None,
+            ref_year=None,
+            value_unit=None,
+            scheduler=None,
+        ):
+        """
+        Parameters
+        ----------
+        data : dict, iterable, DataFrame, ndarray
+        index : Index, array
+        columns : Index, array
+        dtype : dtype
+        copy : bool
+        geometry : array of Point, Polygon, MultiPolygon
+        crs : Any
+            forwarded to pyproj.CRS.from_user_input
+        meta : dict
+        description : str
+        ref_year : int
+        value_unit : str
+        scheduler : str
+            forwarded to dask.dataframe.map_partitions
+            if geometry is going to be calculated
+        """
+        meta = meta or {}
+        if not isinstance(meta, dict):
+            raise TypeError("meta must be of type dict")
+
+        self.description = self._consolidate(meta, 'description', description)
+        self.ref_year = self._consolidate(meta, 'ref_year', ref_year, DEF_REF_YEAR)
+        self.value_unit = self._consolidate(meta, 'value_unit', value_unit, DEF_VALUE_UNIT)
+
+        arg_crs = self._consolidate(meta, 'crs', crs, DEF_CRS)
+        try:
+            data_crs = data.geometry.crs
+        except AttributeError:
+            data_crs = None
+        if data_crs and arg_crs and not u_coord.equal_crs(data_crs, arg_crs):
+            raise ValueError("Inconsistent crs definition in data and data.geometry")
+        self.crs = arg_crs or data_crs or None
+
+        self.geometry = kwargs.pop('geometry', self.from_data('geometry'))
+        self.latitude = kwargs.pop('latitude', self.from_data('latitude'))
+        self.longitude = kwargs.pop('latitude', self.from_data('longitude'))
+
+
 class Exposures():
     """geopandas GeoDataFrame with metada and columns (pd.Series) defined in
     Attributes.
@@ -142,13 +206,20 @@ class Exposures():
             return self.meta.get('crs')
 
     def __init__(self, *args, meta=None, description=None, ref_year=DEF_REF_YEAR,
-                 value_unit=DEF_VALUE_UNIT, crs=None, **kwargs):
+                 value_unit=DEF_VALUE_UNIT, crs=None, scheduler=None, **kwargs):
         """Creates an Exposures object from a GeoDataFrame
 
         Parameters
         ----------
         args :
             Arguments of the GeoDataFrame constructor
+            data : dict, dataframe, ndarray, iterable,
+            index : index, array, 
+            columns : index, array,
+            dtype : dtype,
+            copy : bool,
+            geometry : array, str,
+            crs : Any (pyproj.CRS.from_user_input!)
         kwargs :
             Named arguments of the GeoDataFrame constructor, additionally
         meta : dict, optional
@@ -162,52 +233,34 @@ class Exposures():
         crs : object, anything accepted by pyproj.CRS.from_user_input
             Coordinate reference system. Defaults to the entry of the same name in `meta`, or to
             the CRS of the GeoDataFrame (if provided) or to 'epsg:4326'.
[…]
