### 2024-05-15 57c81e5
climada.entity.Exposures rewrite __init__
