+++ climada/entity/exposures/base.py
@@ -80,67 +80,6 @@ DEF_VAR_MAT = {'sup_field_name': 'entity',
 class _InitArguments():
     """helper class for sorting out `Exposrues.__init__` arguments"""
 
-    def _consolidate(self, meta, nam, val, default=None):
-        if not (val is None or meta.get(nam) is None or val == meta.get(nam)):
-            raise ValueError(f"conflicting arguments: the given value for {nam}"
-                             " is different from its value in meta")
-        return meta.get(nam, val) or default
-
-    def __init__(self,
-            data=None,
-            index=None,
-            columns=None,
-            dtype=None,
-            copy=False,
-            geometry=None,
-            crs=None,
-            meta=None,
-            description=None,
-            ref_year=None,
-            value_unit=None,
-            scheduler=None,
-        ):
-        """
-        Parameters
-        ----------
-        data : dict, iterable, DataFrame, ndarray
-        index : Index, array
-        columns : Index, array
-        dtype : dtype
-        copy : bool
-        geometry : array of Point, Polygon, MultiPolygon
-        crs : Any
-            forwarded to pyproj.CRS.from_user_input
-        meta : dict
-        description : str
-        ref_year : int
-        value_unit : str
-        scheduler : str
-            forwarded to dask.dataframe.map_partitions
-            if geometry is going to be calculated
-        """
-        meta = meta or {}
-        if not isinstance(meta, dict):
-            raise TypeError("meta must be of type dict")
-
-        self.description = self._consolidate(meta, 'description', description)
-        self.ref_year = self._consolidate(meta, 'ref_year', ref_year, DEF_REF_YEAR)
-        self.value_unit = self._consolidate(meta, 'value_unit', value_unit, DEF_VALUE_UNIT)
-
-        arg_crs = self._consolidate(meta, 'crs', crs, DEF_CRS)
-        try:
-            data_crs = data.geometry.crs
-        except AttributeError:
-            data_crs = None
-        if data_crs and arg_crs and not u_coord.equal_crs(data_crs, arg_crs):
-            raise ValueError("Inconsistent crs definition in data and data.geometry")
-        self.crs = arg_crs or data_crs or None
-
-        self.geometry = kwargs.pop('geometry', self.from_data('geometry'))
-        self.latitude = kwargs.pop('latitude', self.from_data('latitude'))
-        self.longitude = kwargs.pop('latitude', self.from_data('longitude'))
-
-
 class Exposures():
     """geopandas GeoDataFrame with metada and columns (pd.Series) defined in
     Attributes.
@@ -205,23 +144,49 @@ class Exposures():
             # In case of gdf without geometry, empty or before set_geometry_points was called
             return self.meta.get('crs')
 
-    def __init__(self, *args, meta=None, description=None, ref_year=DEF_REF_YEAR,
-                 value_unit=DEF_VALUE_UNIT, crs=None, scheduler=None, **kwargs):
-        """Creates an Exposures object from a GeoDataFrame
+    @staticmethod
+    def _consolidate(alternative_data, name, value, default=None):
+        if value is None:
+            return alternative_data.get(name) or default
+        if alternative_data.get(name) is None or alternative_data.get(name) == value:
+            return value
+        raise ValueError(f"conflicting arguments: the given {name}"
+                         " is different from their corresponding value(s) in meta or data")
 
+    def __init__(self,
+            data=None,
+            index=None,
+            columns=None,
+            dtype=None,
+            copy=False,
+            geometry=None,
+            crs=None,
+            meta=None,
+            description=None,
+            ref_year=None,
+            value_unit=None,
+            value=None,
+            lat=None,
+            lon=None,
+            scheduler=None,
+        ):
[…]
