+++ climada/util/coordinates.py
@@ -1055,7 +1055,7 @@ def match_centroids(coord_gdf, centroids, distance='euclidean',
         pass
 
     assigned = match_coordinates(
-        np.stack([coord_gdf['latitude'].values, coord_gdf['longitude'].values], axis=1),
+        np.stack([coord_gdf.geometry.y.values, coord_gdf.geometry.x.values], axis=1),
         centroids.coord, distance=distance, threshold=threshold,
     )
     return assigned
