+++ climada/entity/exposures/base.py
@@ -21,6 +21,7 @@ Define Exposures class.
 
 __all__ = ['Exposures', 'add_sea', 'INDICATOR_IMPF', 'INDICATOR_CENTR']
 
+from collections.abc import Iterable
 import logging
 import copy
 from pathlib import Path
@@ -77,9 +78,6 @@ DEF_VAR_MAT = {'sup_field_name': 'entity',
               }
 """MATLAB variable names"""
 
-class _InitArguments():
-    """helper class for sorting out `Exposrues.__init__` arguments"""
-
 class Exposures():
     """geopandas GeoDataFrame with metada and columns (pd.Series) defined in
     Attributes.
@@ -125,7 +123,7 @@ class Exposures():
     """
     _metadata = ['description', 'ref_year', 'value_unit', 'meta']
 
-    vars_oblig = ['value', 'latitude', 'longitude']
+    vars_oblig = ['value', 'geometry']
     """Name of the variables needed to compute the impact."""
 
     vars_def = [INDICATOR_IMPF, INDICATOR_IMPF_OLD]
@@ -138,18 +136,48 @@ class Exposures():
     @property
     def crs(self):
         """Coordinate Reference System, refers to the crs attribute of the inherent GeoDataFrame"""
-        try:
-            return self.gdf.geometry.crs or self.meta.get('crs')
-        except AttributeError:  # i.e., no geometry, crs is assumed to be a property
-            # In case of gdf without geometry, empty or before set_geometry_points was called
-            return self.meta.get('crs')
+        return self.data.geometry.crs
+
+    @property
+    def gdf(self):
+        """Inherent GeoDataFrame"""
+        return self.data
+
+    @property
+    def latitude(self):
+        """Latitude array of exposures"""
+        return self.data.geometry.y.values
+
+    @property
+    def longitude(self):
+        """Longitude array of exposures"""
+        return self.data.geometry.x.values
+
+    @property
+    def geometry(self):
+        """Geometry array of exposures"""
+        return self.data.geometry.values
+
+    @property
+    def value(self):
+        """Geometry array of exposures"""
+        return self.data["value"].values
 
     @staticmethod
     def _consolidate(alternative_data, name, value, default=None):
+        altvalue = alternative_data.get(name)    
+        if value is None and altvalue is None:
+            return default
         if value is None:
-            return alternative_data.get(name) or default
-        if alternative_data.get(name) is None or alternative_data.get(name) == value:
+            return altvalue
+        if altvalue is None:
             return value
+        try:
+            if all(altvalue == value):
+                return value
+        except TypeError:
+            if altvalue == value:
+                return value
         raise ValueError(f"conflicting arguments: the given {name}"
                          " is different from their corresponding value(s) in meta or data")
 
@@ -158,7 +186,7 @@ class Exposures():
             index=None,
             columns=None,
             dtype=None,
-            copy=False,
+            copy=False,  # pylint: disable=redefined-outer-name
             geometry=None,
             crs=None,
             meta=None,
@@ -210,7 +238,7 @@ class Exposures():
         """
         geodata = GeoDataFrame(data=data, index=index, columns=columns, dtype=dtype, copy=False)
 
-        self.geometry = self._consolidate(geodata, "geometry", geometry)
+        geometry = self._consolidate(geodata, "geometry", geometry)
         value = self._consolidate(geodata, "value", value)
 
         # both column names are accepted, lat and latitude, respectively lon and longitude.
@@ -220,33 +248,15 @@ class Exposures():
[…]
