### 2024-05-15 #6755 @editorialbot
[REVIEW]: A Module for Calibrating Impact Functions in the Climate Risk Modeling Platform CLIMADA
**Submitting author:** @peanutfun (<a href="http://orcid.org/0000-0002-4667-3652">Lukas Riedel</a>)
**Repository:** https://github.com/CLIMADA-project/climada_python
**Branch with paper.md** (empty if default branch): main
**Version:** 5.0.0
