### 2024-05-15 #6755 @observingClouds
@sunt05, @naingeet thank you very much for agreeing to review this submission. Your expertise is very valuable. With the opening of this issue, the official review process can now start. A good starting point is to create the personal review checklist by commenting `@editorialbot generate my checklist` and tick the checked items as you go through the submission. You can find more information about his above or also in the documentation.

Please note that we ask to review the additions made via the PR https://github.com/CLIMADA-project/climada_python/pull/692 as they reflect the new additions and the contents of the submitted manuscript ([see this comment](https://github.com/openjournals/joss-reviews/issues/6719#issuecomment-2092903985)). You might want to checkout the particular branch `calibrate-impact-funcions`.
