### 2024-05-16 cd464ef
climada.entity.Exposures: refactor __init__
