+++ climada/entity/exposures/test/test_base.py
@@ -63,13 +63,13 @@ class TestFuncs(unittest.TestCase):
         haz = Hazard.from_raster([HAZ_DEMO_FL], haz_type='FL', window=Window(10, 20, 50, 60))
         ncentroids = haz.centroids.size
 
-        exp = Exposures(crs=haz.centroids.crs)
-
-        # some are matching exactly, some are geographically close
-        exp.gdf['longitude'] = np.concatenate([
-            haz.centroids.lon, haz.centroids.lon + 0.001 * (-0.5 + np_rand.rand(ncentroids))])
-        exp.gdf['latitude'] = np.concatenate([
-            haz.centroids.lat, haz.centroids.lat + 0.001 * (-0.5 + np_rand.rand(ncentroids))])
+        exp = Exposures(
+            crs=haz.centroids.crs,
+            lon = np.concatenate([
+                haz.centroids.lon, haz.centroids.lon + 0.001 * (-0.5 + np_rand.rand(ncentroids))]),
+            lat = np.concatenate([
+                haz.centroids.lat, haz.centroids.lat + 0.001 * (-0.5 + np_rand.rand(ncentroids))])
+        )
         expected_result = np.concatenate([np.arange(ncentroids), np.arange(ncentroids)])
 
         # make sure that it works for both float32 and float64
@@ -85,14 +85,14 @@ class TestFuncs(unittest.TestCase):
 
     def test__init__meta_type(self):
         """ Check if meta of type list raises a ValueError in __init__"""
-        with self.assertRaises(ValueError) as cm:
+        with self.assertRaises(TypeError) as cm:
             Exposures(meta="{}")
-        self.assertEqual("meta must be a dictionary",
+        self.assertEqual("meta must be of type dict",
                       str(cm.exception))
 
     def test__init__geometry_type(self):
         """Check that initialization fails when `geometry` is given as a `str` argument"""
-        with self.assertRaises(ValueError) as cm:
+        with self.assertRaises(TypeError) as cm:
             Exposures(geometry='myname')
         self.assertEqual("Exposures is not able to handle customized 'geometry' column names.",
                          str(cm.exception))
@@ -110,14 +110,14 @@ class TestFuncs(unittest.TestCase):
         exp = Exposures.from_raster(HAZ_DEMO_FL, window=Window(10, 20, 50, 60))
         exp.check()
         self.assertTrue(u_coord.equal_crs(exp.crs, DEF_CRS))
-        self.assertAlmostEqual(exp.gdf['latitude'].max(),
+        self.assertAlmostEqual(exp.latitude.max(),
                                10.248220966978932 - 0.009000000000000341 / 2)
-        self.assertAlmostEqual(exp.gdf['latitude'].min(),
+        self.assertAlmostEqual(exp.latitude.min(),
                                10.248220966978932 - 0.009000000000000341
                                / 2 - 59 * 0.009000000000000341)
-        self.assertAlmostEqual(exp.gdf['longitude'].min(),
+        self.assertAlmostEqual(exp.longitude.min(),
                                -69.2471495969998 + 0.009000000000000341 / 2)
-        self.assertAlmostEqual(exp.gdf['longitude'].max(),
+        self.assertAlmostEqual(exp.longitude.max(),
                                -69.2471495969998 + 0.009000000000000341
                                / 2 + 49 * 0.009000000000000341)
[…]
