### 2024-05-16 f0cc985
adaptations to changed Exposures structure: gdf.latitude and gdf.longitude have been eliminated
