+++ climada/entity/measures/test/test_base.py
@@ -171,8 +171,8 @@ class TestApply(unittest.TestCase):
         self.assertEqual(new_exp.value_unit, exp.value_unit)
         self.assertEqual(new_exp.description, exp.description)
         self.assertTrue(np.array_equal(new_exp.gdf.value.values, exp.gdf.value.values))
-        self.assertTrue(np.array_equal(new_exp.gdf.latitude.values, exp.gdf.latitude.values))
-        self.assertTrue(np.array_equal(new_exp.gdf.longitude.values, exp.gdf.longitude.values))
+        self.assertTrue(np.array_equal(new_exp.latitude, exp.latitude))
+        self.assertTrue(np.array_equal(new_exp.longitude, exp.longitude))
         self.assertTrue(np.array_equal(exp.gdf[INDICATOR_IMPF + 'TC'].values, np.ones(new_exp.gdf.shape[0])))
         self.assertTrue(np.array_equal(new_exp.gdf[INDICATOR_IMPF + 'TC'].values, np.ones(new_exp.gdf.shape[0]) * 3))
 
@@ -208,8 +208,8 @@ class TestApply(unittest.TestCase):
         self.assertEqual(new_exp.value_unit, ref_exp.value_unit)
         self.assertEqual(new_exp.description, ref_exp.description)
         self.assertTrue(np.array_equal(new_exp.gdf.value.values, ref_exp.gdf.value.values))
-        self.assertTrue(np.array_equal(new_exp.gdf.latitude.values, ref_exp.gdf.latitude.values))
-        self.assertTrue(np.array_equal(new_exp.gdf.longitude.values, ref_exp.gdf.longitude.values))
+        self.assertTrue(np.array_equal(new_exp.latitude, ref_exp.latitude))
+        self.assertTrue(np.array_equal(new_exp.longitude, ref_exp.longitude))
 
     def test_not_filter_exposures_pass(self):
         """Test _filter_exposures method with []"""
@@ -274,8 +274,6 @@ class TestApply(unittest.TestCase):
         self.assertEqual(res_exp.value_unit, exp.value_unit)
         self.assertEqual(res_exp.description, exp.description)
         self.assertTrue(u_coord.equal_crs(res_exp.crs, exp.crs))
-        self.assertFalse(hasattr(exp.gdf, "crs"))
-        self.assertFalse(hasattr(res_exp.gdf, "crs"))
 
         # regions (that is just input data, no need for testing, but it makes the changed and unchanged parts obious)
         self.assertTrue(np.array_equal(res_exp.gdf.region_id.values[0], 4))
@@ -286,15 +284,15 @@ class TestApply(unittest.TestCase):
         self.assertTrue(np.array_equal(res_exp.gdf.value.values[:25], new_exp.gdf.value.values[:25]))
         self.assertTrue(np.all(np.not_equal(res_exp.gdf.value.values[:25], exp.gdf.value.values[:25])))
         self.assertTrue(np.all(np.not_equal(res_exp.gdf.impf_TC.values[:25], new_exp.gdf.impf_TC.values[:25])))
-        self.assertTrue(np.array_equal(res_exp.gdf.latitude.values[:25], new_exp.gdf.latitude.values[:25]))
-        self.assertTrue(np.array_equal(res_exp.gdf.longitude.values[:25], new_exp.gdf.longitude.values[:25]))
+        self.assertTrue(np.array_equal(res_exp.latitude[:25], new_exp.latitude[:25]))
+        self.assertTrue(np.array_equal(res_exp.longitude[:25], new_exp.longitude[:25]))
 
         # unchanged exposures
[…]
