### 2024-05-16 7e8970b
adaptations to changed Exposures structure: gdf.latitude and gdf.longitude have been eliminated
