+++ climada/util/lines_polys_handler.py
@@ -397,10 +397,9 @@ def exp_geom_to_pnt(exp, res, to_meters, disagg_met, disagg_val):
     if disagg_met is DisaggMethod.DIV:
         gdf_pnt = _disagg_values_div(gdf_pnt)
 
-    # set lat lon and centroids
+    # set dataframe
     exp_pnt = exp.copy(deep=False)
     exp_pnt.set_gdf(gdf_pnt)
-    exp_pnt.set_lat_lon()
 
     return exp_pnt
 
