+++ climada/engine/test/test_impact.py
@@ -909,9 +909,9 @@ class TestConvertExp(unittest.TestCase):
 
         imp = dummy_impact()
         exp = imp._build_exp()
-        np.testing.assert_array_equal(imp.eai_exp, exp.gdf['value'])
-        np.testing.assert_array_equal(imp.coord_exp[:, 0], exp.gdf['latitude'])
-        np.testing.assert_array_equal(imp.coord_exp[:, 1], exp.gdf['longitude'])
+        np.testing.assert_array_equal(imp.eai_exp, exp.value)
+        np.testing.assert_array_equal(imp.coord_exp[:, 0], exp.latitude)
+        np.testing.assert_array_equal(imp.coord_exp[:, 1], exp.longitude)
         self.assertTrue(u_coord.equal_crs(exp.crs, imp.crs))
         self.assertEqual(exp.value_unit, imp.unit)
         self.assertEqual(exp.ref_year, 0)
@@ -922,9 +922,9 @@ class TestConvertExp(unittest.TestCase):
         imp = dummy_impact()
         event_id = imp.event_id[1]
         exp = imp._build_exp_event(event_id=event_id)
-        np.testing.assert_array_equal(imp.imp_mat[1].todense().A1, exp.gdf['value'])
-        np.testing.assert_array_equal(imp.coord_exp[:, 0], exp.gdf['latitude'])
-        np.testing.assert_array_equal(imp.coord_exp[:, 1], exp.gdf['longitude'])
+        np.testing.assert_array_equal(imp.imp_mat[1].todense().A1, exp.value)
+        np.testing.assert_array_equal(imp.coord_exp[:, 0], exp.latitude)
+        np.testing.assert_array_equal(imp.coord_exp[:, 1], exp.longitude)
         self.assertTrue(u_coord.equal_crs(exp.crs, imp.crs))
         self.assertEqual(exp.value_unit, imp.unit)
         self.assertEqual(exp.ref_year, 0)

[…]
