### 2024-05-16 55753c9
adaptations to changed Exposures structure: gdf.latitude and gdf.longitude have been eliminated
