### 2024-05-16 7233e34
centroids.Centroids.from_exposures() will always work as geometry column is granted
