### 2024-05-16 b0a0e65
adaptation to eliminated lat/lon columns
