### 2024-05-17 fa4e0f7
hazard.Centroids: region_id and on_land are not necessarily present in the data frame
