### 2024-05-17 1036b79
Exposures() requires geometry or lat/lon
