### 2024-05-17 2f57870
points_to_raster function getting more complicated, considering dataframes with geometry column instead fo lat/lon
