+++ climada/util/coordinates.py
@@ -2401,14 +2401,25 @@ def points_to_raster(points_df, val_names=None, res=0.0, raster_res=0.0, crs=DEF
     """
     if not val_names:
         val_names = ['value']
+
+    if 'geometry' in points_df:
+        latval = points_df.geometry.y
+        lonval = points_df.geometry.y
+    else:
+        latval = points_df.latitude.values
+        lonval = points_df.longitude.values
+
     if not res:
-        res = np.abs(get_resolution(points_df.latitude.values,
-                                    points_df.longitude.values)).min()
+        res = np.abs(get_resolution(latval, lonval)).min()
     if not raster_res:
         raster_res = res
 
-    def apply_box(df_exp):
+    if 'geometry' in points_df:
+        fun = lambda r: r.geometry.buffer(res / 2).envelope
+    else:
         fun = lambda r: Point(r.longitude, r.latitude).buffer(res / 2).envelope
+
+    def apply_box(df_exp):
         return df_exp.apply(fun, axis=1)
 
     LOGGER.info('Raster from resolution %s to %s.', res, raster_res)
@@ -2424,16 +2435,15 @@ def points_to_raster(points_df, val_names=None, res=0.0, raster_res=0.0, crs=DEF
 
     # renormalize longitude if necessary
     if equal_crs(df_poly.crs, DEF_CRS):
-        xmin, ymin, xmax, ymax = latlon_bounds(points_df.latitude.values,
-                                               points_df.longitude.values)
+        xmin, ymin, xmax, ymax = latlon_bounds(latval, lonval)
         x_mid = 0.5 * (xmin + xmax)
         # we don't really change the CRS when rewrapping, so we reset the CRS attribute afterwards
         df_poly = df_poly \
             .to_crs({"proj": "longlat", "lon_wrap": x_mid}) \
             .set_crs(DEF_CRS, allow_override=True)
     else:
-        xmin, ymin, xmax, ymax = (points_df.longitude.min(), points_df.latitude.min(),
-                                  points_df.longitude.max(), points_df.latitude.max())
+        xmin, ymin, xmax, ymax = (lonval.min(), latval.min(),
+                                  lonval.max(), latval.max())
 
     # construct raster
     rows, cols, ras_trans = pts_to_raster_meta((xmin, ymin, xmax, ymax),
