+++ doc/joss/calibration-module/paper.bib
@@ -224,6 +224,7 @@
 	month = feb,
 	year = {2013},
 	keywords = {Science / Environmental Science},
+	url = {https://cambridge.org/9781107006195},
 }
 
 @incollection{smith_human_2014,
