### 2024-05-21 ba9bb7f
JOSS: Remove unused reference
