+++ doc/joss/calibration-module/paper.bib
@@ -67,14 +67,6 @@
 	file = {CRED - 2023 - 2022. Disasters in Numbers.pdf:/Users/ldr.riedel/Zotero/storage/JQRMMVC2/CRED - 2023 - 2022. Disasters in Numbers.pdf:application/pdf},
 }
 
-@misc{emdat,
-	title = {{Emergency Events Database (EM-DAT)}},
-	author = {{CRED}},
-	year = {2024},
-	url = {https://www.emdat.be/},
-	howpublished = {{Centre for Research on the Epidemiology of Disasters (CRED), University of Louvain}},
-}
-
 @article{eberenz_regional_2021,
 	title = {Regional tropical cyclone impact functions for globally consistent risk assessments},
 	volume = {21},
