### 2024-05-21 #6755 @peanutfun
@observingClouds We were able to replace one work marked with "missing DOIs" with a suitable publication with a DOI. For the rest, we provide URLs:

- Bayesian Optimization: Open source constrained glo...: *GitHub repository, no associated publications found*
- 2022 Disasters in Numbers: *No DOI found*
- Emergency Events Database (EM-DAT): *Database website, we replaced it with Delforge et al., https://doi.org/10.21203/rs.3.rs-3807553/v1*
- Risk and Uncertainty Assessment for Natural Hazard... *No DOI found, we now provide a URL to the publisher website*
- Oasis Loss Modelling Framework: *Project website, no associated publications found*
