### 2024-05-21 9c16bc8
Use math in doc string
Co-authored-by: Lukas Riedel <34276446+peanutfun@users.noreply.github.com>
