### 2024-05-21 147ad92
Update docstring with DOI
