### 2024-05-21 91adf65
add description of returned array for angular windspeed functions
