### 2024-05-26 f4190e7
update write raster local return periods
