+++ climada/hazard/base.py
@@ -23,32 +23,85 @@ __all__ = ['Hazard']
 
 import copy
 import datetime as dt
+import itertools
 import logging
-from typing import Optional,List
+import pathlib
+from typing import Union, Optional, Callable, Dict, Any, List
 import warnings
 
+import h5py
+import matplotlib.pyplot as plt
 import numpy as np
+import pandas as pd
 from pathos.pools import ProcessPool as Pool
+import rasterio
+import rasterio.features
+import rasterio.warp
+from rasterio.features import rasterize
+import sparse as sp
 from scipy import sparse
 import xarray as xr
 import netCDF4 as nc
 from scipy.interpolate import griddata
 
-
 from climada import CONFIG
-from climada.hazard.plot import HazardPlot
-from climada.hazard.io import HazardIO
 from climada.hazard.centroids.centr import Centroids
 import climada.util.checker as u_check
 import climada.util.constants as u_const
 import climada.util.coordinates as u_coord
 import climada.util.dates_times as u_dt
-
+import climada.util.hdf5_handler as u_hdf5
+import climada.util.plot as u_plot
 
 LOGGER = logging.getLogger(__name__)
 
+DEF_VAR_EXCEL = {'sheet_name': {'inten': 'hazard_intensity',
+                                'freq': 'hazard_frequency'
+                                },
+                 'col_name': {'cen_id': 'centroid_id/event_id',
+                              'even_id': 'event_id',
+                              'even_dt': 'event_date',
+                              'even_name': 'event_name',
+                              'freq': 'frequency',
+                              'orig': 'orig_event_flag'
+                              },
+                 'col_centroids': {'sheet_name': 'centroids',
+                                   'col_name': {'cen_id': 'centroid_id',
+                                                'latitude': 'lat',
+                                                'longitude': 'lon',
+                                                }
+                                   }
+                 }
+"""Excel variable names"""
 
-class Hazard(HazardIO, HazardPlot):
+DEF_VAR_MAT = {'field_name': 'hazard',
+               'var_name': {'per_id': 'peril_ID',
+                            'even_id': 'event_ID',
+                            'ev_name': 'name',
+                            'freq': 'frequency',
+                            'inten': 'intensity',
+                            'unit': 'units',
+                            'frac': 'fraction',
+                            'comment': 'comment',
+                            'datenum': 'datenum',
+                            'orig': 'orig_event_flag'
+                            },
+               'var_cent': {'field_names': ['centroids', 'hazard'],
+                            'var_name': {'cen_id': 'centroid_ID',
+                                         'lat': 'lat',
+                                         'lon': 'lon'
+                                         }
+                            }
+               }
+"""MATLAB variable names"""
+
+DEF_COORDS = dict(event="time", longitude="longitude", latitude="latitude")
+"""Default coordinates when reading Hazard data from an xarray Dataset"""
+
+DEF_DATA_VARS = ["fraction", "frequency", "event_id", "event_name", "date"]
+"""Default keys for optional Hazard attributes when reading from an xarray Dataset"""
+
+class Hazard():
     """
     Contains events of some hazard type defined at centroids. Loads from
     files with format defined in FILE_EXT.
@@ -221,6 +274,790 @@ class Hazard(HazardIO, HazardPlot):
         """
         self._check_events()
 
+    @classmethod
+    def from_raster(cls, files_intensity, files_fraction=None, attrs=None,
+                    band=None, haz_type=None, pool=None, src_crs=None, window=None,
+                    geometry=None, dst_crs=None, transform=None, width=None,
+                    height=None, resampling=rasterio.warp.Resampling.nearest):
+        """Create Hazard with intensity and fraction values from raster files
+
+        If raster files are masked, the masked values are set to 0.
+
[…]
