### 2024-05-29 b16907e
initial draft for impacts snapshots collections
