+++ climada/engine/impact_trajectories.py
@@ -1,0 +1,187 @@
+"""
+This file is part of CLIMADA.
+
+Copyright (C) 2017 ETH Zurich, CLIMADA contributors listed in AUTHORS.
+
+CLIMADA is free software: you can redistribute it and/or modify it under the
+terms of the GNU General Public License as published by the Free
+Software Foundation, version 3.
+
+CLIMADA is distributed in the hope that it will be useful, but WITHOUT ANY
+WARRANTY; without even the implied warranty of MERCHANTABILITY or FITNESS FOR A
+PARTICULAR PURPOSE.  See the GNU General Public License for more details.
+
+You should have received a copy of the GNU General Public License along
+with CLIMADA. If not, see <https://www.gnu.org/licenses/>.
+
+---
+
+"""
+
+from abc import ABC
+import copy
+
+import numpy as np
+from datetime import datetime
+from scipy.sparse import lil_matrix
+
+from climada.hazard import Hazard
+from climada.engine.impact_calc import ImpactCalc
+
+
+### Utils functions
+
+def get_dates(haz: Hazard):
+    return [datetime.fromordinal(date) for date in haz.date]
+
+def get_years(haz: Hazard):
+    return np.unique(np.array([datetime.fromordinal(date).year for date in haz.date]))
+
+def grow_exp(exp, exp_growth_rate, elapsed):
+    exp_grown =copy.deepcopy(exp)
+    # Exponential growth
+    exp_growth_rate = 0.01
+    exp_grown.gdf.value = exp_grown.gdf.value * (1 + exp_growth_rate)**elapsed
+    return exp_grown
+
+def interpolate_sm(mat_start, mat_end, year, year_start, year_end):
+    if year < year_start or year > year_end:
+        raise ValueError("Year must be within the start and end years")
+
+    # Calculate the ratio of the difference between the target year and the start year
+    # to the total number of years between the start and end years
+    ratio = (year - year_start) / (year_end - year_start)
+
+    # Convert the input matrices to a format that allows efficient modification of its elements
+    mat_start = lil_matrix(mat_start)
+    mat_end = lil_matrix(mat_end)
+
+    # Perform the linear interpolation
+    mat_interpolated = mat_start + ratio * (mat_end - mat_start)
+
+    return mat_interpolated
+
+# Derive the intermediate probability distributions
+def interpolate_years(year_start, year_end):
+    # Generate an array of interpolated values between 0 and 1
+    values = np.linspace(0, 1, num=year_end-year_start+1)
+    return values
+
+
+class TBRSnapshots():
+    def __init__(self, exposure_set, hazard_set, impfset, TBR_snapyears):
+        self.exposure_set = exposure_set
+        self.hazard_set = hazard_set
+        self.impfset = impfset
+        self.TBR_snapyears = TBR_snapyears
+        self.data_dict = {
+            year: [exposure_set[year], hazard_set[year]] for year in TBR_snapyears
+        }
+
+    # Check that at least first and last snap are complete
+    # and otherwise it is ok
+
+    @classmethod
+    def from_dict(cls, snapshots_dict, impfset):
+        TBR_snapyears = list(snapshots_dict.keys())
+        exposure_set = {year:snapshots_dict[year][0] for year in TBR_snapyears}
+        hazard_set = {year:snapshots_dict[year][1] for year in TBR_snapyears}
[…]
