+++ climada/engine/impact_trajectories.py
@@ -18,7 +18,6 @@ with CLIMADA. If not, see <https://www.gnu.org/licenses/>.
 
 """
 
-from abc import ABC
 import copy
 
 import numpy as np
@@ -31,18 +30,22 @@ from climada.engine.impact_calc import ImpactCalc
 
 ### Utils functions
 
+
 def get_dates(haz: Hazard):
     return [datetime.fromordinal(date) for date in haz.date]
+
 
 def get_years(haz: Hazard):
     return np.unique(np.array([datetime.fromordinal(date).year for date in haz.date]))
 
+
 def grow_exp(exp, exp_growth_rate, elapsed):
-    exp_grown =copy.deepcopy(exp)
+    exp_grown = copy.deepcopy(exp)
     # Exponential growth
     exp_growth_rate = 0.01
-    exp_grown.gdf.value = exp_grown.gdf.value * (1 + exp_growth_rate)**elapsed
+    exp_grown.gdf.value = exp_grown.gdf.value * (1 + exp_growth_rate) ** elapsed
     return exp_grown
+
 
 def interpolate_sm(mat_start, mat_end, year, year_start, year_end):
     if year < year_start or year > year_end:
@@ -61,21 +64,22 @@ def interpolate_sm(mat_start, mat_end, year, year_start, year_end):
 
     return mat_interpolated
 
+
 # Derive the intermediate probability distributions
 def interpolate_years(year_start, year_end):
     # Generate an array of interpolated values between 0 and 1
-    values = np.linspace(0, 1, num=year_end-year_start+1)
+    values = np.linspace(0, 1, num=year_end - year_start + 1)
     return values
 
 
-class TBRSnapshots():
-    def __init__(self, exposure_set, hazard_set, impfset, TBR_snapyears):
+class SnapshotsCollection:
+    def __init__(self, exposure_set, hazard_set, impfset, snapshot_years):
         self.exposure_set = exposure_set
         self.hazard_set = hazard_set
         self.impfset = impfset
-        self.TBR_snapyears = TBR_snapyears
+        self.snapshots_years = snapshot_years
         self.data_dict = {
-            year: [exposure_set[year], hazard_set[year]] for year in TBR_snapyears
+            year: [exposure_set[year], hazard_set[year]] for year in snapshot_years
         }
 
     # Check that at least first and last snap are complete
@@ -83,19 +87,30 @@ class TBRSnapshots():
 
     @classmethod
     def from_dict(cls, snapshots_dict, impfset):
-        TBR_snapyears = list(snapshots_dict.keys())
-        exposure_set = {year:snapshots_dict[year][0] for year in TBR_snapyears}
-        hazard_set = {year:snapshots_dict[year][1] for year in TBR_snapyears}
-        return cls(exposure_set=exposure_set, hazard_set=hazard_set, impfset=impfset,TBR_snapyears=TBR_snapyears)
+        snapshot_years = list(snapshots_dict.keys())
+        exposure_set = {year: snapshots_dict[year][0] for year in snapshot_years}
+        hazard_set = {year: snapshots_dict[year][1] for year in snapshot_years}
+        return cls(
+            exposure_set=exposure_set,
+            hazard_set=hazard_set,
+            impfset=impfset,
+            snapshot_years=snapshot_years,
+        )
 
     @classmethod
-    def from_lists(cls, hazard_list, exposure_list, impfset, TBR_snapyears):
-        exposure_set = {year:exposure_list[i] for i,year in enumerate(TBR_snapyears)}
-        hazard_set = {year:hazard_list[i] for i,year in enumerate(TBR_snapyears)}
[…]
