### 2024-05-29 9b62a74
Cast hdmr output to 2D array to make it compatible with unsequa
