### 2024-05-30 b2c5ee6
Correct removing of problematic output from hdmr
