+++ climada/engine/unsequa/calc_base.py
@@ -559,9 +559,13 @@ def _calc_sens_df(method, problem_sa, sensitivity_kwargs, param_labels, X, unc_d
                                           #[param_labels.index(ie[1]), param_labels.index(ie[0])]] = sens_indices['IE'][i]
                 sens_indices['IE'] = interactions
         if method.__name__ == 'SALib.analyze.hdmr':
-            keys_to_remove = []
-            del sens_indices['Em']#need to delete emulator as is a dict
-            del sens_indices['Term']#and useless duplicate of names
+            keys_to_remove = ['Em', 'Term','X', 'Y'] #need to delete emulator as is a dict
+                                            #and useless duplicate of names
+            [keys_to_remove.append(si)  for si, si_val_array in sens_indices.items()
+             if (np.array(si_val_array).size > nparams**2)]
+            sens_indices = {k: v for k, v in sens_indices.items() if k not in keys_to_remove}
+            #del sens_indices['Em']#need to delete emulator as is a dict
+            #del sens_indices['Term']#and useless duplicate of names
             names = sens_indices.pop('names')
             for si, si_val_array in sens_indices.items():
                 if (np.array(si_val_array).ndim == 1 and #for everything that is 1d and has lentgh > n params, refactor to 2D
@@ -575,7 +579,7 @@ def _calc_sens_df(method, problem_sa, sensitivity_kwargs, param_labels, X, unc_d
                 elif (np.array(si_val_array).size > nparams**2):
                     keys_to_remove.append(si) #everything else that has size > nparams**2 is discarded
 
-            sens_indices = {k: v for k, v in sens_indices.items() if k not in keys_to_remove}
+
 
         sens_first_order = np.array([
             np.array(si_val_array)
