### 2024-05-30 8121140
Add tests for hdmr and ff
