### 2024-05-30 a6f938f
Update docstring for delta and hdmr  methods
