### 2024-05-30 8a84183
Remove delta from the compatible methods
