+++ climada/engine/unsequa/unc_output.py
@@ -56,7 +56,7 @@ MAP_CMAP = 'Dark2' #Default color map for the sensitivity map
 #Table of recommended pairing between salib sampling and sensitivity methods
 # NEEDS TO BE UPDATED REGULARLY!! https://salib.readthedocs.io/en/latest/api.html
 SALIB_COMPATIBILITY = {
-    'delta': ['fast_sampler', 'ff', 'finite_diff', 'latin', 'morris', 'saltelli'],
+    #'delta': ['fast_sampler', 'ff', 'finite_diff', 'latin', 'morris', 'saltelli'],
     'dgsm': ['finite_diff'],
     'fast': ['fast_sampler'],
     'ff': ['ff'],
@@ -569,7 +569,7 @@ class UncOutput():
                 print(f"No data to plot for '{col}'.")
                 if ax is not None:
                     ax.text(0.5, 0.5, 'No data to plot', fontsize=18,
-                            horizontalalignment='center', verticalalignment='center', 
+                            horizontalalignment='center', verticalalignment='center',
                             transform=ax.transAxes)
                     ax.set_xlabel(col)
                     ax.set_ylabel('density of samples')
@@ -631,11 +631,11 @@ class UncOutput():
         return axes
 
 
-    def plot_rp_uncertainty(self, orig_list=None, figsize=(16, 6), axes=None, 
+    def plot_rp_uncertainty(self, orig_list=None, figsize=(16, 6), axes=None,
                             calc_delta=False):
         """
         Plot the distribution of return period uncertainty
-    
+
         Parameters
         ----------
         orig_list : list[float], optional
@@ -649,18 +649,18 @@ class UncOutput():
             Axes handles to use for the plot. The default is None.
         calc_delta: boolean, optional
             Adapt axis labels for CalcDeltaImpact unc_output. Default is False.
-    
+
         Raises
         ------
         ValueError
             If no metric distribution was computed the plot cannot be made.
-    
+
         Returns
         -------
         axes : matplotlib.pyplot.axes
             The axis handle of the plot.
         """
-    
+
         try:
             unc_df = self.freq_curve_unc_df
         except AttributeError:
@@ -669,21 +669,21 @@ class UncOutput():
             raise ValueError("No return period uncertainty data present "
                     "Please run an uncertainty analysis with the desired "
                     "return period specified.")
-    
+
         add_orig=True
         if orig_list is None:
             add_orig=False
-    
+
         if axes is None:
             _fig, axes = plt.subplots(figsize=figsize, nrows=1, ncols=2)
-    
+
         [min_l, max_l], m_unit = u_vtm([unc_df.min().min(), unc_df.max().max()], n_sig_dig=4)
-    
+
         # Plotting for the first axes
         ax = axes[0]
         prop_cycle = plt.rcParams['axes.prop_cycle']
         colors = prop_cycle.by_key()['color']
-    
+
         for n, (_name, values) in enumerate(unc_df.items()):
             if values.isna().all() or len(values.dropna()) < 2:
                 print(f"Skipping plot for '{_name}': insufficient data.")
@@ -701,50 +701,50 @@ class UncOutput():
                     color=colors[n], linestyle='dotted', linewidth=2,
                     label="orig=%.2f%s" %(orig_val, m_unit)
                 )
-    
+
         ax.set_xlim(min_l, max_l)
         ax.set_ylim(0, 2*unc_df.shape[1])
         ax.set_yticks(np.arange(0, 2*unc_df.shape[1], 2))
[…]
