### 2024-05-30 a81035d
Decrease precision level for hdmr test because fails otherwise
