### 2024-05-30 576d0d3
Add rp cutoff helper methods
