### 2024-05-30 1876516
Improve docstrings and readability of the code.
