+++ climada/engine/unsequa/calc_base.py
@@ -203,7 +203,7 @@ class Calc():
             Number of samples as used in the sampling method from SALib
         sampling_method : str, optional
             The sampling method as defined in SALib. Possible choices:
-            'dgsm', 'fast_sampler', 'ff', 'finite_diff', 'latin', 'morris', 'saltelli',
+            'saltelli', 'latin', 'morris', 'dgsm', 'fast_sampler', 'ff', 'finite_diff',
              https://salib.readthedocs.io/en/latest/api.html
             The default is 'saltelli'.
         sampling_kwargs : kwargs, optional
@@ -242,24 +242,27 @@ class Calc():
             'names' : param_labels,
             'bounds' : [[0, 1]]*len(param_labels)
             }
-        #not sure if we want to still force users to enter a value for N in the
-        #case of the 'ff' sampler. In the meantime, trigger a warning.
+        #for the ff sampler, no value of N is needed. For API consistency the user
+        #must input a value that is ignored and a warning is given. 
         if sampling_method == 'ff':
             LOGGER.warning("""You are using the 'ff' sampler which does not require
-                           a value for 'N. The entered N value will be ignored
+                           a value for N. The entered N value will be ignored
                            in the sampling process.""")
         uniform_base_sample = self._make_uniform_base_sample(N, problem_sa,
                                                              sampling_method,
                                                              sampling_kwargs)
         df_samples = pd.DataFrame(uniform_base_sample, columns=param_labels)
+        params_ignored = [] #store dummy parameters that will be ignored
         for param in list(df_samples):
-            if param not in self.distr_dict: #dummy_0 param added to uniform_base_sample
-                                             #when using ff method, need to ignore it?
-                continue
-            df_samples[param] = df_samples[param].apply(
-                self.distr_dict[param].ppf
-                )
-
+            if (param not in self.distr_dict and "dummy_" in param):
+                params_ignored.append(param)
+            else:
+                df_samples[param] = df_samples[param].apply(
+                    self.distr_dict[param].ppf
+                    )
+        if params_ignored:
+            LOGGER.warning("""The automatically generated dummy parameters:  """
+                           +str(params_ignored)+ """ are being ignored.""")
         sampling_kwargs = {
             key: str(val)
             for key, val in sampling_kwargs.items()
@@ -290,7 +293,7 @@ class Calc():
             SALib sampling method.
         sampling_method: string
             The sampling method as defined in SALib. Possible choices:
-            'dgsm', 'fast_sampler', 'ff', 'finite_diff', 'latin', 'morris', 'saltelli',
+            'saltelli', 'latin', 'morris', 'dgsm', 'fast_sampler', 'ff', 'finite_diff',
             https://salib.readthedocs.io/en/latest/api.html
         sampling_kwargs: dict()
             Optional keyword arguments passed on to the SALib sampling method.
@@ -315,7 +318,7 @@ class Calc():
         if sampling_method == 'ff': #the ff sampling has a fixed sample size and
                                     #does not require the N parameter
             if (problem_sa['num_vars'] & (problem_sa['num_vars'] - 1) != 0):
-                raise ValueError("The number of parameters must be a power of 2 "
+                raise ValueError("The number of parameters must be a power of 2. "
                                  "to use the ff sampling method. You can generate "
                                  "dummy parameters to overcome this limitation."
                                  " See https://salib.readthedocs.io/en/latest/api.html")
@@ -355,8 +358,8 @@ class Calc():
[…]
