### 2024-05-31 ba30f50
Add test to control for warning and error when calling the ff sampling method
