### 2024-05-31 71077f5
Simplify removal of useless variables from output of hdmr method and add more robust check of name of sensensitivity method
