+++ climada/engine/unsequa/calc_base.py
@@ -243,11 +243,11 @@ class Calc():
             'bounds' : [[0, 1]]*len(param_labels)
             }
         #for the ff sampler, no value of N is needed. For API consistency the user
-        #must input a value that is ignored and a warning is given. 
+        #must input a value that is ignored and a warning is given.
         if sampling_method == 'ff':
-            LOGGER.warning("""You are using the 'ff' sampler which does not require
-                           a value for N. The entered N value will be ignored
-                           in the sampling process.""")
+            LOGGER.warning("You are using the 'ff' sampler which does not require "
+                           "a value for N. The entered N value will be ignored"
+                           "in the sampling process.")
         uniform_base_sample = self._make_uniform_base_sample(N, problem_sa,
                                                              sampling_method,
                                                              sampling_kwargs)
@@ -319,7 +319,7 @@ class Calc():
                                     #does not require the N parameter
             if (problem_sa['num_vars'] & (problem_sa['num_vars'] - 1) != 0):
                 raise ValueError("The number of parameters must be a power of 2. "
-                                 "to use the ff sampling method. You can generate "
+                                 "To use the ff sampling method, you can generate "
                                  "dummy parameters to overcome this limitation."
                                  " See https://salib.readthedocs.io/en/latest/api.html")
 
@@ -358,7 +358,7 @@ class Calc():
         unc_output : climada.engine.unsequa.UncOutput
             Uncertainty data object in which to store the sensitivity indices
         sensitivity_method : str, optional
-            Sensitivity analysis method from SALib.analyse. Possible choices: 'sobol', 'fast', 
+            Sensitivity analysis method from SALib.analyse. Possible choices: 'sobol', 'fast',
             'rbd_fast', 'morris', 'dgsm', 'ff', 'pawn', 'rhdm', 'rsa', 'discrepancy', 'hdmr'.
             Note that in Salib, sampling methods and sensitivity
             analysis methods should be used in specific pairs:
@@ -541,32 +541,34 @@ def _calc_sens_df(method, problem_sa, sensitivity_kwargs, param_labels, X, unc_d
                                                     **sensitivity_kwargs)
         #refactor incoherent SALib output
         nparams = len(param_labels)
-        if method.__name__ == 'SALib.analyze.ff':
+        if method.__name__[-3:] == '.ff': #ff method
             if sensitivity_kwargs['second_order']:
                 #parse interaction terms of sens_indices to a square matrix
                 #to ensure consistency with unsequa
                 interaction_names = sens_indices.pop('interaction_names')
                 interactions = np.full((nparams, nparams), np.nan)
+                #loop over interaction names and extract each param pair,
+                #then match to the corresponding param from param_labels
                 for i,interaction_name in enumerate(interaction_names):
                     interactions[param_labels.index(interaction_name[0]),
                                  param_labels.index(interaction_name[1])] = sens_indices['IE'][i]
                 sens_indices['IE'] = interactions
 
-        if method.__name__ == 'SALib.analyze.hdmr':
-            keys_to_remove = ['Em','Term','X','Y'] #need to delete emulator as is a dict
-                                                #and other useless variables
-            [keys_to_remove.append(si) for si, si_val_array in sens_indices.items()
-            if np.array(si_val_array).size > nparams**2]
+        if method.__name__[-5:] == '.hdmr': #hdmr method
+            #first, remove variables that are incompatible with unsequa output
+            keys_to_remove = ['Em','Term','select', 'RT', 'Y_em', 'idx', 'X', 'Y']
[…]
