### 2024-05-31 82b8958
Add prototype of general testing method for all sensitivity methods
