### 2024-06-03 321e5b3
Replace measure with new class ready for first tests
