### 2024-06-03 8505563
Usable new measure with example
