### 2024-06-03 cbf809a
Add changes to tutorial and new cash_flows.py file
