+++ climada/entity/measures/cash_flows.py
@@ -1,0 +1,410 @@
+"""
+This file is part of CLIMADA.
+
+Copyright (C) 2017 ETH Zurich, CLIMADA contributors listed in AUTHORS.
+
+CLIMADA is free software: you can redistribute it and/or modify it under the
+terms of the GNU General Public License as published by the Free
+Software Foundation, version 3.
+
+CLIMADA is distributed in the hope that it will be useful, but WITHOUT ANY
+WARRANTY; without even the implied warranty of MERCHANTABILITY or FITNESS FOR A
+PARTICULAR PURPOSE.  See the GNU General Public License for more details.
+
+You should have received a copy of the GNU General Public License along
+with CLIMADA. If not, see <https://www.gnu.org/licenses/>.
+ 
+---
+
+Define the Cash Flows class.
+"""
+
+# # Example usage
+# import numpy as np
+# from climada.entity import DiscRates
+
+# # Define custom costs and incomes
+# custom_cash_flows = pd.DataFrame({
+#     'year': [2021, 2021, 2023, 2023],
+#     'Cost': [200, 300, 150, 150],
+#     'Income': [600, 700, 800, 900]
+# })
+
+# cost = Cost(mkt_price_year=2020, initial_cash_flow=1000, yearly_cash_flow=100, annual_growth_rate=0.02, custom_cash_flows_df=custom_cash_flows)
+# income = Income(mkt_price_year=2020, initial_cash_flow=0, yearly_cash_flow=500, annual_growth_rate=0.02, custom_cash_flows_df=custom_cash_flows)
+
+# # Define discount rates
+# years = np.arange(1950, 2100)
+# rates = np.ones(years.size) * 0.014
+# rates[51:55] = 0.025
+# rates[95:120] = 0.035
+# disc = DiscRates(years=years, rates=rates)
+
+# # Create an instance of CostIncomeAnalysis
+# analysis = CostIncomeAnalysis(cost, income)
+
+# # Plot the cash flows
+# analysis.plot_cost_and_income(impl_year=2020, start_year=2020, end_year=2040, discounted=False, disc=disc)
+
+# # Calculate the total net present value
+# total_npv = analysis.calc_total(impl_year=2020, start_year=2020, end_year=2040, discounted=False, disc=disc)
+# print("Total Net Present Value:", total_npv)
+
+
+from typing import Optional, Tuple
+from abc import ABC, abstractmethod
+import numpy as np
+import pandas as pd
+import matplotlib.pyplot as plt
+
+class CashFlow(ABC):
+
+    def __init__(self,
+                 mkt_price_year: float,  # year the market prices are given 
+                 initial_cash_flow: float = 0.0,  # initial cash flow (e.g., implementation cost)
+                 yearly_cash_flow: float = 0.0,  # yearly cash flow (e.g., maintenance cost or yearly income)
+                 annual_growth_rate: Optional[float] = 0.0,  # annual growth rate of the cash flow
+                 custom_cash_flows_df: Optional[pd.DataFrame] = None  # custom cash flows per year
+                 ):
+
+        self.annual_growth_rate = annual_growth_rate
+        self.mkt_price_year = mkt_price_year
+        self.initial_cash_flow = initial_cash_flow
+        self.yearly_cash_flow = yearly_cash_flow
+        self.custom_cash_flows_df = custom_cash_flows_df
+
+        if self.custom_cash_flows_df is not None:
+            self.custom_cash_flows_df = self.custom_cash_flows_df.groupby('year').sum()
+            if 'Cost' not in self.custom_cash_flows_df.columns and 'Income' not in self.custom_cash_flows_df.columns:
+                raise ValueError("Custom cash flows DataFrame must contain 'Cost' or 'Income' column")
+
[…]
