### 2024-06-05 708eb41
Fix indentation issue in ff_fail test
