### 2024-06-07 f5cc9de
u_coord: use set_geometry instead of direct column assignment for GeoDataFrames
