+++ climada/util/coordinates.py
@@ -2412,13 +2412,15 @@ def points_to_raster(points_df, val_names=None, res=0.0, raster_res=0.0, crs=DEF
     LOGGER.info('Raster from resolution %s to %s.', res, raster_res)
     df_poly = gpd.GeoDataFrame(points_df[val_names])
     if not scheduler:
-        df_poly['geometry'] = apply_box(points_df)
+        geometry = apply_box(points_df)
     else:
         ddata = dd.from_pandas(points_df[['latitude', 'longitude']],
                                npartitions=cpu_count())
-        df_poly['geometry'] = ddata.map_partitions(apply_box, meta=Polygon) \
+        geometry = ddata.map_partitions(apply_box, meta=Polygon) \
                                    .compute(scheduler=scheduler)
-    df_poly.set_crs(crs if crs else points_df.crs if points_df.crs else DEF_CRS, inplace=True)
+    df_poly.set_geometry(geometry,
+                         crs=crs if crs else points_df.crs if points_df.crs else DEF_CRS, 
+                         inplace=True)
 
     # renormalize longitude if necessary
     if equal_crs(df_poly.crs, DEF_CRS):
@@ -2655,7 +2657,7 @@ def set_df_geometry_points(df_val, scheduler=None, crs=None):
         try:
             crs = df_val.geometry.crs
         except AttributeError:
-            crs = None
+            pass
 
     # work in parallel
     if scheduler:
@@ -2663,18 +2665,16 @@ def set_df_geometry_points(df_val, scheduler=None, crs=None):
             return df_exp.apply(lambda row: Point(row.longitude, row.latitude), axis=1)
 
         ddata = dd.from_pandas(df_val, npartitions=cpu_count())
-        df_val['geometry'] = ddata.map_partitions(
+        geometry = ddata.map_partitions(
                                  apply_point,
                                  meta=('geometry', gpd.array.GeometryDtype)
                              ).compute(scheduler=scheduler)
     # single process
     else:
-        df_val['geometry'] = gpd.GeoSeries(
+        geometry = gpd.GeoSeries(
             gpd.points_from_xy(df_val.longitude, df_val.latitude), index=df_val.index, crs=crs)
 
-    # set crs
-    if crs:
-        df_val.set_crs(crs, inplace=True)
+    df_val.set_geometry(geometry, crs, inplace=True)
 
 
 def fao_code_def():
