### 2024-06-07 5c3eeec
Include other sensitivity methods in generic test method
