+++ climada/engine/unsequa/test/test_unsequa.py
@@ -564,26 +564,105 @@ class TestCalcImpact(unittest.TestCase):
 
         # dict to store the parameters and results for the tests
         test_dict = {
-            'pawn': {
-                'exp_var' : exp_unc,
-                'impf_var' : impf_unc,
-                'haz_var' : haz,
-                'sampling_method' : 'sobol',
-                'N' : 4,
-                'sensitivity_kwargs' : {
-                    'S' : 10,
-                    'seed' : 12345
-                    },
-                'test_param_name' : 'x_exp', #by default always test parameter at fixed position?
-                'test_si_name' : 'CV', #item
-                'test_si_value' : 0.1875 #item
+           'pawn': {
+               'haz_var' : haz,
+               'sampling_method' : 'saltelli',
+               'sampling_kwargs' : {},
+               'N' : 4,
+               'sensitivity_kwargs' : {
+                   'S' : 10,
+                   'seed' : 12345
+                   },
+               'test_param_name' : ['x_exp',0],
+               'test_si_name' : ['CV', 16],
+               'test_si_value' : [0.25000, 2, 6]
+           },
+           'hdmr': {
+               'haz_var' : haz,
+               'sampling_method' : 'saltelli',
+               'sampling_kwargs' : {},
+               'N' : 100,
+               'sensitivity_kwargs' : {},
+               'test_param_name' : ['x_exp', 2],
+               'test_si_name' : ['Sa', 4],
+               'test_si_value' : [0.004658, 3, 6]
+           },
+           'ff': {
+               'haz_var' : haz_unc,
+               'sampling_method' : 'ff',
+               'sampling_kwargs' : {'seed' : 12345},
+               'N' : 4,
+               'sensitivity_kwargs' : {'second_order': True},
+               'test_param_name' : ['x_exp', 0],
+               'test_si_name' : ['IE', 4],
+               'test_si_value' : [865181825.901295, 10, 6]
+           },
+           'sobol': {
+               'haz_var' : haz,
+               'sampling_method' : 'saltelli',
+               'sampling_kwargs' : {},
+               'N' : 4,
+               'sensitivity_kwargs' : {},
+               'test_param_name' : ['x_paa', 5],
+               'test_si_name' : ['ST', 8],
+               'test_si_value' : [0.313025, 10, 6]
+           },
 
-            },}
+            'dgsm': {
+                'haz_var' : haz,
+                'sampling_method' : 'finite_diff',
+                'N' : 4,
+                'sampling_kwargs' : {'seed':12345},
+                'sensitivity_kwargs' : {'num_resamples': 100,
+                                        'conf_level': 0.95, 'seed': 12345},
+                'test_param_name' : ['x_exp',0],
+                'test_si_name' : ['dgsm', 8],
+                'test_si_value' : [1.697516e-01, 9 , 6]#1.697516e-01
+            },
+            'fast': {
+                'haz_var' : haz,
+                'sampling_method' : 'fast_sampler',
+                'sampling_kwargs' : {'M' : 4, 'seed' : 12345},
+                'N' : 256,
+                'sensitivity_kwargs' : {'M': 4, 'seed': 12345},
+                'test_param_name' : ['x_exp',0],
+                'test_si_name' : ['S1_conf',8],
+                'test_si_value' : [0.671396, 1, 6]#0.117101
+            },
+
+            'rbd_fast': {
+                'haz_var' : haz,
+                'sampling_method' : 'saltelli',
+                'sampling_kwargs' : {},
+                'N' : 24,
+                'sensitivity_kwargs' : {'M': 4, 'seed': 12345},
+                'test_param_name' : ['x_exp', 0],
[…]
