### 2024-06-08 44a86cc
Remove useless input vars from test_dict and add morris
