### 2024-06-08 5f4970a
Remove unnecessary error handling for ff in make_sample
