+++ climada/engine/unsequa/test/test_unsequa.py
@@ -562,7 +562,7 @@ class TestCalcImpact(unittest.TestCase):
         #define input_vars
         exp_unc, impf_unc, haz_unc = make_input_vars()
 
-        # dict to store the parameters and results for the tests
+        # dict to store the parameters and expected results for the tests
         test_dict = {
            'pawn': {
                'sampling_method' : 'saltelli',
@@ -683,263 +683,6 @@ class TestCalcImpact(unittest.TestCase):
         # loop over each method and do test
         for sensitivity_method in test_dict:
             test_sensitivity_method(sensitivity_method, test_dict[sensitivity_method])
-
-    def test_calc_sensitivity_pawn_pass(self):
-        """Test compute sensitivity using the 'pawn' method"""
-
-        exp_unc, impf_unc, _ = make_input_vars()
-        haz = haz_dem()
-        unc_calc = CalcImpact(exp_unc, impf_unc, haz)
-        unc_data = unc_calc.make_sample(N=4)  # Generate samples
-        unc_data = unc_calc.uncertainty(unc_data, calc_eai_exp=False, calc_at_event=False)
-
-        # Call the sensitivity method with PAWN-specific arguments
-        unc_data = unc_calc.sensitivity(
-            unc_data,
-            sensitivity_method='pawn',
-            sensitivity_kwargs={'S': 10, 'seed': 12345}  # Arguments relevant to PAWN
-        )
-
-        self.assertEqual(unc_data.sensitivity_method, 'pawn')
-
-        self.assertEqual('x_exp', unc_data.aai_agg_sens_df.param[0])
-        self.assertEqual('CV', unc_data.aai_agg_sens_df.si[12])
-        self.assertEqual(0.1875, unc_data.aai_agg_sens_df.aai_agg[2])
-
-        self.assertEqual(
-            unc_data.aai_agg_unc_df.size,
-            unc_data.n_samples
-            )
-
-        self.assertEqual(
-            unc_data.freq_curve_unc_df.size,
-            unc_data.n_samples * len(unc_calc.rp)
-            )
-        self.assertTrue(unc_data.eai_exp_unc_df.empty)
-        self.assertTrue(unc_data.at_event_unc_df.empty)
-
-    def test_calc_sensitivity_dgsm_pass(self):
-        """Test compute sensitivity using the DGSM method"""
-
-        exp_unc, impf_unc, _ = make_input_vars()
-        haz = haz_dem()
-        unc_calc = CalcImpact(exp_unc, impf_unc, haz)
-
-        # Generate samples using the finite_diff sampling method
-        unc_data = unc_calc.make_sample(N=4, sampling_method='finite_diff')
-        unc_data = unc_calc.uncertainty(unc_data, calc_eai_exp=False, calc_at_event=False)
-
-        unc_data = unc_calc.sensitivity(
-            unc_data,
-            sensitivity_method='dgsm',
-            sensitivity_kwargs={'num_resamples': 100, 'conf_level': 0.95, 'seed': 12345}
-        )
-
-        self.assertEqual(unc_data.sensitivity_method, 'dgsm')
-
-        self.assertEqual('x_exp', unc_data.freq_curve_sens_df.param[0])
-        self.assertEqual('dgsm', unc_data.freq_curve_sens_df.si[8])
-        self.assertAlmostEqual(0.060215, unc_data.freq_curve_sens_df.rp100[6], places=6)
-
-        self.assertEqual(
-            unc_data.aai_agg_unc_df.size,
-            unc_data.n_samples
-            )
-
-        self.assertEqual(
-            unc_data.freq_curve_unc_df.size,
-            unc_data.n_samples * len(unc_calc.rp)
-            )
[…]
