### 2024-06-10 d4bd9fd
Remove all old code from measures and measure set
