+++ climada/entity/measures/measure_set.py
@@ -21,70 +21,16 @@ Define MeasureSet class.
 
 __all__ = ['MeasureSet']
 
-import ast
-import copy
 import logging
-from typing import Optional, List
+from typing import Iterable
 
-from  matplotlib import colormaps as cm
-import numpy as np
-import pandas as pd
-import xlsxwriter
 
 from climada.entity.measures.base import Measure
-import climada.util.hdf5_handler as u_hdf5
 
 LOGGER = logging.getLogger(__name__)
 
-DEF_VAR_MAT = {'sup_field_name': 'entity',
-               'field_name': 'measures',
-               'var_name': {'name': 'name',
-                            'color': 'color',
-                            'cost': 'cost',
-                            'haz_int_a': 'hazard_intensity_impact_a',
-                            'haz_int_b': 'hazard_intensity_impact_b',
-                            'haz_frq': 'hazard_high_frequency_cutoff',
-                            'haz_set': 'hazard_event_set',
-                            'mdd_a': 'MDD_impact_a',
-                            'mdd_b': 'MDD_impact_b',
-                            'paa_a': 'PAA_impact_a',
-                            'paa_b': 'PAA_impact_b',
-                            'fun_map': 'damagefunctions_map',
-                            'exp_set': 'assets_file',
-                            'exp_reg': 'Region_ID',
-                            'risk_att': 'risk_transfer_attachement',
-                            'risk_cov': 'risk_transfer_cover',
-                            'haz': 'peril_ID'
-                           }
-              }
-"""MATLAB variable names"""
-
-DEF_VAR_EXCEL = {'sheet_name': 'measures',
-                 'col_name': {'name': 'name',
-                              'color': 'color',
-                              'cost': 'cost',
-                              'haz_int_a': 'hazard intensity impact a',
-                              'haz_int_b': 'hazard intensity impact b',
-                              'haz_frq': 'hazard high frequency cutoff',
-                              'haz_set': 'hazard event set',
-                              'mdd_a': 'MDD impact a',
-                              'mdd_b': 'MDD impact b',
-                              'paa_a': 'PAA impact a',
-                              'paa_b': 'PAA impact b',
-                              'fun_map': 'damagefunctions map',
-                              'exp_set': 'assets file',
-                              'exp_reg': 'Region_ID',
-                              'risk_att': 'risk transfer attachement',
-                              'risk_cov': 'risk transfer cover',
-                              'risk_fact': 'risk transfer cost factor',
-                              'haz': 'peril_ID'
-                             }
-                }
-"""Excel variable names"""
-
 class MeasureSet():
-    """Contains measures of type Measure. Loads from
-    files with format defined in FILE_EXT.
+    """Contains measures of type Measure.
 
     Attributes
     ----------
@@ -95,50 +41,20 @@ class MeasureSet():
 
     def __init__(
         self,
-        measure_list: Optional[List[Measure]] = None
+        measures: Iterable[Measure]
     ):
         """Initialize a new MeasureSet object with specified data.
 
         Parameters
         ----------
-        measure_list : list of Measure objects, optional
+        measures : iterable of Measure objects.
             The measures to include in the MeasureSet
 
-        Examples
-        --------
-        Fill MeasureSet with values and check consistency data:
-
-        >>> act_1 = Measure(
-        ...     name='Seawall',
-        ...     color_rgb=np.array([0.1529, 0.2510, 0.5451]),
-        ...     hazard_intensity=(1, 0),
-        ...     mdd_impact=(1, 0),
-        ...     paa_impact=(1, 0),
-        ... )
-        >>> meas = MeasureSet([act_1])
-        >>> meas.check()
-
-        Read measures from file and checks consistency data:
-
[…]
