+++ climada/entity/measures/base.py
@@ -23,8 +23,7 @@ __all__ = ['Measure']
 
 import copy
 import logging
-from pathlib import Path
-from typing import Optional, Tuple, Callable
+from typing import Callable
 
 import numpy as np
 
@@ -35,7 +34,7 @@ from climada.engine import ImpactCalc
 
 LOGGER = logging.getLogger(__name__)
 
-##todo: risk transfer, change hazard/exposures completely
+##todo: risk transfer, change hazard/exposures/impfset completely
 class Measure():
 
     def __init__(
@@ -60,6 +59,54 @@ class Measure():
     @property
     def end_year(self):
         return self.years[1]
+
+    def apply_to_exposures(self, exposures):
+        """
+        Implement measure to exposures.
+
+        Parameters
+        ----------
+        exposures : climada.entity.Exposures
+            exposures instance
+
+        Returns
+        -------
+        new_exp : climada.entity.Exposure
+            Exposure with implemented measure with all defined parameters
+        """
+        return self.exp_map(exposures)
+
+    def apply_to_impfset(self, impfset):
+        """
+        Implement measure to impact function set
+
+        Parameters
+        ----------
+        impfset : climada.entity.ImpactFuncSet
+            impact function set instance
+
+        Returns
+        -------
+        new_impfset : climada.entity.ImpactFuncSet
+            Impact function set with implemented measure with all defined parameters
+        """
+        return self.impfset_map(impfset)
+
+    def apply_to_hazard(self, hazard):
+        """
+        Implement measure to hazard.
+
+        Parameters
+        ----------
+        hazard : climada.hazard.Hazard
+            hazard instance
+
+        Returns
+        -------
+        new_haz : climada.hazard.Hazard
+            Hazard with implemented measure with all defined parameters
+        """
+        return self.haz_map(hazard)
 
     def apply(self, exposures, impfset, hazard):
         """
@@ -92,15 +139,9 @@ class Measure():
 
         return new_exp, new_impfs, new_haz
 
-    def impact(
-            self, exposures, impfset, hazard,
-            save_mat=False, assign_centroids=True
-            ):
+    def impact(self, exposures, impfset, hazard,**kwargs):
         meas_exp, meas_impfset, meas_haz = self.apply(exposures, impfset, hazard)
-        impact = ImpactCalc(meas_exp, meas_impfset, meas_haz)\
-              .impact(save_mat=save_mat, assign_centroids=assign_centroids)
-        return impact
-
+        return ImpactCalc(meas_exp, meas_impfset, meas_haz).impact(**kwargs)
 
 def helper_hazard(
         intensity_multiplier=1, intensity_substract=0
@@ -181,502 +222,3 @@ def helper_exposure(
         exp_modified.gdf['value'][remove_values] = 0
         return exp_modified
     return exposures_change
-
-
-
-
-'''
-IMPF_ID_FACT = 1000
-"""Factor internally used as id for impact functions when region selected."""
-
-NULL_STR = 'nil'
-"""String considered as no path in measures exposures_set and hazard_set or
-no string in imp_fun_map"""
-
-class Measure():
-    """
-    Contains the definition of one measure.
-
-    Attributes
-    ----------
-    name : str
-        name of the measure
-    haz_type : str
-        related hazard type (peril), e.g. TC
-    color_rgb : np.array
-        integer array of size 3. Color code of this measure in RGB
-    cost : float
-        discounted cost (in same units as assets)
-    hazard_set : str
-        file name of hazard to use (in h5 format)
-    hazard_freq_cutoff : float
-        hazard frequency cutoff
-    exposures_set : str  or climada.entity.Exposure
-        file name of exposure to use (in h5 format) or Exposure instance
-    imp_fun_map : str
[…]
