### 2024-06-10 e1735df
Fix few linter issues
