### 2024-06-10 29e3b83
Add test for computation of absolute delta climate
