### 2024-06-10 ffea8e9
fixes uncomplete transition to dataclass snapshots
