+++ climada/engine/impact_trajectories.py
@@ -22,6 +22,7 @@ import copy
 from dataclasses import dataclass
 
 import numpy as np
+import itertools
 from datetime import datetime
 from scipy.sparse import lil_matrix
 
@@ -33,6 +34,12 @@ from climada.engine.impact_calc import ImpactCalc
 
 
 ### Utils functions
+def pairwise(iterable):
+    "s -> (s0, s1), (s1, s2), (s2, s3), ..."
+    a, b = itertools.tee(iterable)
+    next(b, None)
+    return zip(a, b)
+
 
 def get_dates(haz: Hazard):
     return [datetime.fromordinal(date) for date in haz.date]
@@ -74,6 +81,7 @@ def interpolate_years(year_start, year_end):
     values = np.linspace(0, 1, num=year_end - year_start + 1)
     return values
 
+
 def bayesian_viktypliers(year0, year1):
     prop_H1 = interpolate_years(year0, year1)
     prop_H0 = 1 - prop_H1
@@ -83,7 +91,7 @@ def bayesian_viktypliers(year0, year1):
 def snapshot_combinaisons(snapshot0, snapshot1):
     impfset0 = snapshot0.impfset
     impfset1 = snapshot1.impfset
-    assert impfset0 is impfset1 # We don't allow for different impfset
+    assert impfset0 is impfset1  # We don't allow for different impfset
 
     exp_y0 = snapshot0.exposure
     exp_y1 = snapshot1.exposure
@@ -103,6 +111,7 @@ def snapshot_combinaisons(snapshot0, snapshot1):
 
     return imp_E0H0, imp_E1H0, imp_E0H1, imp_E1H1
 
+
 # Would it be better to have a Snapshot Class that hold a singular snapshot?
 @dataclass
 class Snapshot:
@@ -118,9 +127,10 @@ class SnapshotsCollection:
         self.hazard_set = hazard_set
         self.impfset = impfset
         self.snapshots_years = snapshot_years
-        self.data_dict = {}
-        for year in self.snapshots_years:
-            self.data_dict[year] = Snapshot(exposure_set[year], hazard_set[year], impfset, year)
+        self.data = [
+            Snapshot(exposure_set[year], hazard_set[year], impfset, year)
+            for year in snapshot_years
+        ]
 
     # Check that at least first and last snap are complete
     # and otherwise it is ok
@@ -159,10 +169,11 @@ class CalcImpactsSnapshots:
 
     def calc_impacts_snapshots(self):
         impacts_list = {}
-        for year, [exp, haz] in self.snapshots.data_dict.items():
-            impacts_list[year] = ImpactCalc(exp, self.snapshots.impfset, haz).impact()
+        for snapshot in self.snapshots.data:
+            impacts_list[snapshot.year] = ImpactCalc(
+                snapshot.exposure, self.snapshots.impfset, snapshot, hazard
+            ).impact()
         return impacts_list
-
 
     # Calculate the eai_exp for each year
     def interpolate_eai_exp(self, imp0, imp1, start_year, end_year, frequency):
@@ -176,20 +187,20 @@ class CalcImpactsSnapshots:
             yearly_eai_exp.append(eai_exp)
         return yearly_eai_exp
 
-    def bayesian_mixer(self, start_year, end_year):
+    def bayesian_mixer(self, start_snapshot, end_snapshot):
         # 1. Interpolate in between years
-        prop_H0, prop_H1 = bayesian_viktypliers(start_year, end_year)
-        imp_E0H0, imp_E1H0, imp_E0H1, imp_E1H1 = (
-            snapshot_combinaisons(start_year, end_year)
+        prop_H0, prop_H1 = bayesian_viktypliers(start_snapshot.year, end_snapshot.year)
[…]
