### 2024-06-10 ece0b07
fixes typo in ImpactCalc
