### 2024-06-11 fdb7d8a
Remove unnecessary lines in _calc_sens_df
