### 2024-06-11 3eb3fc5
Minor code improvements in generic sensitivity test function
