### 2024-06-12 10854c4
Add year to measures apply method
