### 2024-06-12 677b024
Add combined measures
