### 2024-06-13 ca8338d
Format docstring of TropCyclone.from_tracks
