+++ climada/hazard/trop_cyclone.py
@@ -201,9 +201,9 @@ class TropCyclone(Hazard):
         Create new TropCyclone instance that contains windfields from the specified tracks.
 
         This function sets the ``intensity`` attribute to contain, for each centroid,
-        the maximum wind speed (1-minute sustained winds at 10 meters above ground) experienced
-        over the whole period of each TC event in m/s. The wind speed is set to 0 if it doesn't
-        exceed the threshold ``intensity_thres``.
+        the maximum wind speed (1-minute sustained winds at 10 meters above ground)
+        experienced over the whole period of each TC event in m/s. The wind speed is set
+        to 0 if it doesn't exceed the threshold ``intensity_thres``.
 
         The ``category`` attribute is set to the value of the ``category``-attribute
         of each of the given track data sets.
@@ -219,90 +219,96 @@ class TropCyclone(Hazard):
         tracks : climada.hazard.TCTracks
             Tracks of storm events.
         centroids : Centroids, optional
-            Centroids where to model TC. Default: global centroids at 360 arc-seconds resolution.
+            Centroids where to model TC. Default: global centroids at 360 arc-seconds
+            resolution.
         pool : pathos.pool, optional
-            Pool that will be used for parallel computation of wind fields. Default: None
+            Pool that will be used for parallel computation of wind fields. Default:
+            None
         model : str, optional
-            Parametric wind field model to use:
-            "H1980" (the prominent Holland 1980 model) from the paper:
-                Holland, G.J. (1980): An Analytic Model of the Wind and Pressure Profiles
-                in Hurricanes. Monthly Weather Review 108(8): 1212–1218.
-                https://doi.org/10.1175/1520-0493(1980)108<1212:AAMOTW>2.0.CO;2,
+            Parametric wind field model to use. Default: "H08".
 
-            "H08" (Holland 1980 with b-value from Holland 2008) from the paper:
-                Holland, G. (2008). A revised hurricane pressure-wind model. Monthly
-                Weather Review, 136(9), 3432–3445. https://doi.org/10.1175/2008MWR2395.1,
-            "H10" (Holland et al. 2010) from the paper:
-                Holland et al. (2010): A Revised Model for Radial Profiles of Hurricane Winds. Monthly
-                Weather Review 138(12): 4393–4401. https://doi.org/10.1175/2010MWR3317.1,
-            or
-            "ER11" (Emanuel and Rotunno 2011) from the paper:
-                Emanuel, K., Rotunno, R. (2011): Self-Stratification of Tropical Cyclone Outflow. Part I:
-                Implications for Storm Structure. Journal of the Atmospheric Sciences 68(10): 2236–2249.
-                https://dx.doi.org/10.1175/JAS-D-10-05024.1.
-            Default: "H08".
+            * ``"H1980"`` (the prominent Holland 1980 model) from the paper:
+                    Holland, G.J. (1980): An Analytic Model of the Wind and Pressure
+                    Profiles in Hurricanes. Monthly Weather Review 108(8): 1212–1218.
+                    ``https://doi.org/10.1175/1520-0493(1980)108<1212:AAMOTW>2.0.CO;2``
+            * ``"H08"`` (Holland 1980 with b-value from Holland 2008) from the paper:
+                    Holland, G. (2008). A revised hurricane pressure-wind model. Monthly
+                    Weather Review, 136(9), 3432–3445.
+                    https://doi.org/10.1175/2008MWR2395.1
+            * ``"H10"`` (Holland et al. 2010) from the paper:
+                    Holland et al. (2010): A Revised Model for Radial Profiles of
+                    Hurricane Winds. Monthly Weather Review 138(12): 4393–4401.
[…]
