### 2024-06-13 69961ee
Add pre-commit to dependencies with appropriate config
