### 2024-06-14 #6755 @peanutfun
@sunt05 @naingeet We are looking forward to receiving your feedback on the code and the paper draft in https://github.com/CLIMADA-project/climada_python/pull/692. If you have any trouble installing CLIMADA or executing the provided tutorial Jupyter script, please let us know.
