### 2024-06-14 4f6952e
Ensure that csr_matrices are in canonical format
* Add getter/setter attributes for csr_matrices in Hazard.
* Check format and sum duplicates when assigning matrices.
* Update unit tests.
