### 2024-06-14 91c52e9
Ensure canonical format in init and streamline checks
