### 2024-06-14 3e27ac8
Explicitly remove zeros from csr matrices when assigning
