### 2024-06-14 0465790
added tests for new functions
