+++ climada/test/test_hazard.py
@@ -24,6 +24,8 @@ import numpy as np
 import datetime as dt
 from pathlib import Path
 from scipy import sparse
+from rioxarray import open_rasterio
+from xarray import load_dataset
 
 from climada import CONFIG
 from climada.hazard import tc_tracks as tc
@@ -188,6 +190,40 @@ class TestCentroids(unittest.TestCase):
 
         DATA_DIR.joinpath(intensity_file).unlink()
         DATA_DIR.joinpath(fraction_file).unlink()
+        
+    def test_write_raster_local_exceedance_inten(self):
+        """Test write TIFF file from local exceedance intensity"""
+        haz = Hazard.from_hdf5(HAZ_TEST_TC)
+        haz.write_raster_local_exceedance_inten([10, 20, 30, 40], filename = f'{CONFIG.test_data.dir()}/test_write_raster_local_exceedance_inten.tif')
+        dataarray = open_rasterio(f'{CONFIG.test_data.dir()}/test_write_raster_local_exceedance_inten.tif')
+        np.testing.assert_array_almost_equal(
+            dataarray.data[[0, 1, 2, 3], [0, 8, 3, 6], [2, 5, 7, 1]],
+            np.array([35.829193, 37.64515 , 61.62324 , 58.595936])
+        )
+        np.testing.assert_array_equal(dataarray.data.shape, (4, 10, 10))
+
+    def test_write_raster_local_return_periods(self):
+        """Test write TIFF file from local exceedance intensity"""
+        haz = Hazard.from_hdf5(HAZ_TEST_TC)
+        haz.write_raster_local_return_periods([10., 20., 30.], filename = f'{CONFIG.test_data.dir()}/test_write_raster_local_return_periods.tif')
+        dataarray = open_rasterio(f'{CONFIG.test_data.dir()}/test_write_raster_local_return_periods.tif')
+        np.testing.assert_array_almost_equal(
+            dataarray.data[[0, 1, 2], [2, 5, 8], [1, 6, 7]],
+            np.array([2.42105263, 2.59677419, 4.01496259])
+        )
+        np.testing.assert_array_equal(dataarray.data.shape, (3, 10, 10))
+
+    def test_write_netcdf_local_return_periods(self):
+        """Test write netcdf file from local exceedance intensity"""
+        haz = Hazard.from_hdf5(HAZ_TEST_TC)
+        haz.write_netcdf_local_return_periods([10., 20., 30.], filename = f'{CONFIG.test_data.dir()}/test_write_netcdf_local_return_periods.nc')
+        dataset = load_dataset(f'{CONFIG.test_data.dir()}/test_write_netcdf_local_return_periods.nc')
+        np.testing.assert_array_almost_equal(
+            dataset.to_array().data[[0, 1, 2, 3, 4], [21, 45, 86, 65, 7]],
+            np.array([22., -80. ,1.483871, 3.108108, 5.366667])
+        )
+        np.testing.assert_array_equal(dataset.to_array().data.shape, (5, 100))
+
 
 
 
