### 2024-06-14 405f50a
Add isort to pre-commit configuration
