### 2024-06-17 99d4bf5
Support reading non-string impact event names from H5
