+++ climada/engine/test/test_impact.py
@@ -1120,6 +1120,15 @@ class TestImpactH5IO(unittest.TestCase):
         impact = Impact.from_hdf5(self.filepath)
         npt.assert_array_equal(impact.imp_mat.toarray(), [[0, 1, 2], [3, 0, 0]])
 
+        # Check with non-string event_name
+        event_name = [1.2, 2]
+        with h5py.File(self.filepath, "r+") as file:
+            del file["event_name"]
+            file.create_dataset("event_name", data=event_name)
+        with self.assertLogs("climada.engine.impact", "WARNING") as cm:
+            impact = Impact.from_hdf5(self.filepath)
+        self.assertIn("'event_name' could not be decoded to a string", cm.output[0])
+        self.assertListEqual(impact.event_name, event_name)
 
 # Execute Tests
 if __name__ == "__main__":
