### 2024-06-17 4ced09c
Add explicit check for updating matrices
