### 2024-06-17 e9d5f1d
add hazard crs for testing haz.concat
