### 2024-06-17 dc56284
add test for haz.cocat for failing crs
