### 2024-06-17 de06f2e
Add util function for pruning csr_matrices
Update hazard docstring
