### 2024-06-17 ebb2cc8
Check matrices when instantiating ImpactCalc
