### 2024-06-17 a3279d4
include admin-0 in plot function
