+++ climada/hazard/centroids/centr.py
@@ -35,6 +35,7 @@ import pandas as pd
 from pyproj.crs.crs import CRS
 import rasterio
 from shapely.geometry.point import Point
+import cartopy.feature as cfeature
 
 from climada.util.constants import DEF_CRS
 import climada.util.coordinates as u_coord
@@ -477,48 +478,9 @@ class Centroids():
               (self.lat >= lat_min) & (self.lat <= lat_max)
             )
         return sel_cen
-
-    #TODO replace with nice GeoDataFrame util plot method.
-    def plot_deprecated(self, axis=None, figsize=(9, 13), **kwargs):
-        """Plot centroids scatter points over earth
-
-        Parameters
-        ----------
-        axis : matplotlib.axes._subplots.AxesSubplot, optional
-            axis to use
-        figsize: (float, float), optional
-            figure size for plt.subplots
-            The default is (9, 13)
-        kwargs : optional
-            arguments for scatter matplotlib function
-
-        Returns
-        -------
-        axis : matplotlib.axes._subplots.AxesSubplot
-        """
-        pad = np.abs(u_coord.get_resolution(self.lat, self.lon)).min()
-
-        proj_data, _ = u_plot.get_transformation(self.crs)
-        proj_plot = proj_data
-        if isinstance(proj_data, ccrs.PlateCarree):
-            # use different projections for plot and data to shift the central lon in the plot
-            xmin, ymin, xmax, ymax = u_coord.latlon_bounds(self.lat, self.lon, buffer=pad)
-            proj_plot = ccrs.PlateCarree(central_longitude=0.5 * (xmin + xmax))
-        else:
-            xmin, ymin, xmax, ymax = (self.lon.min() - pad, self.lat.min() - pad,
-                                      self.lon.max() + pad, self.lat.max() + pad)
-
-        if not axis:
-            _fig, axis, _fontsize = u_plot.make_map(proj=proj_plot, figsize=figsize)
-
-        axis.set_extent((xmin, xmax, ymin, ymax), crs=proj_data)
-        u_plot.add_shapes(axis)
-        axis.scatter(self.lon, self.lat, transform=proj_data, **kwargs)
-        plt.tight_layout()
-        return axis
     
     def plot(self, figsize=(9, 13), *args, **kwargs):
-        """Plot centroids geodataframe using geopandas plotting function
+        """Plot centroids geodataframe using geopandas and cartopy plotting functions
 
         Parameters
         ----------
@@ -526,15 +488,23 @@ class Centroids():
             figure size for plt.subplots
             The default is (9, 13)
         args : optional
-            positional arguments for geopandas plot function
+            positional arguments for geopandas.GeoDataFrame.plot
         kwargs : optional
-            keyword arguments for geopandas plot function
+            keyword arguments for geopandas.GeoDataFrame.plot
             
         Returns
         -------
-        ax : matplotlib.axes instance
+        ax : cartopy.mpl.geoaxes.GeoAxes instance
         """
-        return self.gdf.plot(figsize, *args, **kwargs)
+        
+        fig, ax = plt.subplots(figsize=figsize, subplot_kw={"projection": ccrs.PlateCarree()})
+
+        ax.add_feature(cfeature.BORDERS)
+        ax.add_feature(cfeature.COASTLINE)
+        ax.gridlines(draw_labels=True, dms=True, x_inline=False, y_inline=False)
+        self.gdf.plot(ax=ax, *args, **kwargs)
+
+        return ax
 
     def set_region_id(self, level='country', overwrite=False):
         """Set region_id as country ISO numeric code attribute for every pixel or point.
