### 2024-06-18 8a792ec
Replace assertRaises with assertRaisesRegex
