### 2024-06-18 4d438c7
Update the assertRaises to assertRaisesRegex for test_concat_raise_value_error.
