### 2024-06-18 1a3bcb0
bugfix for test_concat_raise_value_error
