### 2024-06-18 770e941
Instantiate centroids in Hazard.concat() with CRS from the first hazard (#881)
* add test for Hazard.concat for failing crs
* Update error message for Centroids.append
