### 2024-06-18 24e27fb
copy and reproject gdf to DEF_CRS before plotting
