### 2024-06-18 124af45
clean up according to pylint
