### 2024-06-19 fcb406f
Added the costIncome to Measure and updated the steps in the NPV tutorial
