+++ climada/entity/measures/cost_income.py
@@ -1,0 +1,267 @@
+"""
+This file is part of CLIMADA.
+
+Copyright (C) 2017 ETH Zurich, CLIMADA contributors listed in AUTHORS.
+
+CLIMADA is free software: you can redistribute it and/or modify it under the
+terms of the GNU General Public License as published by the Free
+Software Foundation, version 3.
+
+CLIMADA is distributed in the hope that it will be useful, but WITHOUT ANY
+WARRANTY; without even the implied warranty of MERCHANTABILITY or FITNESS FOR A
+PARTICULAR PURPOSE.  See the GNU General Public License for more details.
+
+You should have received a copy of the GNU General Public License along
+with CLIMADA. If not, see <https://www.gnu.org/licenses/>.
+ 
+---
+
+Define the Cash Flows class.
+"""
+
+
+# Define default discount rates
+# DISC_RATES = DiscRates(years=np.arange(1900, 2100), rates=np.ones(np.arange(1900, 2100).size))
+
+
+from typing import Optional, Tuple
+import numpy as np
+import pandas as pd
+import matplotlib.pyplot as plt
+from datetime import datetime
+
+
+
+class CostIncome:
+
+    def __init__(self,
+                 mkt_price_year: float = datetime.today().year, # Get the year of today as integer
+                 cost_growth_rate: float = 0.0,
+                 init_cost: float = 0.0,
+                 annual_cost: float = 0.0,
+                 annual_income: float = 0.0,
+                 income_growth_rate: float = 0.0,
+                 custom_cash_flows_df: Optional[pd.DataFrame] = None):
+        
+        self.mkt_price_year = mkt_price_year
+        # Cost parameters
+        self.cost_growth_rate = cost_growth_rate
+        self.init_cost = -abs(init_cost)
+        self.annual_cost = -abs(annual_cost)
+        # Income parameters
+        self.annual_income = annual_income
+        self.income_growth_rate = income_growth_rate
+        # Custom cash flows
+        # Updaet the cost columns to be negative
+        if custom_cash_flows_df is not None and 'cost' in custom_cash_flows_df.columns:
+                custom_cash_flows_df['cost'] = -abs(custom_cash_flows_df['cost'])
+        self.custom_cash_flows_df = custom_cash_flows_df
+
+        # Custom cash flows
+        if self.custom_cash_flows_df is not None:
+            self.custom_cash_flows_df = self.custom_cash_flows_df.groupby('year').sum()
+            if 'cost' not in self.custom_cash_flows_df.columns and 'income' not in self.custom_cash_flows_df.columns:
+                raise ValueError("Custom cash flows DataFrame must contain 'cost' or 'income' column")
+
+    def _calc_cash_flow_at_year(self, impl_year, year):
+        '''
+        Calculate the net cash flow at a specific year
+
+        Parameters:
+        -----------
+        impl_year: int
+            the year the measure is implemented
+        year: int
+            the year to calculate the cash flow
+
+        Returns:
+        --------
+        Tuple[float, float, float]
+            the net cash flow, cost, and income at the given year
+        '''
+
+        # Calculate the increased cost and income multipliers
+        cost_incr_multi = (1 + self.cost_growth_rate)**(year - self.mkt_price_year)
+        income_incr_multi = (1 + self.income_growth_rate)**(year - self.mkt_price_year)
+
+        # Calculate the cash flows at the given year
+        if year < impl_year:
+            cost = 0
+            income = 0
+        elif year == impl_year:
+            cost = self.init_cost * cost_incr_multi
+            income = self.annual_income * income_incr_multi
+        else:
+            cost = self.annual_cost * cost_incr_multi
+            income = self.annual_income * income_incr_multi
+
+        # Custom cash flows
+        custom_cost = self._get_custom_cash_flow(year, 'cost')
[…]

[…]
