### 2024-06-20 8709a71
added tutorial and better tests
