+++ climada/hazard/test/test_base.py
@@ -1013,15 +1013,21 @@ class TestStats(unittest.TestCase):
 
     def test_local_return_period(self):
         """Compare local return periods against reference."""
-        haz = Hazard.from_hdf5(HAZ_TEST_TC)
-        threshold_intensities = np.array([10., 20.])
+        haz = dummy_hazard()
+        haz.intensity = sparse.csr_matrix([[1.1, 2.1, 3.1],
+                                   [0.1, 3.1, 2.1],
+                                   [4.1, 1.1, 0.1],
+                                   [2.1, 3.1, 3.1]])
+        haz.frequency = np.full(4, .5)
+        threshold_intensities = np.array([1., 2., 4.])
         return_stats = haz.local_return_period(threshold_intensities)
         np.testing.assert_allclose(
-            return_stats[:, [15,54,87]],
-            np.array(
-                [[1.82746879, 2.22375691, 1.31000814],
-                [3.33333333, 3.87019231, 2.34693878]]
-                )
+            return_stats,
+            np.array([
+                [0.66666667, 0.5, 0.66666667],
+                [1., 0.66666667, 0.66666667],
+                [2., np.nan, np.nan]
+                ])
         )
 
 class TestYearset(unittest.TestCase):
