### 2024-06-20 1e91def
added new functions
