+++ climada/hazard/io.py
@@ -27,6 +27,7 @@ import pathlib
 from typing import Union, Optional, Callable, Dict, Any
 
 import h5py
+import netCDF4 as nc
 import numpy as np
 import pandas as pd
 import rasterio
@@ -1037,6 +1038,128 @@ class HazardIO():
                             var_name, var_val.__class__.__name__, var_name
                         )
         self.centroids.write_hdf5(file_name, mode='a')
+
+    def write_raster_local_exceedance_inten(self, return_periods, filename):
+        """
+        Generates exceedance intensity data for specified return periods and 
+        saves it into a GeoTIFF file.
+    
+        Parameters
+        ----------
+        return_periods : np.array or list
+            Array or list of return periods (in years) for which to calculate 
+            and store exceedance intensities.
+        filename : str
+            Path and name of the file to write in tif format.
+        """
+        inten_stats = self.local_exceedance_inten(return_periods=return_periods)
+        num_bands = inten_stats.shape[0]
+        
+        if not self.centroids.get_meta():
+            raise ValueError("centroids.get_meta() is required but not set.")
+
+        ### this code is to replace pixel_geom = self.centroids.calc_pixels_polygons()
+        if abs(abs(self.centroids.get_meta()['transform'].a) -
+               abs(self.centroids.get_meta()['transform'].e)) > 1.0e-5:
+            raise ValueError('Area can not be computed for not squared pixels.')
+        pixel_geom = self.centroids.geometry.buffer(self.centroids.get_meta()['transform'].a / 2).envelope
+        ###
+        profile = self.centroids.get_meta().copy()
+        profile.update(driver='GTiff', dtype='float32', count=num_bands)
+        
+        with rasterio.open(filename, 'w', **profile) as dst:
+            LOGGER.info('Writing %s', filename)
+            for band in range(num_bands):
+                raster = rasterio.features.rasterize(
+                    [(x, val) for (x, val) in zip(pixel_geom, inten_stats[band].reshape(-1))],
+                    out_shape=(profile['height'], profile['width']),
+                    transform=profile['transform'], fill=0,
+                    all_touched=True, dtype=profile['dtype'])
+                dst.write(raster, band + 1)
+                
+                band_name = f"Exceedance intensity for RP {return_periods[band]} years"
+                dst.set_band_description(band + 1, band_name)   
+                
+    def write_raster_local_return_periods(self, hazard_intensities, filename, output_resolution=None):
+        """Write local return periods map as GeoTIFF file.
+    
+        Parameters
+        ----------
+        hazard_intensities: np.array
+            Hazard intensities to consider for calculating return periods.
+        filename: str
+            File name to write in tif format.
+        output_resolution: int
+            If not None, the data is rasterized to this resolution.
+            Default is None (resolution is estimated from the data).
+        """
+        
+        # Calculate the local return periods
+        variable = self.local_return_period(hazard_intensities)
+        
+        # Obtain the meta information for the raster file
+        meta = self.centroids.get_meta(resolution=output_resolution)
+        meta.update(driver='GTiff', dtype=rasterio.float32, count=len(hazard_intensities))
+        res = meta["transform"][0]  # resolution from lon coordinates
+    
+        if meta['height'] * meta['width'] == self.centroids.size:
+            # centroids already in raster format
+            u_coord.write_raster(filename, variable, meta)
+        else:
+            geometry = self.centroids.get_pixel_shapes(res=res)
+            with rasterio.open(filename, 'w', **meta) as dst:
+                LOGGER.info('Writing %s', filename)
+                for i_ev in range(len(hazard_intensities)):
+                    raster = rasterio.features.rasterize(
+                        (
+                            (geom, value)
+                            for geom, value
[…]
