### 2024-06-20 #6755 @sunt05
Thanks @peanutfun for the nice work!

Could you please address [this issue](https://github.com/CLIMADA-project/climada_python/issues/899) while I was testing your notebook?
