### 2024-06-20 81dc151
added unit test for new functions
