### 2024-06-20 f5c3d4e
change from rioxarray to rasterio in testing
