### 2024-06-20 CLIMADA-project/climada_python#899 @emanuel-schmid
Thanks for reporting this @sunt05.
In order to run this notebook you will eventually have to switch to the "advanced installation" method, as you will need to install climada from sources (from the calibrate-impact-function branch, to be more precise).

However failing at the `import climada` step cannot be attributed to the "simple installation" approach!
Usually this happens because the ipykernel is wrong. Have you checked that?
