### 2024-06-21 c588124
added test for non-raster data and new variable name
