### 2024-06-21 086aee2
loosens pre-commit python version requirement
