### 2024-06-21 CLIMADA-project/climada_python#899 @peanutfun
@sunt05 This looks like a mismatch between the versions of `dask` and `pandas`. As @emanuel-schmid already mentioned, you need to follow the [Advanced Installation Instructions](https://climada-python.readthedocs.io/en/latest/guide/install.html#advanced-instructions) for running the script from #692. In step 4, replace
```
git checkout develop
```
with
```
git checkout calibrate-impact-functions
```
and continue with the instructions. Following them will *not* install Jupyter into the environment, so you can then choose to [install Jupyterlab](https://climada-python.readthedocs.io/en/latest/guide/install.html#jupyterlab) or install the more lightweight Jupyter with
```
mamba install -n climada_env jupyter
```
You should then be ready to execute the tutorial. In any case, you should choose the Conda environment of Climada as Kernel when you run Jupyter.

Please report back if the problem persists if you follow these instructions. In this case, we might need to update the environment specifications.
