### 2024-06-21 3537206
Fix links in pull request template
