### 2024-06-21 0711526
remove setting centroids.geometry in plot funciton
