### 2024-06-21 c58b48d
Update climada/hazard/base.py
Co-authored-by: Chahan M. Kropf <chahan.kropf@usys.ethz.ch>
