### 2024-06-21 e0ebdef
Format docstring suggestion from review
