### 2024-06-21 71ce00e
Update unit tests for matrix pruning and checks
