### 2024-06-21 48c44cc
Issue debug message if Holland 2010 model is called with cyclostrophic=False
