### 2024-06-21 4b48511
WIP: Introduce a second version of the Holland 2010 model (#846)
* Introduce a second version of the Holland 2010 model
* TropCyclone: implement model_kwargs
* Update unit tests
* Update CHANGELOG.md

---------

Co-authored-by: Thomas Roosli <thomas.roeoesli@meteoswiss.ch>
Co-authored-by: Thomas Vogt <thomas.vogt@pik-potsdam.de>
Co-authored-by: Lukas Riedel <34276446+peanutfun@users.noreply.github.com>
