+++ climada/hazard/test/test_trop_cyclone.py
@@ -33,8 +33,8 @@ from climada.test import get_test_file
 from climada.hazard.tc_tracks import TCTracks
 from climada.hazard.trop_cyclone import (
     TropCyclone, get_close_centroids, _vtrans, _B_holland_1980, _bs_holland_2008,
-    _v_max_s_holland_2008, _x_holland_2010, _stat_holland_1980, _stat_holland_2010,
-    _stat_er_2011, tctrack_to_si, MBAR_TO_PA, KM_TO_M, H_TO_S,
+    _v_max_s_holland_2008, _x_holland_2010, _stat_holland_1980, _stat_holland_2010, _stat_er_2011,
+    tctrack_to_si, MBAR_TO_PA, KM_TO_M, H_TO_S,
 )
 from climada.hazard.centroids.centr import Centroids
 import climada.hazard.test as hazard_test
@@ -150,32 +150,51 @@ class TestReader(unittest.TestCase):
     def test_windfield_models(self):
         """Test _tc_from_track function with different wind field models."""
         intensity_idx = [0, 1, 2,  3,  80, 100, 120, 200, 220, 250, 260, 295]
-        intensity_values = {
-            "H08": [
-                22.74903,  23.784691, 24.82255,  22.67403,  27.218706, 30.593959,
-                18.980878, 24.540069, 27.826407, 26.846293,  0.,       34.568898,
-            ],
-            "H10": [
-                24.745521, 25.596484, 26.475329, 24.690914, 28.650107, 31.584395,
-                21.723546, 26.140293, 28.94964,  28.051915, 18.49378, 35.312152,
-            ],
-            # Holland 1980 and Emanuel & Rotunno 2011 use recorded wind speeds, while the above use
-            # pressure values only. That's why the results are so different:
-            "H1980": [21.376807, 21.957217, 22.569568, 21.284351, 24.254226, 26.971303,
-                      19.220149, 21.984516, 24.196388, 23.449116,  0, 31.550207],
-            "ER11": [23.565332, 24.931413, 26.360758, 23.490333, 29.601171, 34.522795,
-                     18.996389, 26.102109, 30.780737, 29.498453,  0, 38.368805],
-        }
+        intensity_values = [
+             ("H08", None, [
+                 22.74903, 23.784691, 24.82255, 22.67403, 27.218706, 30.593959,
+                 18.980878, 24.540069, 27.826407, 26.846293, 0., 34.568898,
+             ]),
+             ("H10", None, [
+                 24.745521, 25.596484, 26.475329, 24.690914, 28.650107, 31.584395,
+                 21.723546, 26.140293, 28.94964, 28.051915, 18.49378, 35.312152,
+             ]),
+             # The following model configurations use recorded wind speeds, while the above use
+             # pressure values only. That's why some of the values are so different.
+             ("H10", dict(vmax_from_cen=False, rho_air_const=1.2), [
+                 23.702232, 24.327615, 24.947161, 23.589233, 26.616085, 29.389295,
+                 21.338178, 24.257067, 26.472543, 25.662313, 18.535842, 31.886041,
+             ]),
+             ("H10", dict(vmax_from_cen=False, rho_air_const=None), [
+                 24.244162, 24.835561, 25.432454, 24.139294, 27.127457, 29.719196,
+                 21.910658, 24.692637, 26.783575, 25.971516, 19.005555, 31.904048,
+             ]),
+             ("H10", dict(vmax_from_cen=False, rho_air_const=None, vmax_in_brackets=True), [
[…]
