### 2024-06-21 12b2e2d
Update documentation to present pre-commit hooks
