### 2024-06-21 fa0821e
Update climada/util/test/test_checker.py
Co-authored-by: Chahan M. Kropf <chahan.kropf@usys.ethz.ch>
