### 2024-06-21 f7eaa01
Attempts at implementing all risk measures
   - Methods that should be functions are now functions
   - Adds `calc_freq_curve` from notebook
   - Adds `calc_yearly_<metric>s` for eai, aai, rp
   - Changes `bayesian_mixer` to now return a DataFrame with all risk metrics
   per year

Code was not yet tested, expect bugs
