+++ climada/engine/impact_trajectories.py
@@ -21,6 +21,7 @@ with CLIMADA. If not, see <https://www.gnu.org/licenses/>.
 import copy
 from dataclasses import dataclass
 
+import pandas as pd
 import numpy as np
 import itertools
 from datetime import datetime
@@ -112,7 +113,127 @@ def snapshot_combinaisons(snapshot0, snapshot1):
     return imp_E0H0, imp_E1H0, imp_E0H1, imp_E1H1
 
 
-# Would it be better to have a Snapshot Class that hold a singular snapshot?
+def interpolate_imp_mat(imp0, imp1, start_year, end_year):
+        return [
+            interpolate_sm(imp0.imp_mat, imp1.imp_mat, year, start_year, end_year)
+            for year in range(start_year, end_year + 1)
+        ]
+
+def calc_freq_curve(imp_mat_intrpl, frequency, return_per=None):
+    '''
+    Calculate the frequency curve
+
+    Parameters:
+    imp_mat_intrpl (np.array): The interpolated impact matrix
+    frequency (np.array): The frequency of the hazard
+    return_per (np.array): The return period
+
+    Returns:
+    ifc_return_per (np.array): The impact exceeding frequency
+    ifc_impact (np.array): The impact exceeding the return period
+    '''
+
+    # Calculate the at_event make the np.array
+    at_event = np.sum(imp_mat_intrpl, axis=1).A1
+
+    # Sort descendingly the impacts per events
+    sort_idxs = np.argsort(at_event)[::-1]
+    # Calculate exceedence frequency
+    exceed_freq = np.cumsum(frequency[sort_idxs])
+    # Set return period and impact exceeding frequency
+    ifc_return_per = 1 / exceed_freq[::-1]
+    ifc_impact = at_event[sort_idxs][::-1]
+
+    if return_per is not None:
+        interp_imp = np.interp(return_per, ifc_return_per, ifc_impact)
+        ifc_return_per = return_per
+        ifc_impact = interp_imp
+
+    return ifc_impact
+
+
+def calc_yearly_eais(imp_mats_0, imp_mats_1, frequency_0, frequency_1):
+    yearly_eai_exp_0 = [
+        ImpactCalc.eai_exp_from_mat(imp_mat, frequency_0) for imp_mat in imp_mats_0
+    ]
+    yearly_eai_exp_1 = [
+        ImpactCalc.eai_exp_from_mat(imp_mat, frequency_1) for imp_mat in imp_mats_1
+    ]
+    return yearly_eai_exp_0, yearly_eai_exp_1
+
+def calc_yearly_rps(imp_mats_0, imp_mats_1, frequency_0, frequency_1, return_periods):
+    rp_0 = [
+        calc_freq_curve(imp_mat, frequency_0, return_periods) for imp_mat in imp_mats_0
+    ]
+    rp_1 = [
+        calc_freq_curve(imp_mat, frequency_1, return_periods) for imp_mat in imp_mats_1
+    ]
+    return rp_0, rp_1
+
+def calc_yearly_aais(yearly_eai_exp_0, yearly_eai_exp_1):
+    yearly_aai_0 = [
+            ImpactCalc.aai_agg_from_eai_exp(eai_exp) for eai_exp in yearly_eai_exp_0
+    ]
+    yearly_aai_1 = [
+            ImpactCalc.aai_agg_from_eai_exp(eai_exp) for eai_exp in yearly_eai_exp_1
+    ]
+    return yearly_aai_0, yearly_aai_1
+
+
+def bayesian_mixer(start_snapshot, end_snapshot, metrics=["eai", "aai", "rp"], return_periods=[100, 500, 1000]):
+        # 1. Interpolate in between years
+        prop_H0, prop_H1 = bayesian_viktypliers(start_snapshot.year, end_snapshot.year)
[…]
