### 2024-06-21 e4c4dc9
Fixes bugs and miscomprehensions for rps
   eai by regions will come!
