### 2024-06-23 CLIMADA-project/climada_python#899 @sunt05
Many thanks for the guidance above.

However, I now encounter a new error when running verification after installing `climada`, which should be addressed as numpy 2.0 has been formally released:

```
python -m unittest climada.engine.test.test_impact
```

<summary>Error message below (Click to expand)</summary>
```
Traceback (most recent call last):
  File "/Users/tingsun/micromamba/envs/climada_env/lib/python3.9/runpy.py", line 197, in _run_module_as_main
    return _run_code(code, main_globals, None,
  File "/Users/tingsun/micromamba/envs/climada_env/lib/python3.9/runpy.py", line 87, in _run_code
    exec(code, run_globals)
  File "/Users/tingsun/micromamba/envs/climada_env/lib/python3.9/unittest/__main__.py", line 18, in <module>
    main(module=None)
  File "/Users/tingsun/micromamba/envs/climada_env/lib/python3.9/unittest/main.py", line 100, in __init__
    self.parseArgs(argv)
  File "/Users/tingsun/micromamba/envs/climada_env/lib/python3.9/unittest/main.py", line 147, in parseArgs
    self.createTests()
  File "/Users/tingsun/micromamba/envs/climada_env/lib/python3.9/unittest/main.py", line 158, in createTests
    self.test = self.testLoader.loadTestsFromNames(self.testNames,
  File "/Users/tingsun/micromamba/envs/climada_env/lib/python3.9/unittest/loader.py", line 220, in loadTestsFromNames
    suites = [self.loadTestsFromName(name, module) for name in names]
  File "/Users/tingsun/micromamba/envs/climada_env/lib/python3.9/unittest/loader.py", line 220, in <listcomp>
    suites = [self.loadTestsFromName(name, module) for name in names]
  File "/Users/tingsun/micromamba/envs/climada_env/lib/python3.9/unittest/loader.py", line 154, in loadTestsFromName
    module = __import__(module_name)
  File "/Users/tingsun/Dropbox (Personal)/Mac/Downloads/20240617-climada/climada_python/climada/__init__.py", line 24, in <module>
    from .util.config import CONFIG
  File "/Users/tingsun/Dropbox (Personal)/Mac/Downloads/20240617-climada/climada_python/climada/util/__init__.py", line 22, in <module>
    from pint import UnitRegistry
  File "/Users/tingsun/.local/lib/python3.9/site-packages/pint/__init__.py", line 28, in <module>
    from .formatting import formatter, register_unit_format
  File "/Users/tingsun/.local/lib/python3.9/site-packages/pint/formatting.py", line 17, in <module>
    from .babel_names import _babel_lengths, _babel_units
  File "/Users/tingsun/.local/lib/python3.9/site-packages/pint/babel_names.py", line 11, in <module>
    from .compat import HAS_BABEL
  File "/Users/tingsun/.local/lib/python3.9/site-packages/pint/compat.py", line 166, in <module>
    from xarray import DataArray, Dataset, Variable
  File "/Users/tingsun/.local/lib/python3.9/site-packages/xarray/__init__.py", line 1, in <module>
    from . import testing, tutorial
  File "/Users/tingsun/.local/lib/python3.9/site-packages/xarray/testing.py", line 9, in <module>
    from xarray.core import duck_array_ops, formatting, utils
  File "/Users/tingsun/.local/lib/python3.9/site-packages/xarray/core/duck_array_ops.py", line 35, in <module>
    from . import dask_array_ops, dtypes, nputils
  File "/Users/tingsun/.local/lib/python3.9/site-packages/xarray/core/dask_array_ops.py", line 3, in <module>
    from . import dtypes, nputils
  File "/Users/tingsun/.local/lib/python3.9/site-packages/xarray/core/dtypes.py", line 43, in <module>
    {np.bytes_, np.unicode_},  # numpy promotes to unicode
  File "/Users/tingsun/micromamba/envs/climada_env/lib/python3.9/site-packages/numpy/__init__.py", line 397, in __getattr__
    raise AttributeError(
AttributeError: `np.unicode_` was removed in the NumPy 2.0 release. Use `np.str_` instead.
```
