### 2024-06-24 CLIMADA-project/climada_python#899 @emanuel-schmid
Many thanks, @sunt05! This is unexpected, normally numpy 1.xx is bein installed. I suppose you can workaround by running
```bash
mamba install numpy=1.26
```
Meanwhile we have to investigate this and likely update the environment file. To speed this up, could you perhaps post the output of
```bash
mamba repoquery whoneeds numpy
```
here?
