### 2024-06-24 CLIMADA-project/climada_python#899 @sunt05
Please see below for the output of `mamba repoquery whoneeds numpy`:

```

Using local repodata...

Loaded current active prefix: "/Users/tingsun/micromamba/envs/climada_env"

No entries matching "numpy" found

numpy may not be installed. Try giving a channel with '-c,--channel' option for remote repoquery

```

From the above error message, it looks like the issue was caused by `xarray`.
