### 2024-06-24 CLIMADA-project/climada_python#899 @emanuel-schmid
Er - sorry - but that doesn't look good. In particular the output of `mamba repoquery` is inexplicable. I'm afraid something went wrong with the installation. At this point I'd suggest to restart from scratch. Do `cd <path/to/workspace>` and run

```bash
cd climada_python
git checkout calibrate-impact-functions
git pull origin calibrate-impact-functions
mamba env remove -n climada_env
mamba create -n climada_env python=3.9
mamba env update -n climada_env -f requirements/env_climada.yml
mamba activate climada_env
python -m pip install -e "./[dev]"
```

If you get an error for any of the commands, please post the command and error output here.
If all goes smoothly, run the `mamba repoquery whoneeds numpy` command again and post the output.
