### 2024-06-24 CLIMADA-project/climada_python#899 @sunt05
Thanks for updated instructions.

Please see below:

```
❯ mamba repoquery whoneeds numpy
Using local repodata...

Loaded current active prefix: "/Users/tingsun/micromamba/envs/climada_env"

 Name            Version  Build                    Depends     Channel   Subdir
────────────────────────────────────────────────────────────────────────────────
 bokeh           3.4.1    pyhd8ed1ab_0             conda-forge noarch
 bottleneck      1.4.0    py39h37867e2_0           conda-forge osx-arm64
 cartopy         0.23.0   py39h998126f_1           conda-forge osx-arm64
 cfgrib          0.9.9.1  pyhd8ed1ab_2             conda-forge noarch
 cftime          1.6.4    py39h161d348_0           conda-forge osx-arm64
 contourpy       1.2.1    py39h48c5dd5_0           conda-forge osx-arm64
 dask            2024.2.1 pyhd8ed1ab_0             conda-forge noarch
 fiona           1.9.1    py39h9e5269e_0           conda-forge osx-arm64
 folium          0.17.0   pyhd8ed1ab_0             conda-forge noarch
 gdal            3.6.2    py39h766d3fc_6           conda-forge osx-arm64
 h5py            3.8.0    nompi_py39hc9149d8_100   conda-forge osx-arm64
 mapclassify     2.6.1    pyhd8ed1ab_0             conda-forge noarch
 matplotlib-base 3.8.4    py39h15359f4_2           conda-forge osx-arm64
 netcdf4         1.6.2    nompi_py39h8ded8ba_100   conda-forge osx-arm64
 numba           0.60.0   py39h2d4ef1e_0           conda-forge osx-arm64
 numexpr         2.10.0   py39h998126f_0           conda-forge osx-arm64
 pandas          2.1.4    py39hf8cecc8_0           conda-forge osx-arm64
 patsy           0.5.6    pyhd8ed1ab_0             conda-forge noarch
 pyarrow         12.0.1   py39hf40061a_7_cpu       conda-forge osx-arm64
 pytables        3.7.0    py39h8abd629_3           conda-forge osx-arm64
 python-eccodes  1.5.1    py39h4d8bf0d_0           conda-forge osx-arm64
 rasterio        1.3.6    py39h157378c_0           conda-forge osx-arm64
 salib           1.5.0    pyhd8ed1ab_0             conda-forge noarch
 scikit-learn    1.5.0    py39h3c33c8b_1           conda-forge osx-arm64
 scipy           1.13.1   py39h3d5391c_0           conda-forge osx-arm64
 seaborn-base    0.13.2   pyhd8ed1ab_2             conda-forge noarch
 shapely         2.0.1    py39h472ea82_0           conda-forge osx-arm64
 snuggs          1.4.7    py_0                     conda-forge noarch
 sparse          0.15.4   pyhd8ed1ab_0             conda-forge noarch
 statsmodels     0.14.2   py39h161d348_0           conda-forge osx-arm64
 xarray          2024.6.0 pyhd8ed1ab_1             conda-forge noarch
(climada_env)
```
