### 2024-06-24 CLIMADA-project/climada_python#899 @peanutfun
@sunt05 This still does not look right. Note that the traceback shows that the `dask` module loaded by your interpreter is located in `/Users/tingsun/.local/lib/python3.9/site-packages/dask/` (your default interpeter location, I presume), and not in the Mamba environment `/Users/tingsun/micromamba/envs/climada_env`.

It seems you are using Micromamba, and we have no experience with its particular behavior. Can you switch to Conda or Mamba?
