### 2024-06-24 CLIMADA-project/climada_python#899 @sunt05
I see - thanks for pointing that out - will switch now and update you shortly.
