### 2024-06-24 af8bad4
add axis argument, custom error message
