### 2024-06-24 CLIMADA-project/climada_python#899 @sunt05
Thanks for the above instructions - I just managed to install the package and run the entire impact notebook.

One suggestion for the installation instruction is to change the recommended Python version from 3.9 to 3.11, as the latter has better compatibility for ARM64 Mac. Python 3.9 has various issues due to several packages lacking ARM libraries for legacy reasons.

I'll now close this issue and move forward with the [review task](https://github.com/openjournals/joss-reviews/issues/6755).
