### 2024-06-25 CLIMADA-project/climada_python#899 @emanuel-schmid
👍 Glad it worked out. Thanks for the suggestion. We wanted to do that anyway, eventually. Considering your input this should happen rather sooner than later.
