### 2024-06-25 22dddaa
Update climada/hazard/centroids/centr.py
Co-authored-by: Chahan M. Kropf <chahan.kropf@usys.ethz.ch>
