+++ climada/engine/impact_trajectories.py
@@ -304,6 +304,7 @@ class CalcImpactsSnapshots:
         for start_snapshot, end_snapshot in pairwise(self.snapshots.data):
             results_df.append(bayesian_mixer(start_snapshot, end_snapshot, metrics, return_periods))
         results_df = pd.concat(results_df,axis=0)
+        results_df.drop_duplicates()
         return results_df[["group","year","metric","result"]]
 
 
