### 2024-06-26 32166b2
Adds eai_group risk metric
