### 2024-06-28 926d114
typo and docstring cosmetics
