### 2024-06-28 24ba039
output and plot gdf
