+++ climada/hazard/test/test_io.py
@@ -686,75 +686,75 @@ class TestHDF5(unittest.TestCase):
         self.assertTrue(np.array_equal(hazard.date, hazard_read.date))
         self.assertTrue(np.array_equal(hazard_read.event_id, np.array([])))  # Empty array
 
-class TestWriteExceedAndRP(unittest.TestCase):
-    """Test writing raster and netCDF files from exceedance intensitiy and return period maps""" 
+# class TestWriteExceedAndRP(unittest.TestCase):
+#     """Test writing raster and netCDF files from exceedance intensitiy and return period maps""" 
 
-    def test_write_raster_exceed_inten(self):
-        """Test write TIFF file from local exceedance intensity"""
-        self.temp_dir = TemporaryDirectory()
-        self.test_file_path = os.path.join(self.temp_dir.name, 'test_file.tif')
+#     def test_write_raster_exceed_inten(self):
+#         """Test write TIFF file from local exceedance intensity"""
+#         self.temp_dir = TemporaryDirectory()
+#         self.test_file_path = os.path.join(self.temp_dir.name, 'test_file.tif')
 
-        haz = Hazard.from_hdf5(HAZ_TEST_TC)
-        haz.write_raster_local_exceedance_inten([10, 20, 50], filename = self.test_file_path)
-        raster = rasterio.open(self.test_file_path)
-        dataarray = np.array([ raster.read(i + 1) for i in range(raster.count)])
+#         haz = Hazard.from_hdf5(HAZ_TEST_TC)
+#         haz.write_raster_local_exceedance_inten([10, 20, 50], filename = self.test_file_path)
+#         raster = rasterio.open(self.test_file_path)
+#         dataarray = np.array([ raster.read(i + 1) for i in range(raster.count)])
 
-        np.testing.assert_array_almost_equal(
-            np.transpose(np.flip(dataarray, axis = 1), axes= [0, 2, 1]),
-            haz.local_exceedance_inten([10, 20, 50]).reshape((3, 10, 10)),
-            decimal=4
-        )
-        self.temp_dir.cleanup()
+#         np.testing.assert_array_almost_equal(
+#             np.transpose(np.flip(dataarray, axis = 1), axes= [0, 2, 1]),
+#             haz.local_exceedance_inten([10, 20, 50]).reshape((3, 10, 10)),
+#             decimal=4
+#         )
+#         self.temp_dir.cleanup()
     
-    def test_write_raster_local_return_periods(self):
-        """Test write TIFF file from local return periods intensity"""
-        self.temp_dir = TemporaryDirectory()
-        self.test_file_path = os.path.join(self.temp_dir.name, 'test_file.tif')
+#     def test_write_raster_local_return_periods(self):
+#         """Test write TIFF file from local return periods intensity"""
+#         self.temp_dir = TemporaryDirectory()
+#         self.test_file_path = os.path.join(self.temp_dir.name, 'test_file.tif')
 
-        haz = Hazard.from_hdf5(HAZ_TEST_TC)
-        haz.write_raster_local_return_periods([10., 20., 30.], filename = self.test_file_path)
-        raster = rasterio.open(self.test_file_path)
-        dataarray = np.array([ raster.read(i + 1) for i in range(raster.count)])
+#         haz = Hazard.from_hdf5(HAZ_TEST_TC)
+#         haz.write_raster_local_return_periods([10., 20., 30.], filename = self.test_file_path)
+#         raster = rasterio.open(self.test_file_path)
+#         dataarray = np.array([ raster.read(i + 1) for i in range(raster.count)])
 
-        np.testing.assert_array_almost_equal(
-            dataarray,
-            haz.local_return_period([10., 20., 30.]).reshape((3, 10, 10)),
-            decimal=4
-        )
[…]
