+++ climada/engine/impact_metrics.py
@@ -1,0 +1,31 @@
+"""
+This file is part of CLIMADA.
+
+Copyright (C) 2017 ETH Zurich, CLIMADA contributors listed in AUTHORS.
+
+CLIMADA is free software: you can redistribute it and/or modify it under the
+terms of the GNU General Public License as published by the Free
+Software Foundation, version 3.
+
+CLIMADA is distributed in the hope that it will be useful, but WITHOUT ANY
+WARRANTY; without even the implied warranty of MERCHANTABILITY or FITNESS FOR A
+PARTICULAR PURPOSE.  See the GNU General Public License for more details.
+
+You should have received a copy of the GNU General Public License along
+with CLIMADA. If not, see <https://www.gnu.org/licenses/>.
+
+---
+
+"""
+
+import pandas as pd
+
+
+class ImpactMetrics:
+
+    def __init__(self, 
+                 arm_df: pd.DataFrame,
+                 all_arms_df: pd.DataFrame,
+                 ):
+        self.arm_df = arm_df
+        self.all_arms_df = all_arms_df
