### 2024-07-01 #6755 @sunt05
Congratulations to the CLIMADA team on delivering this excellent work, which I believe will be very useful to the climate risk community.

After trying out the new development, I have only two minor points for the climada team to consider:

- The main demo notebook `climate_util_calibrate.ipynb` is quite informative but reads a bit lengthy. I'd suggest providing a `quickstart` section to quickly cover the essential steps. Readers can then revisit the whole for a more thorough understanding.

- The notebook has some typos and unused imports - I've left comments in the PR thread for the authors to address. But please get some fresh eyes to proofread it.

Slightly off-topic, while I understand CLIMADA is a comprehensive codebase with various dependencies, I believe the dependency structure could be simplified for easier installation. Even as an experienced Python user and developer, it took me a while to get it working. New users interested in climate risk analysis rather than Python development might struggle even more. Also, since the complexity largely comes from geospatial analysis packages, I'd suggest focusing on climate risk analysis and leaving the geospatial part to more experienced users.
