+++ climada/engine/impact_metrics.py
@@ -17,8 +17,969 @@ with CLIMADA. If not, see <https://www.gnu.org/licenses/>.
 ---
 
 """
+from climada.entity.measures import MeasureSet
 
 import pandas as pd
+from typing import Optional
+import copy
+
+import numpy as np
+from tabulate import tabulate
+
+import matplotlib.pyplot as plt
+import seaborn as sns
+
+
+#%% Utility functions
+
+
+## Generate active measures for a specific year
+def get_meas_times_df(measure_set, incl_combo=False):
+    """
+    Get a DataFrame with the start and end years of the measures in a MeasureSet
+
+    Parameters
+    ----------
+    MeasureSet : MeasureSet
+        The measure set
+
+    Returns
+    -------
+    pd.DataFrame
+        DataFrame with the start and end years of the measures
+    """
+
+    cols = ['measure', 'start_year', 'end_year']
+    meas_times_df = pd.DataFrame(columns=cols)
+
+    # Populate the df
+    for _,meas in measure_set.measures().items():
+
+        # Skip combo measures if not included
+        if not incl_combo and meas.combo:
+            continue
+        
+        # Create a new DataFrame for the current measure
+        temp_df = pd.DataFrame({
+            'measure': [meas.name],
+            'start_year': [meas.start_year],
+            'end_year': [meas.end_year]
+        })
+        
+        # Append to the df
+        if meas_times_df.empty:
+            meas_times_df = temp_df
+        else:
+            meas_times_df = pd.concat([meas_times_df, temp_df])
+
+    return meas_times_df
+
+
+def get_active_measure_combinations(meas_times_df):
+
+    # Get the range of years
+    min_year = meas_times_df['start_year'].min()
+    max_year = meas_times_df['end_year'].max()
+    years = range(min_year, max_year + 1)
+    
+    # Initialize a list to store the active measures for each year
+    active_measures_by_year = []
+
+    for year in years:
+        active_measures = meas_times_df[(meas_times_df['start_year'] <= year) & (meas_times_df['end_year'] >= year)]['measure'].tolist()
+        if not active_measures:
+            active_measures = ['no_measure']
+        active_measures_by_year.append(frozenset(active_measures))  # Use frozenset for unique combinations
+
+    # Get the unique combinations of active measures
+    unique_combinations = set(active_measures_by_year)
+    
+    # Convert the frozensets to sorted lists for readability and filter out lists with size equal to one
+    unique_combinations = [sorted(list(comb)) for comb in unique_combinations if len(comb) > 1]
+
+    return unique_combinations
+
+def include_combos_in_measure_set(measure_set, *other_combos, all_measures=True ):
+    """
+    Generate possible combinations of measures in a measure set
+
+    Parameters
+    ----------
+    measure_set : MeasureSet
+        The measure set to be combined
+    all_measures : bool
+        If True, all measures are combined
+    other_combos : list
+        List of lists of measures to be combined
+
+    Returns
+    -------
+    dict
+        Dictionary of combined measures
+    """
+
+    # Make a copy of the measure set
+    new_measure_set = copy.deepcopy(measure_set)
+
+    # Combine all measures
+    if all_measures:
+        meas_all = new_measure_set.combine(combo_name='all')
+        #new_measure_set.append(meas_combo)
+
+    # Combine other measures
+    for combo in other_combos:
+        meas_combo = new_measure_set.combine(names=combo)
+        new_measure_set.append(meas_combo)
+
+    # Add the 'all' measure
+    if all_measures:
+        new_measure_set.append(meas_all)
+    
+    return new_measure_set
+
+
+
[…]
