### 2024-07-01 d1dc9a5
Initiate works on CalcImpactMetrics class
