+++ climada/engine/calc_impact_metrics.py
@@ -1,0 +1,358 @@
+"""
+This file is part of CLIMADA.
+
+Copyright (C) 2017 ETH Zurich, CLIMADA contributors listed in AUTHORS.
+
+CLIMADA is free software: you can redistribute it and/or modify it under the
+terms of the GNU General Public License as published by the Free
+Software Foundation, version 3.
+
+CLIMADA is distributed in the hope that it will be useful, but WITHOUT ANY
+WARRANTY; without even the implied warranty of MERCHANTABILITY or FITNESS FOR A
+PARTICULAR PURPOSE.  See the GNU General Public License for more details.
+
+You should have received a copy of the GNU General Public License along
+with CLIMADA. If not, see <https://www.gnu.org/licenses/>.
+
+---
+
+"""
+import copy
+
+from climada.engine.impact_trajectories import CalcImpactsSnapshots, SnapshotsCollection
+from climada.engine.impact_metrics import ImpactMetrics
+import pandas as pd
+
+# BASE_arm_df_COLUMNS = ['measure', 'group', 'year', 'metric', 'result']
+
+def get_meas_times_df(measure_set, incl_combo=False):
+    """
+    Get a DataFrame with the start and end years of the measures in a MeasureSet
+
+    Parameters
+    ----------
+    MeasureSet : MeasureSet
+        The measure set
+
+    Returns
+    -------
+    pd.DataFrame
+        DataFrame with the start and end years of the measures
+    """
+
+    cols = ['measure', 'start_year', 'end_year']
+    meas_times_df = pd.DataFrame(columns=cols)
+
+    # Populate the df
+    for _,meas in measure_set.measures().items():
+
+        # Skip combo measures if not included
+        if not incl_combo and meas.combo:
+            continue
+
+        # Create a new DataFrame for the current measure
+        temp_df = pd.DataFrame({
+            'measure': [meas.name],
+            'start_year': [meas.start_year],
+            'end_year': [meas.end_year]
+        })
+
+        # Append to the df
+        if meas_times_df.empty:
+            meas_times_df = temp_df
+        else:
+            meas_times_df = pd.concat([meas_times_df, temp_df])
+
+    return meas_times_df
+
+
+def get_active_measure_combinations(meas_times_df):
+
+    # Get the range of years
+    min_year = meas_times_df['start_year'].min()
+    max_year = meas_times_df['end_year'].max()
+    years = range(min_year, max_year + 1)
+
+    # Initialize a list to store the active measures for each year
+    active_measures_by_year = []
+
+    for year in years:
+        active_measures = meas_times_df[(meas_times_df['start_year'] <= year) & (meas_times_df['end_year'] >= year)]['measure'].tolist()
+        if not active_measures:
+            active_measures = ['no_measure']
+        active_measures_by_year.append(frozenset(active_measures))  # Use frozenset for unique combinations
+
+    # Get the unique combinations of active measures
+    unique_combinations = set(active_measures_by_year)
+
+    # Convert the frozensets to sorted lists for readability and filter out lists with size equal to one
+    unique_combinations = [sorted(list(comb)) for comb in unique_combinations if len(comb) > 1]
+
+    return unique_combinations
+
+def include_combos_in_measure_set(measure_set, *other_combos, all_measures=True ):
+    """
+    Generate possible combinations of measures in a measure set
+
+    Parameters
+    ----------
+    measure_set : MeasureSet
+        The measure set to be combined
+    all_measures : bool
+        If True, all measures are combined
+    other_combos : list
+        List of lists of measures to be combined
+
+    Returns
+    -------
+    dict
+        Dictionary of combined measures
+    """
+
+    # Make a copy of the measure set
[…]
