### 2024-07-01 #6755 @observingClouds
@peanutfun please have a look at the last comments of sunt05. Please improve the demo notebook accordingly and improve the user experience when installing CLIMADA.

I expect the reviews from @naingeet also coming in this week, but please work in the meantime on above's issues to help the interactivity of the review process.

Cheers!
