### 2024-07-02 #6755 @peanutfun
@sunt05 Thank you for your review. I am very glad about your positive feedback. We will add a quickstart section to the tutorial, as you suggested, in the next few days.

I understand that the setup for the development version of Climada is quite involved, especially if you have a working setup that does not rely on Conda. Unfortunately, this is a somewhat historical problem we cannot resolve easily. First, we think that geospatial analysis combined with the comparably simple task of impact calculation is the main advantage of Climada, and we would not want to make it optional. Second, the trouble comes from a few C(++)-library dependencies that are used for some submodules of Climada. We plan to move these to the adjacent repository https://github.com/CLIMADA-project/climada_petals, see https://github.com/CLIMADA-project/climada_python/discussions/729. Our goal is to make the "Core" https://github.com/CLIMADA-project/climada_python repository and all its dependencies installable only via Pip, which is currently not the case. However, we decided to update these submodules before moving, as they have grown outdated, and are waiting for feedback from collaborators, see https://github.com/CLIMADA-project/climada_python/discussions/811. Third, our current module import patterns do not allow well for optional dependencies. Also, testing (combinations of) optional dependencies is a larger task that would overburden our current resources.

Finally, we provide a [conda-forge package](https://anaconda.org/conda-forge/climada) for all Climada releases. We hope this simplifies installation for inexperienced users enough to get started with Climada easily:
```
conda create -n climada_env climada
```
Of course, they then cannot switch to a development or feature branch, as you were required to.

I hope this clarifies why we currently do not see an easy, quick way to improve the user experience regarding development version installation. We are well aware of the issue, and resolving it is a somewhat long-term goal for us. Still, if you encountered some pitfalls our installation instructions do not cover, we would gladly incorporate any suggestions for improvement.
