### 2024-07-02 e56af25
Only write H5 file if Impact.event_name is strings
Try to convert to strings when reading.
