+++ climada/engine/test/test_impact.py
@@ -1020,9 +1020,10 @@ class TestImpactH5IO(unittest.TestCase):
     def test_write_hdf5_type_fail(self):
         """Test that writing attributes with varying types results in an error"""
         self.impact.event_name = [1, "a", 1.0, "b", "c", "d"]
-        with self.assertRaises(TypeError) as cm:
+        with self.assertRaisesRegex(
+            TypeError, "'event_name' must be a list of strings"
+        ):
             self.impact.write_hdf5(self.filepath)
-        self.assertIn("No conversion path for dtype", str(cm.exception))
 
     def test_cycle_hdf5(self):
         """Test writing and reading the same object"""
@@ -1127,8 +1128,8 @@ class TestImpactH5IO(unittest.TestCase):
             file.create_dataset("event_name", data=event_name)
         with self.assertLogs("climada.engine.impact", "WARNING") as cm:
             impact = Impact.from_hdf5(self.filepath)
-        self.assertIn("'event_name' could not be decoded to a string", cm.output[0])
-        self.assertListEqual(impact.event_name, event_name)
+        self.assertIn("'event_name' is not stored as strings", cm.output[0])
+        self.assertListEqual(impact.event_name, ["1.2", "2.0"])
 
 # Execute Tests
 if __name__ == "__main__":
