### 2024-07-02 067c9be
Fix bug where only the first event name was checked
Update docs
