### 2024-07-04 6ad7ab4
Apply suggestions from code review
