### 2024-07-05 fc19727
remove *args from plotting method
