### 2024-07-05 fe3becc
first draft general fit method
