+++ climada/engine/option_appraisal/MCDM/utils.py
@@ -1,0 +1,503 @@
+#!/usr/bin/env python3
+# -*- coding: utf-8 -*-
+"""
+Created on Fri Feb  9 14:58:15 2024
+
+@author: vwattin
+"""
+
+from climada.entity import Entity
+from climada.util.api_client import Client
+from climada.engine import CostBenefit
+from climada.engine.cost_benefit import risk_aai_agg, risk_rp_100, risk_rp_250
+from climada.engine.unsequa import InputVar, UncOutput
+from IPython.display import clear_output
+
+from climada.engine.unsequa import CalcCostBenefit
+
+
+import pandas as pd
+import copy
+import os
+import itertools
+import scipy as sp
+from functools import partial
+import numpy as np
+
+#import functions as fcn
+
+# Constants
+CURRENT_YEAR = 2018
+FUTURE_YEAR = 2040
+# Define the risk functions dictionary
+RISK_FNCS_DICT = { 'aai': risk_aai_agg,
+                  'rp250': risk_rp_250}
+
+
+import inspect
+
+def filter_dataframe(df, filter_conditions=None, derived_columns=None, base_cols=None):
+    """
+    This function filters a DataFrame based on provided conditions and calculates derived columns.
+
+    Parameters:
+    - df (pandas.DataFrame): The input DataFrame.
+    - filter_conditions (dict): A dictionary specifying filtering conditions for columns.
+    - derived_columns (dict): A dictionary specifying derived columns and their functions.
+
+    Returns:
+    - filtered_df (pandas.DataFrame): The filtered DataFrame based on conditions and derived columns.
+    - boolean_df (pandas.DataFrame): A boolean DataFrame indicating whether values satisfy conditions.
+    """
+
+    # Create a copy of the input DataFrame
+    filtered_df = df.copy()
+    unfiltered_df = df.copy()
+    
+    # If conditions or derived columns are not provided, initialize them as empty dictionaries
+    if filter_conditions is None:
+        filter_conditions = {}
+    
+    if derived_columns is None:
+        derived_columns = {}
+
+    # Calculate and add derived columns to the filtered DataFrame
+    if derived_columns:
+        for new_col, function in derived_columns.items():
+            filtered_df[new_col] = function(df)
+            unfiltered_df[new_col] = function(df)
+
+    # Create a boolean DataFrame to track conditions satisfaction
+    if base_cols:
+        boolean_df = df[base_cols].copy()
+    else:
+        boolean_df = df.copy()
+    
+    # Apply filtering conditions and update boolean DataFrame accordingly
+    if filter_conditions:
+        for col, cond in filter_conditions.items():
+            
+            if isinstance(cond, list):
+                # Filter data based on whether column values are equal to the provided value
+                filtered_df = filtered_df[filtered_df[col] == cond['equal']]
+                boolean_df[col] = unfiltered_df[col] == cond['equal']
+            elif 'equal' in cond:
+                # Filter data based on whether column values are equal to the provided value
+                filtered_df = filtered_df[filtered_df[col].isin(cond['equal'])]
+                boolean_df[col] = unfiltered_df[col].isin(cond['equal'])
+            elif 'in' in cond:
+                # Filter data based on whether column values are in the provided list
+                filtered_df = filtered_df[filtered_df[col].isin(cond['in'])]
+                boolean_df[col] = unfiltered_df[col].isin(cond['in'])
+            elif 'greater' in cond:
+                # Filter data based on whether column values are greater than the provided value
+                filtered_df = filtered_df[filtered_df[col] > cond['greater']]
+                boolean_df[col] = unfiltered_df[col] > cond['greater']
+            elif 'less' in cond:
+                # Filter data based on whether column values are less than the provided value
[…]
