### 2024-07-07 #6755 @naingeet
I finished my review on the code and the paper draft in https://github.com/CLIMADA-project/climada_python/pull/692.  Installation proceeded as outlined in advanced instructions for installation and functionally were correctly implemented. The demo notebook code was quite comprehensive and easy to follow.
Great work, team CLIMADA!
