### 2024-07-08 f50e148
Explicitly convert `event_name` to strings in `Impact.from_hdf5` (#894)
* Try converting non-string impact event names from H5
* Only write H5 file if Impact.event_name is strings
* Fix bug where only the first event name was checked
* Update CHANGELOG.md

---------

Co-authored-by: emanuel-schmid <schmide@ethz.ch>
