+++ climada/engine/option_appraisal/impact_metrics.py
@@ -248,6 +248,7 @@ def _calc_waterfall_plot_df(arm_df, all_arms_df, measure=None, metric='aai', ref
         
     return waterfall_df
 
+# Plot the yearly waterfall plot
 def _plot_yearly_waterfall(waterfall_df, metric='aai', value_unit='USD'):
     fig, ax = plt.subplots(figsize=(15, 10))
 
@@ -315,7 +316,7 @@ def _plot_yearly_waterfall(waterfall_df, metric='aai', value_unit='USD'):
     plt.show()
 
 
-
+# Plot the waterfall plot for two years
 def _plot_two_years_waterfall(waterfall_df, metric='aai', value_unit='USD'):
     # Get the reference and future years
     ref_year = waterfall_df['year'].min()
@@ -389,19 +390,14 @@ def _plot_two_years_waterfall(waterfall_df, metric='aai', value_unit='USD'):
 
     plt.show()
 
-## Yearly risk and cost functions
-# Get the active measures for a specific year
-def get_active_measures_for_year(measure_set, year, incl_combo=False):
+# 
+def get_active_measures_for_year(meas_times_df, year):
     
-    # Get the DataFrame of the individul measure times
-    meas_times_df = get_meas_times_df(measure_set, incl_combo)
-
     # Filter the DataFrame for the specified year
     active_measures = meas_times_df[(meas_times_df['start_year'] <= year) & (meas_times_df['end_year'] >= year)]['measure'].tolist()
     if not active_measures:
         active_measures = ['no_measure']
     return active_measures
-
 
 # Get the name of the active combo measure
 def get_name_of_active_combo(measure_set, active_measures):
@@ -416,6 +412,7 @@ def get_name_of_active_combo(measure_set, active_measures):
         if meas.combo and set(meas.combo) == set(active_measures):
                 return meas.name
 
+# Calculate the averted risk metrics
 def calc_averted_risk_metrics(arm_df):
     # Copy the DataFrame to avoid modifying the original
     averted_arm_df = pd.DataFrame(columns=arm_df.columns)
@@ -462,6 +459,7 @@ def calc_averted_risk_metrics(arm_df):
 
     return averted_arm_df
 
+# Plot the risk metrics
 def plot_risk_metrics(arm_df, metric='aai', group=np.nan, averted=False, discounted=False, plot_type='line', measures=None, value_unit='USD'):
     # Filter the DataFrame
     if pd.isna(group):
@@ -580,7 +578,10 @@ def calc_npv_arm_df(df, disc=None):
     
     return disc_df, npv_df
 
-def create_indv_meas_mod_arm_df(arm_df, measure_set, consider_measure_times=True):
+#%% Functions for the modified arm_df
+
+# Create the modified arm_df for the individual measures
+def create_indv_meas_mod_arm_df(arm_df, measure_set, measure_times_df, consider_measure_times=True):
 
     # Use start_year and end_year if they are provided and filter the arm_df
     start_year = arm_df['year'].min()
@@ -603,9 +604,10 @@ def create_indv_meas_mod_arm_df(arm_df, measure_set, consider_measure_times=True
         if meas.combo:
             continue
 
+        # Get the measure name and the start and end years from the measure_times_df
         measure_name = meas.name
-        meas_start_year = meas.start_year
-        meas_end_year = meas.end_year
+        meas_start_year = measure_times_df[measure_times_df['measure'] == measure_name]['start_year'].values[0]
+        meas_end_year = measure_times_df[measure_times_df['measure'] == measure_name]['end_year'].values[0]
 
         # Iterate over the years
         for year in range(start_year, end_year + 1):
[…]

[…]
