+++ climada/hazard/trop_cyclone/__init__.py
@@ -1,0 +1,5 @@
+from climada.hazard.trop_cyclone.trop_cyclone import *
+from climada.hazard.trop_cyclone.trop_cyclone_windfields import _compute_windfields_sparse, compute_angular_windspeeds, tctrack_to_si, \
+    get_close_centroids, KN_TO_MS, KM_TO_M, KM_TO_M, H_TO_S, NM_TO_KM, KMH_TO_MS, MBAR_TO_PA, \
+    DEF_MAX_DIST_EYE_KM, DEF_INTENSITY_THRES, DEF_MAX_MEMORY_GB, MODEL_VANG, DEF_RHO_AIR, DEF_GRADIENT_TO_SURFACE_WINDS, \
+    T_ICE_K, V_ANG_EARTH
