+++ climada/hazard/test/test_trop_cyclone.py
@@ -26,24 +26,20 @@ import unittest
 
 import numpy as np
 from scipy import sparse
-import xarray as xr
 
+from climada.hazard import test as hazard_test
 from climada.util import ureg
 from climada.test import get_test_file
 from climada.hazard.tc_tracks import TCTracks
-from climada.hazard.trop_cyclone import (
-    TropCyclone, get_close_centroids, _vtrans, _B_holland_1980, _bs_holland_2008,
-    _v_max_s_holland_2008, _x_holland_2010, _stat_holland_1980, _stat_holland_2010, _stat_er_2011,
-    tctrack_to_si, MBAR_TO_PA, KM_TO_M, H_TO_S,
-)
+from climada.hazard.trop_cyclone.trop_cyclone import (
+    TropCyclone, )
 from climada.hazard.centroids.centr import Centroids
-import climada.hazard.test as hazard_test
-
 
 DATA_DIR = Path(hazard_test.__file__).parent.joinpath('data')
 
 TEST_TRACK = DATA_DIR.joinpath("trac_brb_test.csv")
 TEST_TRACK_SHORT = DATA_DIR.joinpath("trac_short_test.csv")
+
 
 CENTR_TEST_BRB = Centroids.from_hdf5(get_test_file('centr_test_brb', file_format='hdf5'))
 
@@ -294,236 +290,6 @@ class TestReader(unittest.TestCase):
         self.assertEqual(tc_haz.fraction.nonzero()[0].size, 0)
         self.assertEqual(tc_haz.intensity.nonzero()[0].size, 0)
 
-class TestWindfieldHelpers(unittest.TestCase):
-    """Test helper functions of TC wind field model"""
-
-    def test_get_close_centroids_pass(self):
-        """Test get_close_centroids function."""
-        si_track = xr.Dataset({
-            "lat": ("time", np.array([0, -0.5, 0])),
-            "lon": ("time", np.array([0.9, 2, 3.2])),
-        }, attrs={"mid_lon": 0.0})
-        centroids = np.array([
-            [0, -0.2], [0, 0.9], [-1.1, 1.2], [1, 2.1], [0, 4.3], [0.6, 3.8], [0.9, 4.1],
-        ])
-        centroids_close, mask_close, mask_close_alongtrack = (
-            get_close_centroids(si_track, centroids, 112.0)
-        )
-        self.assertEqual(centroids_close.shape[0], mask_close.sum())
-        self.assertEqual(mask_close_alongtrack.shape[0], si_track.sizes["time"])
-        self.assertEqual(mask_close_alongtrack.shape[1], centroids_close.shape[0])
-        np.testing.assert_equal(mask_close_alongtrack.any(axis=0), True)
-        np.testing.assert_equal(mask_close, np.array(
-            [False, True, True, False, False, True, False]
-        ))
-        np.testing.assert_equal(mask_close_alongtrack, np.array([
-            [True, False, False],
-            [False, True, False],
-            [False, False, True],
-        ]))
-        np.testing.assert_equal(centroids_close, centroids[mask_close])
-
-        # example where antimeridian is crossed
-        si_track = xr.Dataset({
-            "lat": ("time", np.linspace(-10, 10, 11)),
-            "lon": ("time", np.linspace(170, 200, 11)),
-        }, attrs={"mid_lon": 180.0})
-        centroids = np.array([[-11, 169], [-7, 176], [4, -170], [10, 170], [-10, -160]])
-        centroids_close, mask_close, mask_close_alongtrack = (
-            get_close_centroids(si_track, centroids, 600.0)
-        )
-        self.assertEqual(centroids_close.shape[0], mask_close.sum())
-        self.assertEqual(mask_close_alongtrack.shape[0], si_track.sizes["time"])
[…]
