+++ climada/hazard/trop_cyclone.py
@@ -1,2031 +1,0 @@
-"""
-This file is part of CLIMADA.
-
-Copyright (C) 2017 ETH Zurich, CLIMADA contributors listed in AUTHORS.
-
-CLIMADA is free software: you can redistribute it and/or modify it under the
-terms of the GNU General Public License as published by the Free
-Software Foundation, version 3.
-
-CLIMADA is distributed in the hope that it will be useful, but WITHOUT ANY
-WARRANTY; without even the implied warranty of MERCHANTABILITY or FITNESS FOR A
-PARTICULAR PURPOSE.  See the GNU General Public License for more details.
-
-You should have received a copy of the GNU General Public License along
-with CLIMADA. If not, see <https://www.gnu.org/licenses/>.
-
----
-
-Define TC wind hazard (TropCyclone class).
-"""
-
-__all__ = ['TropCyclone']
-
-import copy
-import datetime as dt
-import itertools
-import logging
-import time
-from typing import Optional, Tuple, List, Union
-
-import numpy as np
-from scipy import sparse
-import matplotlib.animation as animation
-from tqdm import tqdm
-import pathos.pools
-import xarray as xr
-
-from climada.hazard.base import Hazard
-from climada.hazard.tc_tracks import TCTracks, estimate_rmw
-from climada.hazard.tc_clim_change import get_knutson_criterion, calc_scale_knutson
-from climada.hazard.centroids.centr import Centroids
-from climada.util import ureg
-import climada.util.constants as u_const
-import climada.util.coordinates as u_coord
-import climada.util.plot as u_plot
-
-LOGGER = logging.getLogger(__name__)
-
-HAZ_TYPE = 'TC'
-"""Hazard type acronym for Tropical Cyclone"""
-
-DEF_MAX_DIST_EYE_KM = 300
-"""Default value for the maximum distance (in km) of a centroid to the TC center at which wind
-speed calculations are done."""
-
-DEF_INTENSITY_THRES = 17.5
-"""Default value for the threshold below which wind speeds (in m/s) are stored as 0."""
-
-DEF_MAX_MEMORY_GB = 8
-"""Default value of the memory limit (in GB) for windfield computations (in each thread)."""
-
-MODEL_VANG = {'H08': 0, 'H1980': 1, 'H10': 2, 'ER11': 3}
-"""Enumerate different symmetric wind field models."""
-
-DEF_RHO_AIR = 1.15
-"""Default value for air density (in kg/m³), following Holland 1980."""
-
-DEF_GRADIENT_TO_SURFACE_WINDS = 0.9
-"""Default gradient-to-surface wind reduction factor, following the 90%-rule mentioned in:
-
-Franklin, J.L., Black, M.L., Valde, K. (2003): GPS Dropwindsonde Wind Profiles in Hurricanes and
-Their Operational Implications. Weather and Forecasting 18(1): 32–44.
-https://doi.org/10.1175/1520-0434(2003)018<0032:GDWPIH>2.0.CO;2
-
-According to Table 2, this is a reasonable factor for the 750 hPa level in the eyewall region. For
-other regions and levels, values of 0.8 or even 0.75 might be justified.
-"""
-
-KMH_TO_MS = (1.0 * ureg.km / ureg.hour).to(ureg.meter / ureg.second).magnitude
-KN_TO_MS = (1.0 * ureg.knot).to(ureg.meter / ureg.second).magnitude
-NM_TO_KM = (1.0 * ureg.nautical_mile).to(ureg.kilometer).magnitude
-KM_TO_M = (1.0 * ureg.kilometer).to(ureg.meter).magnitude
-H_TO_S = (1.0 * ureg.hours).to(ureg.seconds).magnitude
[…]

[…]
