### 2024-07-08 #6755 @peanutfun
@naingeet Thank you for your positive feedback and review!

@sunt05 We added a quickstart section and improved the links from the tutorial to the respective Python module documentation sections. We also incorporated your suggestions for fixing typos. See https://climada-python--692.org.readthedocs.build/en/692/tutorial/climada_util_calibrate.html for the latest version of the tutorial.
