### 2024-07-08 #6755 @sunt05
Many thanks for addressing my comments @peanutfun!

The revised notebook looks brilliant. I'm very happy to suggest the acceptance of your work 🎉
