+++ climada/engine/option_appraisal/impact_metrics.py
@@ -853,7 +853,7 @@ def calc_CB_df(arm_df, measure_set, measure_times_df, start_year=None, end_year=
     disc_filt_arm_df, _ = calc_npv_arm_df(filt_arm_df, disc=risk_disc)
 
     # Get the base CB dataframe
-    ann_CB_df = filt_arm_df[['measure', 'year', 'group', 'metric', 'result']].copy()
+    ann_CB_df = disc_filt_arm_df[['measure', 'year', 'group', 'metric', 'result']].copy()
     ann_CB_df.columns = ['measure', 'year', 'group', 'metric', 'total risk']
 
     # Step 4 - Calculate the averted risk metrics
@@ -865,7 +865,7 @@ def calc_CB_df(arm_df, measure_set, measure_times_df, start_year=None, end_year=
     ann_CB_df = ann_CB_df.merge(averted_arm_df, on=['measure', 'year', 'group', 'metric'], how='left')
 
     # Calculate the residual risk
-    ann_CB_df['residual risk'] = ann_CB_df['total risk'] - ann_CB_df['averted risk']
+    #ann_CB_df['residual risk'] = ann_CB_df['total risk'] - ann_CB_df['averted risk']
 
     # Calculate the measure cash flows
     costincome_df = calc_measure_cash_flows_df(measure_set, measure_times_df, start_year, end_year, consider_measure_times, disc=cost_disc)
@@ -889,44 +889,53 @@ def calc_CB_df(arm_df, measure_set, measure_times_df, start_year=None, end_year=
     # Drop the 'year' column
     tot_CB_df = tot_CB_df.drop(columns=['year'])
 
+    # Add the average annual risk
+    tot_CB_df['average annual risk'] = tot_CB_df['total risk'] / (end_year - start_year + 1)
+
     # Step 7 - Calculate the CB ratio
     # Calculate the CB ratio
-    tot_CB_df['B/C ratio'] = tot_CB_df['averted risk'] / -tot_CB_df['cost (net)']
+    tot_CB_df['B/C ratio'] = tot_CB_df['averted risk'] / tot_CB_df['cost (net)']
 
     return ann_CB_df, tot_CB_df
 
-def print_CB_summary_table(tot_CB_df, metric='aai', value_unit='USD', group=np.nan):
+def print_CB_summary_table(tot_CB_df, measure_set, metric='aai', value_unit='USD', group=np.nan):
     df_filtered = tot_CB_df[tot_CB_df['metric'] == metric]
     group = str(group)  # Cast the 'group' to string type
     df_filtered = df_filtered[df_filtered['group'] == group]
 
-    total_climate_risk = df_filtered[df_filtered['measure'] == 'no_measure']['total risk']
-    annual_risk = total_climate_risk / (df_filtered['year'].max() - df_filtered['year'].min() + 1)
-    residual_risk = df_filtered[df_filtered['measure'] != 'no_measure']['residual risk'].min()
+    # Get the total climate risk, annual risk, and residual risk
+    total_climate_risk = df_filtered[df_filtered['measure'] == 'no_measure']['total risk'].values[0]
+    annual_risk = df_filtered[df_filtered['measure'] == 'no_measure']['average annual risk'].values[0]
+    residual_risk = df_filtered[df_filtered['measure'] != 'no_measure']['total risk'].min()
 
 
     # Set unit to billions ('Bn') regardless of the magnitude
     unit = 'Bn'
     divisor = 1e9
 
-
+    # Prepare the measure data
     measure_data = df_filtered[df_filtered['measure'] != 'no_measure']
     measure_summary = []
-    for measure in measure_data['measure'].unique():
+    for measure in measure_set.measures():
         measure_df = measure_data[measure_data['measure'] == measure]
         total_cost = measure_df['cost (net)'].sum()
[…]
