### 2024-07-09 0a6c627
try fixing petals 3.11 compatibility error (petals issue #135)
