### 2024-07-09 f359ed9
skip scheduler in set_df_geometry_points
