+++ climada/util/coordinates.py
@@ -2653,7 +2653,7 @@ def set_df_geometry_points(df_val, scheduler=None, crs=None):
     """
     LOGGER.info('Setting geometry points.')
     if scheduler is not None:
-        LOGGER.warn("The scheduler argument is obsolete, dask support has been removed from the"
+        LOGGER.warning("The scheduler argument is obsolete, dask support has been removed from the"
                     " `set_df_geometry_points` function.")
 
     # keep the original crs if any
