### 2024-07-10 5583104
Update climada/util/coordinates.py
Co-authored-by: Lukas Riedel <34276446+peanutfun@users.noreply.github.com>
