+++ climada/util/coordinates.py
@@ -2646,8 +2646,12 @@ def set_df_geometry_points(df_val, scheduler=None, crs=None):
     ----------
     df_val : GeoDataFrame
         contains latitude and longitude columns
-    scheduler : str
-        deprecated, obsolete
+    scheduler : str, optional
+        Scheduler type for dask map_partitions.
+
+        .. deprecated:: 4.1
+           This function does not use dask features anymore. The parameter has no effect
+           and will be removed in a future version.
     crs : object (anything readable by pyproj4.CRS.from_user_input), optional
         Coordinate Reference System, if omitted or None: df_val.geometry.crs
     """
