### 2024-07-10 6e73cf8
warn through warnings not through logs, fix deprecation version
