### 2024-07-10 6a15a6f
Update installation instructions and guides for pre-commit
