### 2024-07-10 7840521
remove obsolete multi_processing test, make test_points_to_raster mutltiprocessing instead
