### 2024-07-10 9e33969
fix petals 3.11 compatibility error (petals issue 135)
