+++ climada/hazard/tc_tracks.py
@@ -1383,7 +1383,7 @@ class TCTracks():
         df_attrs = pd.DataFrame([t.attrs for t in data], index=ds_combined["storm"].to_series())
         ds_combined = xr.merge([ds_combined, df_attrs.to_xarray()])
 
-        encoding = {v: dict(zlib=False, complevel=complevel) for v in ds_combined.data_vars}
+        encoding = {v: dict(zlib=_zlib_from_DataArray(ds_combined[v]), complevel=complevel) for v in ds_combined.data_vars}
         LOGGER.info('Writing %d tracks to %s', self.size, file_name)
         ds_combined.to_netcdf(file_name, encoding=encoding)
 
@@ -2435,3 +2435,18 @@ def set_category(max_sus_wind, wind_unit='kn', saffir_scale=None):
         return (np.argwhere(max_wind < saffir_scale) - 1)[0][0]
     except IndexError:
         return -1
+
+def _zlib_from_DataArray(data_var: xr.DataArray) -> bool:
+    """Return true if data_var is of numerical type, return False otherwise
+
+    Parameters
+    ----------
+    data_var : xr.DataArray
+
+    Returns
+    -------
+    bool
+    """
+    if np.issubdtype(data_var.dtype, float) or np.issubdtype(data_var.dtype, int):
+        return True
+    return False
