### 2024-07-10 32a0c9c
compute_windfields_sparse should not be a private function due to split in two modules
