### 2024-07-10 d79b670
create gdf meta data as named tuple
