+++ climada/util/plot.py
@@ -45,6 +45,7 @@ from cartopy.mpl.gridliner import LONGITUDE_FORMATTER, LATITUDE_FORMATTER
 from rasterio.crs import CRS
 import requests
 from geopandas import GeoDataFrame
+from collections import namedtuple
 
 from climada.util.constants import CMAP_EXPOSURES, CMAP_CAT, CMAP_RASTER
 from climada.util.files_handler import to_list
@@ -876,14 +877,20 @@ def multibar_plot(ax, data, colors=None, total_width=0.8, single_width=1,
     if legend:
         ax.legend(bars, data.keys())
 
+# Create GdfMeta class (as a named tuple) for GeoDataFrame meta data
+GdfMeta = namedtuple('GdfMeta', ['name', 'unit', 'col_name', 'col_unit'])
 
-def subplots_from_gdf(gdf, smooth=True, axis=None, figsize=(9, 13), adapt_fontsize=True, **kwargs):
+def subplots_from_gdf(gdf: GeoDataFrame, gdf_meta: GdfMeta = None, smooth=True, axis=None, figsize=(9, 13), adapt_fontsize=True, **kwargs):
         """Plot hazard local return periods for given hazard intensities.
     
         Parameters
         ----------
         gdf: gpd.GeoDataFrame
             return periods per threshold intensity
+        gdf_meta: 
+            climada.util.plot.GdfMeta
+            gdf meta data in named tuple with attributes 'name' (quantity in gdf), 'unit', (unit thereof)
+            'col_name' (quantity in column labels), 'col_unit' (unit thereof)
         smooth: bool, optional
             Smooth plot to plot.RESOLUTION x plot.RESOLUTION. Default is True
         axis: matplotlib.axes._subplots.AxesSubplot, optional
@@ -901,32 +908,41 @@ def subplots_from_gdf(gdf, smooth=True, axis=None, figsize=(9, 13), adapt_fontsi
         axis: matplotlib.axes._subplots.AxesSubplot
             Matplotlib axis with the plot.
         """
+        # check if inputs are correct types
         if not isinstance(gdf, GeoDataFrame):
             raise ValueError("gdf is not a GeoDataFrame")
         gdf = gdf[['geometry', *[col for col in gdf.columns if col != 'geometry']]]
-        try:
-            meta = {key: val for key, val in gdf.columns.name}
-            colbar_name = f"{meta['name']} ({meta['unit']})"
-            title_subplots = [f"{meta['col_name']}: {thres_inten} {meta['col_unit']}" 
+
+        # read meta data for fig and axis labels
+        if not isinstance(gdf_meta, GdfMeta):
+            #warnings.warn("gdf_meta variable is not of type climada.util.plot.GdfMeta. Figure and axis labels will be missing.")
+            print("gdf_meta variable is not of type climada.util.plot.GdfMeta. Figure and axis labels will be missing.")
+            colbar_name, title_subplots = None, [f"{col}" for col in gdf.columns[1:]] 
+        else:
+            colbar_name = f"{gdf_meta.name} ({gdf_meta.unit})"
+            title_subplots = [f"{gdf_meta.col_name}: {thres_inten} {gdf_meta.col_unit}" 
                               for thres_inten in gdf.columns[1:]]
-            
+        
             # change default plot kwargs if plotting return periods
-            if meta['name'] == 'Return Period':
-                if 'camp' not in kwargs.keys():
+            if gdf_meta.name == 'Return Period':
+                if 'cmap' not in kwargs.keys():
                     kwargs.update({'cmap': 'viridis_r'})
                 if 'norm' not in kwargs.keys():
                     kwargs.update(
[…]
