+++ CHANGELOG.md
@@ -59,8 +59,8 @@ latitude and longitude column are no longer persent there (the according arrays 
 - climada.hazard.centroids.centr.Centroids.empty_geometry_points
 - climada.hazard.centroids.centr.Centroids.set_meta_to_lat_lon
 - climada.hazard.centroids.centr.Centroids.set_lat_lon_to_meta
-- climada.entity.exposures.Exposures.set_lat_lon`
-- climada.entity.exposures.Exposures.set_geometry_points`
+- climada.entity.exposures.Exposures.set_lat_lon
+- climada.entity.exposures.Exposures.set_geometry_points
 
 ### Removed
 
