### 2024-07-11 5104503
introduce exposures properties, region_id etc. consequently use region_id property where applicable
