+++ climada/entity/measures/base.py
@@ -415,7 +415,7 @@ class Measure():
         if self.exp_region_id:
             # compute impact only in selected region
             in_reg = np.logical_or.reduce(
-                [exposures.gdf.region_id.values == reg for reg in self.exp_region_id]
+                [exposures.region_id == reg for reg in self.exp_region_id]
             )
             exp_imp = Exposures(exposures.gdf[in_reg], crs=exposures.crs)
         else:
