### 2024-07-11 cb666dc
rename and use hazard_impf and hazard_centroids
