+++ climada/entity/exposures/test/test_mat.py
@@ -41,32 +41,32 @@ class TestReader(unittest.TestCase):
         self.assertEqual(expo.gdf.index[0], 0)
         self.assertEqual(expo.gdf.index[n_expos - 1], n_expos - 1)
 
-        self.assertEqual(expo.gdf.value.shape, (n_expos,))
-        self.assertEqual(expo.gdf.value[0], 13927504367.680632)
-        self.assertEqual(expo.gdf.value[n_expos - 1], 12624818493.687229)
+        self.assertEqual(expo.value.shape, (n_expos,))
+        self.assertEqual(expo.value[0], 13927504367.680632)
+        self.assertEqual(expo.value[n_expos - 1], 12624818493.687229)
 
-        self.assertEqual(expo.gdf.deductible.shape, (n_expos,))
-        self.assertEqual(expo.gdf.deductible[0], 0)
-        self.assertEqual(expo.gdf.deductible[n_expos - 1], 0)
+        self.assertEqual(expo.deductible.shape, (n_expos,))
+        self.assertEqual(expo.deductible[0], 0)
+        self.assertEqual(expo.deductible[n_expos - 1], 0)
 
-        self.assertEqual(expo.gdf.cover.shape, (n_expos,))
-        self.assertEqual(expo.gdf.cover[0], 13927504367.680632)
-        self.assertEqual(expo.gdf.cover[n_expos - 1], 12624818493.687229)
+        self.assertEqual(expo.cover.shape, (n_expos,))
+        self.assertEqual(expo.cover[0], 13927504367.680632)
+        self.assertEqual(expo.cover[n_expos - 1], 12624818493.687229)
 
-        self.assertIn('int', str(expo.gdf.impf_.dtype))
-        self.assertEqual(expo.gdf.impf_.shape, (n_expos,))
-        self.assertEqual(expo.gdf.impf_[0], 1)
-        self.assertEqual(expo.gdf.impf_[n_expos - 1], 1)
+        self.assertIn('int', str(expo.hazard_impf().dtype))
+        self.assertEqual(expo.hazard_impf().shape, (n_expos,))
+        self.assertEqual(expo.hazard_impf()[0], 1)
+        self.assertEqual(expo.hazard_impf()[n_expos - 1], 1)
 
-        self.assertIn('int', str(expo.gdf.category_id.dtype))
-        self.assertEqual(expo.gdf.category_id.shape, (n_expos,))
-        self.assertEqual(expo.gdf.category_id[0], 1)
-        self.assertEqual(expo.gdf.category_id[n_expos - 1], 1)
+        self.assertIn('int', str(expo.category_id.dtype))
+        self.assertEqual(expo.category_id.shape, (n_expos,))
+        self.assertEqual(expo.category_id[0], 1)
+        self.assertEqual(expo.category_id[n_expos - 1], 1)
 
-        self.assertIn('int', str(expo.gdf.centr_.dtype))
-        self.assertEqual(expo.gdf.centr_.shape, (n_expos,))
-        self.assertEqual(expo.gdf.centr_[0], 47)
-        self.assertEqual(expo.gdf.centr_[n_expos - 1], 46)
+        self.assertIn('int', str(expo.hazard_centroids().dtype))
+        self.assertEqual(expo.hazard_centroids().shape, (n_expos,))
+        self.assertEqual(expo.hazard_centroids()[0], 47)
+        self.assertEqual(expo.hazard_centroids()[n_expos - 1], 46)
 
         self.assertTrue('region_id' not in expo.gdf)
 
