+++ climada/entity/exposures/test/test_base.py
@@ -77,11 +77,11 @@ class TestFuncs(unittest.TestCase):
             haz.centroids.gdf["lat"] = haz.centroids.lat.astype(test_dtype)
             haz.centroids.gdf["lon"] = haz.centroids.lon.astype(test_dtype)
             exp.assign_centroids(haz)
-            self.assertEqual(exp.gdf.shape[0], len(exp.gdf[INDICATOR_CENTR + 'FL']))
-            np.testing.assert_array_equal(exp.gdf[INDICATOR_CENTR + 'FL'].values, expected_result)
+            self.assertEqual(exp.gdf.shape[0], len(exp.hazard_centroids('FL')))
+            np.testing.assert_array_equal(exp.hazard_centroids('FL'), expected_result)
             exp.assign_centroids(Hazard(), overwrite=False)
-            self.assertEqual(exp.gdf.shape[0], len(exp.gdf[INDICATOR_CENTR + 'FL']))
-            np.testing.assert_array_equal(exp.gdf[INDICATOR_CENTR + 'FL'].values, expected_result)
+            self.assertEqual(exp.gdf.shape[0], len(exp.hazard_centroids('FL')))
+            np.testing.assert_array_equal(exp.hazard_centroids('FL'), expected_result)
 
     def test__init__meta_type(self):
         """ Check if meta of type list raises a ValueError in __init__"""
