+++ doc/joss/calibration-module/paper.bib
@@ -128,18 +128,19 @@
 }
 
 @article{riedel_fluvial_2024,
-	title = {Fluvial {Flood} {Inundation} and {Humanitarian} {Impact} {Model} {Based} {On} {Open} {Data}},
-	url = {https://egusphere.copernicus.org/preprints/2024/egusphere-2024-93/},
-	doi = {10.5194/egusphere-2024-93},
-	abstract = {{\textless}p{\textgreater}{\textless}strong class="journal-contentHeaderColor"{\textgreater}Abstract.{\textless}/strong{\textgreater} Fluvial floods are destructive hazards that affect millions of people worldwide each year. Forecasting flood events and their potential impacts therefore is crucial for disaster preparation and mitigation. Modeling flood inundation based on extreme value analysis of river discharges is an alternative to physical models of flood dynamics, which are computationally expensive. We present the implementation of a globally applicable, open-source fluvial flood model within a state-of-the-art natural catastrophe modeling framework. It uses openly available data to rapidly compute flood inundation footprints of historic and forecasted events for the estimation of associated impacts. At the example of Pakistan, we use this flood model to compute flood depths and extents, and employ it to estimate population displacement due to floods. Comparing flood extents to satellite data reveals that incorporating estimated flood protection standards does not necessarily improve the flood footprint computed by the model. We further show that, after calibrating the vulnerability of the impact model to a single event, the estimated displacement caused by past floods is in good agreement with disaster reports. Finally, we demonstrate that this calibrated model is suited for probabilistic impact-based forecasting.{\textless}/p{\textgreater}},
+	title = {Fluvial flood inundation and socio-economic impact model based on open data},
+	volume = {17},
+	issn = {1991-959X},
+	doi = {10.5194/gmd-17-5291-2024},
+	abstract = {Fluvial floods are destructive hazards that affect millions of people worldwide each year. Forecasting flood events and their potential impacts therefore is crucial for disaster preparation and mitigation. Modeling flood inundation based on extreme value analysis of river discharges is an alternative to physical models of flood dynamics, which are computationally expensive. We present the implementation of a globally applicable, open-source fluvial flood model within a state-of-the-art risk modeling framework. It uses openly available data to rapidly compute flood inundation footprints of historic and forecasted events for the estimation of associated impacts. For the example of Pakistan, we use this flood model to compute flood depths and extents and employ it to estimate population displacement due to floods. Comparing flood extents to satellite data reveals that incorporating estimated flood protection standards does not necessarily improve the flood footprint computed by the model. We further show that, after calibrating the vulnerability of the impact model to a single event, the estimated displacement caused by past floods is in good agreement with disaster reports. Finally, we demonstrate that this calibrated model is suited for probabilistic impact-based forecasting.},
 	language = {English},
-	urldate = {2024-04-29},
-	journal = {EGUsphere [preprint]},
+	number = {13},
+	urldate = {2024-07-10},
+	journal = {Geoscientific Model Development},
 	author = {Riedel, Lukas and R{\"o}{\"o}sli, Thomas and Vogt, Thomas and Bresch, David N.},
-	month = jan,
+	month = jul,
 	year = {2024},
-	pages = {1--29},
-	file = {Full Text PDF:/Users/ldr.riedel/Zotero/storage/RB72JJWL/Riedel et al. - 2024 - Fluvial Flood Inundation and Humanitarian Impact M.pdf:application/pdf},
+	pages = {5291--5308},
 }
 
 @misc{kam_impact-based_2023,
