### 2024-07-11 0183307
ci: Remove JOSS paper build job
