+++ .github/workflows/ci.yml
@@ -67,27 +67,6 @@ jobs:
           name: coverage-report-core-unittests-py${{ matrix.python-version }}
           path: coverage/
 
-  joss-paper:
-    runs-on: ubuntu-latest
-    name: Paper Draft
-    steps:
-      - name: Checkout
-        uses: actions/checkout@v4
-      - name: Build draft PDF
-        uses: openjournals/openjournals-draft-action@master
-        with:
-          journal: joss
-          # This should be the path to the paper within your repo.
-          paper-path: doc/joss/calibration-module/paper.md
-      - name: Upload
-        uses: actions/upload-artifact@v1
-        with:
-          name: paper
-          # This is the output path where Pandoc will write the compiled
-          # PDF. Note, this should be the same directory as the input
-          # paper.md
-          path: doc/joss/calibration-module/paper.pdf
-
   test-petals:
     name: Petals Compatibility
     uses: CLIMADA-project/climada_petals/.github/workflows/testing.yml@develop
