+++ climada/engine/option_appraisal/calc_impact_metrics.py
@@ -23,8 +23,105 @@ from .impact_trajectories import CalcImpactsSnapshots, SnapshotsCollection
 from .impact_metrics import ImpactMetrics
 import pandas as pd
 import numpy as np
+from climada.entity.measures import MeasureSet
 
 # BASE_arm_df_COLUMNS = ['measure', 'group', 'year', 'metric', 'result']
+
+
+class CalcImpactMetrics:
+
+    def __init__(self, snapshots, group_col_str=None):
+        self.snapshots = snapshots
+        # Get groups
+        self.group_map_exp_dict = create_group_map_exp_dict(snapshots, group_col_str) if group_col_str else None
+
+    # Calculate the annual risk metrics
+    def calc_arm_df(self, calc_static_scenario_metrics = True, previous_impact_metrics = None, measure_set = None):
+        # Calculate the risk metrics for the non-measure case
+        if previous_impact_metrics:
+            _df = previous_impact_metrics.arm_df.copy()
+        else: 
+            _df = calc_annual_risk_metrics(self.snapshots, group_map_exp_dict=self.group_map_exp_dict)
+
+        # Calculate the static risk metrics (used for waterfall levels)
+        if previous_impact_metrics and previous_impact_metrics.all_arms_df is not None:
+            _all_df = previous_impact_metrics.all_arms_df.copy()
+        elif calc_static_scenario_metrics:
+            _all_df = calc_annual_risk_metrics_static(self.snapshots, group_map_exp_dict=self.group_map_exp_dict)
+        else:
+            _all_df = None
+
+        # Calculate the risk metrics for the measure set
+        if measure_set:
+            # ADHOC!!! This should be done in a more elegant way
+            # Remove all the rows in the dataframe that have measure 'All'
+            _df = _df[_df['measure'] != 'All']
+            # Reduce the measure set if previous_impact_metrics is given
+            measure_set = reduce_meas_set(measure_set, previous_impact_metrics.measure_set) if previous_impact_metrics else measure_set
+            # Calculate the risk metrics for the measure set
+            _temp_df = calc_annual_risk_metrics_measure_set(self.snapshots, measure_set, group_map_exp_dict=self.group_map_exp_dict)
+            # Concatenate with arm_df
+            _df = pd.concat([_df, _temp_df], ignore_index=True)
+
+        return _df, _all_df
+
+    # Generate an impact metrics object
+    def generate_impact_metrics(self, measure_set=None, calc_static_scenario_metrics = True, previous_impact_metrics = None, planner = None, combine_all = True):
+
+        # Generate the measure times DataFrame
+        # Check if measure_times_df is given
+        if previous_impact_metrics and previous_impact_metrics.measure_times_df is not None:
+            measure_times_df = previous_impact_metrics.measure_times_df.copy()
+            # Update the measure times DataFrame based on the planner
+            if planner:
+                measure_times_df = update_measure_times_df(measure_times_df, planner)  
+        elif measure_set:
+            measure_times_df = generate_meas_times_df(measure_set, self.snapshots, planner) if measure_set else None
+        
+        # Generate the necessary measure set including the sub_combos
+        if measure_set and combine_all:
+            measure_set = generate_necessary_combo_measure_set(measure_set, measure_times_df)
+        else:
+            measure_set = measure_set if measure_set else None
+       
+        # Get the impact metrics
+        _df, _all_df = self.calc_arm_df(calc_static_scenario_metrics, previous_impact_metrics, measure_set)
+
+        # Get the exposure value units
+        exp_value_unit = self.snapshots.data[0].exposure.value_unit
+
[…]
