+++ climada/engine/option_appraisal/impact_metrics.py
@@ -30,8 +30,221 @@ import matplotlib.pyplot as plt
 import seaborn as sns
 
 
+
+
+class ImpactMetrics:
+
+    def __init__(self, 
+                 arm_df: pd.DataFrame,
+                 all_arms_df: Optional[pd.DataFrame] = None,
+                 measure_set: Optional[MeasureSet] = None,
+                 measure_times_df: Optional[pd.DataFrame] = None,
+                 value_unit: str = 'USD'
+                 ):
+        self.arm_df = arm_df.copy()
+        self.all_arms_df = copy.deepcopy(all_arms_df)
+        self.measure_set = copy.deepcopy(measure_set)
+        self.measure_times_df = copy.deepcopy(measure_times_df)
+        self.value_unit = value_unit 
+
+
+    ## Waterfall plot functions
+    # Plot the waterfall plot
+    def waterfall_plot(self, yearly=False, measure=None, metric='aai', group=np.nan):
+
+        # Check if the all_arms_df is provided
+        if self.all_arms_df is None:
+            raise ValueError("The all_arms_df is required for the waterfall plot. Please reggerate the all_arms_df using CalcImpactMetrics.")
+
+        # Check if yearly or classic waterfall plot
+        if yearly:
+            # Calculate the averted risk metrics for yearly waterfall plot
+            waterfall_df = _calc_waterfall_plot_df(self.arm_df, self.all_arms_df, measure=measure, metric=metric, ref_year=None, fut_year=None, group=group, subtract=True)
+            _plot_yearly_waterfall(waterfall_df, self.value_unit)
+        else:
+            # Calculate the averted risk metrics for the two points waterfall plot
+            waterfall_df = _calc_waterfall_plot_df(self.arm_df, self.all_arms_df, measure=measure, metric=metric, ref_year=None, fut_year=None, group=group)
+            _plot_two_years_waterfall(waterfall_df, self.value_unit)
+
+        return
+    
+    ## CB analysis
+    # Calculate the cost-benefit analysis
+    def calc_CB(self, start_year=None, end_year=None, consider_measure_times=True, risk_disc=None, cost_disc=None):
+        return calc_CB_df(self.arm_df, self.measure_set, self.measure_times_df, start_year, end_year, consider_measure_times, risk_disc, cost_disc)
+    
+    # Print the CB summary table
+    def print_CB_summary_table(self, metric = 'aai', start_year=None, end_year=None, consider_measure_times=True, risk_disc=None, cost_disc=None):
+        # Get the CB analysis
+        _, tot_CB_df = self.calc_CB(start_year, end_year, consider_measure_times, risk_disc, cost_disc)
+        print_CB_summary_table(tot_CB_df, self.measure_set, metric=metric, value_unit= self.value_unit)
+        return
+    
+    # Plot the yearly averted cost for a specific measure
+    def plot_benefit_vs_cost(self, measure, metric='aai', start_year=None, end_year=None, consider_measure_times=True, risk_disc=None, cost_disc=None):
+        ann_CB_df, tot_CB_df = self.calc_CB(start_year, end_year, consider_measure_times, risk_disc, cost_disc)
+        plot_yearly_averted_cost(ann_CB_df, measure, metric, group=None)
+        return
+    
+    ## Risk analysis
+    # Plot the risk metrics for each measure
+    def plot_yearly_risk_metrics(self, metric='aai', averted= False, consider_measure_times=False, risk_disc=None, plot_type='line', measures=None, group=np.nan):
+        # Copy the DataFrame
+        plot_df = self.arm_df.copy()
+
+        # Check if the time should be considered
+        if consider_measure_times:
+            plot_df = create_meas_mod_arm_df(plot_df, self.measure_set, self.measure_times_df, consider_measure_times=True)
+
[…]
