### 2024-07-11 5198646
updated docstrings and variable frequency unit
