### 2024-07-11 fe99121
remove obsolete plot function
