### 2024-07-11 759ce99
Apply suggestions from code review
Co-authored-by: Lukas Riedel <34276446+peanutfun@users.noreply.github.com>
