### 2024-07-11 b3d85d6
Apply code review suggestions
Co-authored-by: Lukas Riedel <34276446+peanutfun@users.noreply.github.com>
