+++ climada/hazard/trop_cyclone/trop_cyclone.py
@@ -41,7 +41,6 @@ from climada.hazard.tc_clim_change import get_knutson_criterion, calc_scale_knut
 from climada.hazard.centroids.centr import Centroids
 from .trop_cyclone_windfields import DEF_MAX_DIST_EYE_KM, DEF_INTENSITY_THRES, \
     DEF_MAX_MEMORY_GB, compute_windfields_sparse
-from climada.util import ureg
 import climada.util.constants as u_const
 import climada.util.coordinates as u_coord
 import climada.util.plot as u_plot
