### 2024-07-11 eab38ac
updated fit func and new loc_exceedance func
