+++ climada/util/fit_methods.py
@@ -22,16 +22,16 @@ define fit functionalities for (local) exceedance frequencies and return periods
 
 import numpy as np
 
-def calc_fit_interp(x_test, x_train, y_train, method=None, x_scale=None, y_scale=None):
+def calc_fit_interp(x_test, x_train: np.array, y_train: np.array, method=None, x_scale=None, y_scale=None, x_thres = None, y_thres = None):
     """Save variable with provided file name. Uses configuration save_dir folder
     if no absolute path provided.
 
     Parameters
     ----------
     x : np.array
-        x values 
+        x values (1-D array of x values to fit)
     y : np.array
-        y values
+        y values (1-D array of y values to fir)
     method : object
         variable to save in pickle format
 
@@ -45,37 +45,42 @@ def calc_fit_interp(x_test, x_train, y_train, method=None, x_scale=None, y_scale
         raise ValueError(f'Unknown method: {method}. Use "interp" or "fit" instead')
     if x_train.shape != y_train.shape:
         raise ValueError(f'Incompatible shapes of input data, x_train {x_train.shape} and y_train {y_train.shape}. Should be the same')
-    if len(x_train.shape) == 1:
-        x_train, y_train = x_train.reshape(-1,1), y_train.reshape(-1,1)
     
+    # cut x and y above threshold
+    if x_thres:
+        x_th = np.asarray(x_train > x_thres).squeeze()
+        x_train = x_train[x_th]
+        y_train = y_train[x_th]
     
+    if y_thres:
+        y_th = np.asarray(y_train > y_thres).squeeze()
+        x_train = x_train[y_th]
+        y_train = y_train[y_th]
+
+
     # adapt x and y scale
     if x_scale == 'log':
         x_train, x_test = np.log10(x_train), np.log10(x_test)
     if y_scale == 'log':
         y_train = np.log10(y_train)
 
+
     # calculate fit/interpolation
-    y_test = np.zeros((x_test.shape[0], x_train.shape[1]))
-    for i in range(x_train.shape[1]):
-        if method == 'interp':
-            y_test[:, i] = np.interp(x_test, x_train[:, i], y_train[:, i])
-        elif method == 'fit':
-            try: 
-                pol_coef = np.polyfit(x_train[:, i], y_train[:, i], deg=1)
-            except ValueError:
-                pol_coef = np.polyfit(x_train[:, i], y_train[:, i], deg=0)
-            y_test[:, i] = np.polyval(pol_coef, x_test)
+    if method == 'interp':
+        if not np.all(x_train[:-1] <= x_train[1:]):
+            raise ValueError(f'Input array x_train must be sorted.')
+        y_test = np.interp(x_test, x_train, y_train)
+    elif method == 'fit':
+        try: 
+            pol_coef = np.polyfit(x_train, y_train, deg=1)
+        except ValueError:
+            pol_coef = np.polyfit(x_train, y_train, deg=0)
+        y_test = np.polyval(pol_coef, x_test)
 
     # adapt output scale
     if y_scale == 'log':
         y_test = np.exp(y_test)
     
-    return np.squeeze(y_test)
-
-def local_fit_interp(x_test, x_train, y_train, method=None, x_scale=None, y_scale=None):
-
-    return    
-
+    return y_test
     
 
