### 2024-07-12 b93842e
Split tropical_cyclone module into submodules (#911)
* split tropical_cyclone module
* fix petals 3.11 compatibility error (petals issue 135)
* compute_windfields_sparse should not be a private function due to split in two modules
* Apply suggestions from code review
* Include docs of split module
* Update CHANGELOG.md

---------

Co-authored-by: Thomas Roosli <thomas.roeoesli@meteoswiss.ch>
Co-authored-by: Lukas Riedel <34276446+peanutfun@users.noreply.github.com>
