### 2024-07-12 1511cc1
Fix CHANGELOG.md and add entry in Citation guide
