### 2024-07-12 #6755 @peanutfun
@observingClouds

We updated one reference to a preprint in the manuscript, because the associated publication was released. I'll ask the bot to check references again.

> * Double check authors and affiliations (including ORCIDs)

We double checked.

> * Make a release of the software with the latest changes from the review and post the version number here. This is the version that will be used in the JOSS paper.

We will merge the reviewed pull request and make a release. This will probably take place next week, and I'll report back afterward.

> * Archive the release on Zenodo/figshare/etc and post the DOI here.
> * Make sure that the title and author list (including ORCIDs) in the archive match those in the JOSS paper.

All Climada releases are automatically released on this Zenodo record: https://zenodo.org/doi/10.5281/zenodo.4598943 The author list is intended to reflect the authors of the entire repository. The JOSS author papers are included in this list. Does that suffice, or should we create a separate Zenodo record for the new release?
