### 2024-07-12 316ca45
added local_rp fct and stepfunction fit
