+++ climada/util/fit_methods.py
@@ -22,7 +22,16 @@ define fit functionalities for (local) exceedance frequencies and return periods
 
 import numpy as np
 
-def calc_fit_interp(x_test, x_train: np.array, y_train: np.array, method=None, x_scale=None, y_scale=None, x_thres = None, y_thres = None):
+def calc_fit_interp(
+        x_test, 
+        x_train: np.array, 
+        y_train: np.array, 
+        method=None, 
+        x_scale=None, 
+        y_scale=None, 
+        x_thres = None, 
+        y_thres = None
+    ):
     """Save variable with provided file name. Uses configuration save_dir folder
     if no absolute path provided.
 
@@ -39,10 +48,11 @@ def calc_fit_interp(x_test, x_train: np.array, y_train: np.array, method=None, x
     -------
     np.array
     """
+    # check if inputs are valid
     if not method:
         method = 'interp'
-    if not method in ['interp', 'fit']:
-        raise ValueError(f'Unknown method: {method}. Use "interp" or "fit" instead')
+    if not method in ['interp', 'fit', 'stepfunction']:
+        raise ValueError(f'Unknown method: {method}. Use "interp", "fit", or "stepfunction" instead')
     if x_train.shape != y_train.shape:
         raise ValueError(f'Incompatible shapes of input data, x_train {x_train.shape} and y_train {y_train.shape}. Should be the same')
     
@@ -65,17 +75,28 @@ def calc_fit_interp(x_test, x_train: np.array, y_train: np.array, method=None, x
         y_train = np.log10(y_train)
 
 
-    # calculate fit/interpolation
+    # calculate interpolation
     if method == 'interp':
         if not np.all(x_train[:-1] <= x_train[1:]):
             raise ValueError(f'Input array x_train must be sorted.')
         y_test = np.interp(x_test, x_train, y_train)
+
+    # calculate linear fit
     elif method == 'fit':
         try: 
             pol_coef = np.polyfit(x_train, y_train, deg=1)
         except ValueError:
             pol_coef = np.polyfit(x_train, y_train, deg=0)
         y_test = np.polyval(pol_coef, x_test)
+    
+    # calculate stepfunction fit
+    elif method == 'stepfunction':
+        # find indeces of x_test if sorted into x_train
+        if not np.all(x_train[:-1] <= x_train[1:]):
+            raise ValueError(f'Input array x_train must be sorted.')
+        indx = np.searchsorted(x_train, x_test)
+        y_test = y_train[indx.clip(max = len(x_train) - 1)]
+        y_test[indx == len(x_train)] = np.nan
 
     # adapt output scale
     if y_scale == 'log':
