+++ climada/hazard/base.py
@@ -404,7 +404,10 @@ class Hazard(HazardIO, HazardPlot):
             u_coord.latlon_bounds(lat=lat_nz, lon=lon_nz, buffer=buffer)
         ))
 
-    def local_exceedance_inten(self, return_periods=(25, 50, 100, 250)):
+    def local_exceedance_inten(
+            self, 
+            return_periods=(25, 50, 100, 250),
+            ):
         """Compute exceedance intensity map for given return periods.
 
         Parameters
@@ -460,33 +463,186 @@ class Hazard(HazardIO, HazardPlot):
         """
         num_cen = self.intensity.shape[1]
         inten_stats = np.zeros((len(return_periods), num_cen))
-        
-        sort_pos = np.argsort(self.intensity.toarray(), axis=0)[::-1, :]
-        columns = np.ones(sort_pos.shape, int)
-        columns *= np.arange(num_cen)
-        inten_sort = self.intensity.toarray()[sort_pos, columns]
-        freq_sort = self.frequency[sort_pos]
-        freq_sort = np.cumsum(freq_sort, axis=0)
-
-        # inten_sort = self.intensity.toarray()[sort_pos, np.arange(num_cen)]
-        # freq_sort = self.frequency[sort_pos]
-        # freq_sort = np.cumsum(freq_sort, axis=0)
-        
-        inten_stats = np.zeros((len(return_periods), num_cen))
-
-        
 
         for i in range(num_cen):
+            # sort intensties and frequencies at given centroid
+            sorted_idxs = np.argsort(np.squeeze(self.intensity[:,i].toarray()))[::-1]
+            intensity_sorted = np.squeeze(self.intensity[:,i].toarray()[sorted_idxs])
+            frequency_sorted = self.frequency[sorted_idxs]
+            frequency_sorted = np.cumsum(frequency_sorted)
+
+            # fit intensities to cummulative frequencies
             inten_stats[:,i] = u_fit.calc_fit_interp(
                 1/np.array(return_periods),
-                freq_sort[:,i],
-                inten_sort[:,i],
+                frequency_sorted,
+                intensity_sorted,
                 method='fit',
                 x_scale='log',
                 y_thres=self.intensity_thres
             )
 
         return inten_stats
+
+    # def local_return_period(self, threshold_intensities):
+    #     """Compute local return periods for given hazard intensities. The used method is fitting the 
+    #     ordered intensitites per centroid to the corresponding cummulated frequency with a step function
+
+    #     Parameters
+    #     ----------
+    #     threshold_intensities : np.array
+    #         User-specified hazard intensities for which the return period should be calculated 
+    #         locally (at each centroid)
+
+    #     Returns
+    #     -------
+    #     gdf : gpd.GeoDataFrame
+    #         GeoDataFrame containing return periods for given threshold intensities. Each column 
+    #         corresponds to a threshold_intensity value, each row corresponds to a centroid. Values
+    #         in the gdf correspond to the return period for the given centroid and threshold_intensity value
+    #     gdf_meta : climada.util.plot.GdfMeta
+    #         named tuple providing meta data of which quantities (and their units) are written the gdf
+    #     """
+    #     #check frequency unit
+    #     if self.frequency_unit in ['1/year', 'annual', '1/y', '1/a']:
+    #         rp_unit = 'Years'
+    #     elif self.frequency_unit in ['1/month', 'monthly', '1/m']:
+    #         rp_unit = 'Months'
+    #     elif self.frequency_unit in ['1/week', 'weekly', '1/w']:
+    #         rp_unit = 'Weeks'
+    #     else:
+    #         LOGGER.warning(f"Hazard's frequency unit {self.frequency_unit} is not known, "
+    #                         "years will be used as return period unit.")
+    #         rp_unit = 'Years'
+
[…]
