### 2024-07-12 9538bb5
Add module for calibrating impact functions (#692)
* Add climada.util.calibrate module with tests
* Add calibration tutorial
* Add JOSS publication
* Add 'bayesian-optimization' and  'seaborn' as dependencies
* Update CHANGELOG.md

---------

Co-authored-by: emanuel-schmid <schmide@ethz.ch>
Co-authored-by: Thomas Vogt <57705593+tovogt@users.noreply.github.com>
Co-authored-by: Timo Schmid <55481374+timschmi95@users.noreply.github.com>
Co-authored-by: Schmid  Timo <timschmi@ethz.ch>
Co-authored-by: Chahan M. Kropf <chahan.kropf@usys.ethz.ch>
