### 2024-07-12 30764c0
coordinates, dask may or not return a Polygon array (could be str)
