### 2024-07-12 8dc544f
divide gdf metadata into label and colum_label
