### 2024-07-15 79ab66f
Revert changes to attribute structure of Hazard
* Only call `Hazard.check_matrices` from `ImpactCalc.__init__`.
* Update tests and docstrings accordingly.
