### 2024-07-15 7f9649d
update local_exceedance_imp func
