### 2024-07-15 8b0d482
updated test, docstrings, var names, and tutorial
