+++ climada/hazard/base.py
@@ -451,21 +451,23 @@ class Hazard(HazardIO, HazardPlot):
             self.event_id = np.arange(1, self.event_id.size + 1)
 
     def local_return_period(self, threshold_intensities=(5., 10., 20.)):
-        """Compute local return periods for given hazard intensities. The used method is fitting the 
-        ordered intensitites per centroid to the corresponding cummulated frequency with a step function
-    
+        """Compute local return periods for given hazard intensities. The used method
+        is fitting the ordered intensitites per centroid to the corresponding cummulated
+        frequency with a step function.
+
         Parameters
         ----------
         threshold_intensities : np.array
-            User-specified hazard intensities for which the return period should be calculated 
+            User-specified hazard intensities for which the return period should be calculated
             locally (at each centroid). Defaults to (5, 10, 20)
-    
+
         Returns
         -------
         gdf : gpd.GeoDataFrame
-            GeoDataFrame containing return periods for given threshold intensities. Each column 
+            GeoDataFrame containing return periods for given threshold intensities. Each column
             corresponds to a threshold_intensity value, each row corresponds to a centroid. Values
-            in the gdf correspond to the return period for the given centroid and threshold_intensity value
+            in the gdf correspond to the return period for the given centroid and
+            threshold_intensity value
         label : str
             GeoDataFrame label, for reporting and plotting
         column_label : function
@@ -479,22 +481,23 @@ class Hazard(HazardIO, HazardPlot):
         elif self.frequency_unit in ['1/week', 'weekly', '1/w']:
             rp_unit = 'Weeks'
         else:
-            LOGGER.warning(f"Hazard's frequency unit {self.frequency_unit} is not known, "
-                           "years will be used as return period unit.")
+            LOGGER.warning("Hazard's frequency unit %s is not known, "
+                           "years will be used as return period unit.", self.frequency_unit)
             rp_unit = 'Years'
 
         # Ensure threshold_intensities is a numpy array
         threshold_intensities = np.array(threshold_intensities)
-        
+
         num_cen = self.intensity.shape[1]
         return_periods = np.zeros((len(threshold_intensities), num_cen))
-        
-        # batch_centroids = number of centroids that are handled in parallel: maximal matrix size // number of events
-        batch_centroids = CONFIG.max_matrix_size.int() // self.intensity.shape[0] 
+
+        # batch_centroids = number of centroids that are handled in parallel: 
+        # batch_centroids = maximal matrix size // number of events
+        batch_centroids = CONFIG.max_matrix_size.int() // self.intensity.shape[0]
         if batch_centroids < 1:
             raise ValueError('Increase max_matrix_size configuration parameter to >'
                              f'{self.intensity.shape[0]}')
-        
+
         # Process the intensities in chunks of centroids
         for start_col in range(0, num_cen, batch_centroids):
             end_col = min(start_col + batch_centroids, num_cen)
@@ -502,18 +505,20 @@ class Hazard(HazardIO, HazardPlot):
                 threshold_intensities,
                 self.intensity[:, start_col:end_col].toarray()
                 )
-        
+
         # create the output GeoDataFrame
-        gdf = gpd.GeoDataFrame(geometry = self.centroids.gdf['geometry'], crs = self.centroids.gdf.crs)
+        gdf = gpd.GeoDataFrame(geometry = self.centroids.gdf['geometry'],
+                               crs = self.centroids.gdf.crs)
         col_names = [f'{tresh_inten}' for tresh_inten in threshold_intensities]
         gdf[col_names] = return_periods.T
 
         # create label and column_label
         label = f'Return Periods ({rp_unit})'
-        column_label = lambda column_names: [f'Threshold Intensity: {col} {self.units}' for col in column_names]
[…]
