+++ climada/hazard/io.py
@@ -27,7 +27,6 @@ import pathlib
 from typing import Union, Optional, Callable, Dict, Any
 
 import h5py
-import netCDF4 as nc
 import numpy as np
 import pandas as pd
 import rasterio
