+++ climada/util/plot.py
@@ -877,75 +877,78 @@ def multibar_plot(ax, data, colors=None, total_width=0.8, single_width=1,
         ax.legend(bars, data.keys())
 
 def subplots_from_gdf(
-        gdf: gpd.GeoDataFrame, 
+        gdf: gpd.GeoDataFrame,
         colorbar_name: str = None,
         title_subplots: callable = None,
-        smooth=True, 
-        axis=None, 
-        figsize=(9, 13), 
-        adapt_fontsize=True, 
-        **kwargs   
+        smooth=True,
+        axis=None,
+        figsize=(9, 13),
+        adapt_fontsize=True,
+        **kwargs
 ):
-        """Plot hazard local return periods for given hazard intensities.
-    
-        Parameters
-        ----------
-        gdf: gpd.GeoDataFrame
-            return periods per threshold intensity
-        colorbar_name: str
-            title of the subplots' colorbars
-        title_subplots: function
-            function that generates the titles of the different subplots using the columns' names
-        smooth: bool, optional
-            Smooth plot to plot.RESOLUTION x plot.RESOLUTION. Default is True
-        axis: matplotlib.axes._subplots.AxesSubplot, optional
-            Axis to use. Default is None
-        figsize: tuple, optional
-            Figure size for plt.subplots. Default is (9, 13)
-        adapt_fontsize: bool, optional
-            If set to true, the size of the fonts will be adapted to the size of the figure.
-            Otherwise the default matplotlib font size is used. Default is True.
-        kwargs: optional
-            Arguments for pcolormesh matplotlib function used in event plots.
-    
-        Returns
-        -------
-        axis: matplotlib.axes._subplots.AxesSubplot
-            Matplotlib axis with the plot.
-        """
-        # check if inputs are correct types
-        if not isinstance(gdf, gpd.GeoDataFrame):
-            raise ValueError("gdf is not a GeoDataFrame")
-        gdf = gdf[['geometry', *[col for col in gdf.columns if col != 'geometry']]]
+    """Plot several subplots from different columns of a GeoDataFrame, e.g., for 
+    plotting local return periods or local exceedance intensities.
 
-        # read meta data for fig and axis labels
-        if not isinstance(colorbar_name, str):
-            print("Unknown colorbar name. Colorbar label will be missing.")
-            colorbar_name = ''
-        if not callable(title_subplots):
-            print("Unknown subplot-title-generation function. Subplot titles will be column names.")
-            title_subplots = lambda cols: [f"{col}" for col in cols] 
-        
-        # change default plot kwargs if plotting return periods
-        if colorbar_name.strip().startswith('Return Period'):
-            if 'cmap' not in kwargs.keys():
-                kwargs.update({'cmap': 'viridis_r'})
-            if 'norm' not in kwargs.keys():
-                kwargs.update(
-                    {'norm': mpl.colors.LogNorm(vmin=gdf.values[:,1:].min(), vmax=gdf.values[:,1:].max()),
-                    'vmin': None, 'vmax': None}
-                )
+    Parameters
+    ----------
+    gdf: gpd.GeoDataFrame
+        return periods per threshold intensity
+    colorbar_name: str
+        title of the subplots' colorbars
+    title_subplots: function
+        function that generates the titles of the different subplots using the columns' names
+    smooth: bool, optional
+        Smooth plot to plot.RESOLUTION x plot.RESOLUTION. Default is True
+    axis: matplotlib.axes._subplots.AxesSubplot, optional
+        Axis to use. Default is None
+    figsize: tuple, optional
[…]
