### 2024-07-15 0e50ebe
two more pylint changes
