+++ climada/hazard/base.py
@@ -41,6 +41,7 @@ import climada.util.constants as u_const
 import climada.util.coordinates as u_coord
 import climada.util.dates_times as u_dt
 
+
 LOGGER = logging.getLogger(__name__)
 
 
