### 2024-07-15 482a207
change fcts to return gdfs
