+++ climada/hazard/base.py
@@ -404,51 +404,6 @@ class Hazard(HazardIO, HazardPlot):
             u_coord.latlon_bounds(lat=lat_nz, lon=lon_nz, buffer=buffer)
         ))
 
-    # def local_exceedance_inten(
-    #         self, 
-    #         return_periods=(25, 50, 100, 250),
-    #         ):
-    #     """Compute exceedance intensity map for given return periods.
-
-    #     Parameters
-    #     ----------
-    #     return_periods : np.array
-    #         return periods to consider
-
-    #     Returns
-    #     -------
-    #     inten_stats: np.array
-    #     """
-    #     # warn if return period is above return period of rarest event:
-    #     for period in return_periods:
-    #         if period > 1 / self.frequency.min():
-    #             LOGGER.warning('Return period %1.1f exceeds max. event return period.', period)
-    #     LOGGER.info('Computing exceedance intenstiy map for return periods: %s',
-    #                 return_periods)
-    #     num_cen = self.intensity.shape[1]
-    #     inten_stats = np.zeros((len(return_periods), num_cen))
-    #     cen_step = CONFIG.max_matrix_size.int() // self.intensity.shape[0]
-    #     if not cen_step:
-    #         raise ValueError('Increase max_matrix_size configuration parameter to >'
-    #                          f' {self.intensity.shape[0]}')
-    #     # separte in chunks
-    #     chk = -1
-    #     for chk in range(int(num_cen / cen_step)):
-    #         self._loc_return_inten(
-    #             np.array(return_periods),
-    #             self.intensity[:, chk * cen_step:(chk + 1) * cen_step].toarray(),
-    #             inten_stats[:, chk * cen_step:(chk + 1) * cen_step])
-    #     self._loc_return_inten(
-    #         np.array(return_periods),
-    #         self.intensity[:, (chk + 1) * cen_step:].toarray(),
-    #         inten_stats[:, (chk + 1) * cen_step:])
-    #     # set values below 0 to zero if minimum of hazard.intensity >= 0:
-    #     if np.min(inten_stats) < 0 <= self.intensity.min():
-    #         LOGGER.warning('Exceedance intenstiy values below 0 are set to 0. \
-    #                Reason: no negative intensity values were found in hazard.')
-    #         inten_stats[inten_stats < 0] = 0
-    #     return inten_stats
-    
     def local_exceedance_inten(self, return_periods=(25, 50, 100, 250)):
         """Compute exceedance intensity map for given return periods.
 
@@ -461,13 +416,26 @@ class Hazard(HazardIO, HazardPlot):
         -------
         inten_stats: np.array
         """
+        #check frequency unit
+        if self.frequency_unit in ['1/year', 'annual', '1/y', '1/a']:
+            rp_unit = 'years'
+        elif self.frequency_unit in ['1/month', 'monthly', '1/m']:
+            rp_unit = 'months'
+        elif self.frequency_unit in ['1/week', 'weekly', '1/w']:
+            rp_unit = 'weeks'
+        else:
+            LOGGER.warning(f"Hazard's frequency unit {self.frequency_unit} is not known, "
+                            "years will be used as return period unit.")
+            rp_unit = 'years'
+        
         num_cen = self.intensity.shape[1]
         inten_stats = np.zeros((len(return_periods), num_cen))
 
         for i in range(num_cen):
             # sort intensties and frequencies at given centroid
-            sorted_idxs = np.argsort(np.squeeze(self.intensity[:,i].toarray()))[::-1]
-            intensity_sorted = np.squeeze(self.intensity[:,i].toarray()[sorted_idxs])
+            intensity_sorted = np.squeeze(self.intensity[:,i].toarray())
+            sorted_idxs = np.argsort(intensity_sorted)[::-1]
[…]
