+++ climada/hazard/base.py
@@ -461,72 +461,6 @@ class Hazard(HazardIO, HazardPlot):
 
         return gdf, label, column_label
 
-    def local_return_period(self, threshold_intensities):
-        """Compute local return periods for given hazard intensities. The used method is fitting the 
-        ordered intensitites per centroid to the corresponding cummulated frequency with a step function
-
-        Parameters
-        ----------
-        threshold_intensities : np.array
-            User-specified hazard intensities for which the return period should be calculated 
-            locally (at each centroid)
-
-        Returns
-        -------
-        gdf : gpd.GeoDataFrame
-            GeoDataFrame containing return periods for given threshold intensities. Each column 
-            corresponds to a threshold_intensity value, each row corresponds to a centroid. Values
-            in the gdf correspond to the return period for the given centroid and threshold_intensity value
-        gdf_meta : climada.util.plot.GdfMeta
-            named tuple providing meta data of which quantities (and their units) are written the gdf
-        """
-        #check frequency unit
-        if self.frequency_unit in ['1/year', 'annual', '1/y', '1/a']:
-            rp_unit = 'Years'
-        elif self.frequency_unit in ['1/month', 'monthly', '1/m']:
-            rp_unit = 'Months'
-        elif self.frequency_unit in ['1/week', 'weekly', '1/w']:
-            rp_unit = 'Weeks'
-        else:
-            LOGGER.warning(f"Hazard's frequency unit {self.frequency_unit} is not known, "
-                            "years will be used as return period unit.")
-            rp_unit = 'Years'
-
-        # Ensure threshold_intensities is a numpy array
-        threshold_intensities = np.array(threshold_intensities)
-        
-        num_cen = self.intensity.shape[1]
-        return_periods = np.zeros((len(threshold_intensities), num_cen))
-        
-        for i in range(num_cen):
-            # sort intensties and frequencies at given centroid
-            intensity_sorted = np.squeeze(self.intensity[:,i].toarray())
-            sorted_idxs = np.argsort(intensity_sorted)[::-1]
-            intensity_sorted = np.squeeze(intensity_sorted[sorted_idxs])
-            frequency_sorted = self.frequency[sorted_idxs]
-            frequency_sorted = np.cumsum(frequency_sorted)
-
-            # fit intensities to cummulative frequencies
-            return_periods[:,i] = np.divide(1., u_fit.calc_fit_interp(
-                threshold_intensities,
-                intensity_sorted[::-1],
-                frequency_sorted[::-1],
-                method='stepfunction',
-                x_scale='log'
-            ))
-        
-        # create the output GeoDataFrame
-        gdf = gpd.GeoDataFrame(geometry = self.centroids.gdf['geometry'], crs = self.centroids.gdf.crs)
-        col_names = [f'{tresh_inten}' for tresh_inten in threshold_intensities]
-        gdf[col_names] = return_periods.T
-
-        # create label and column_label
-        label = f'Return Periods ({rp_unit})'
-        column_label = lambda column_names: [f'Threshold Intensity: {col} {self.units}'
-                                             for col in column_names]
-
-        return gdf, label, column_label
-
     def sanitize_event_ids(self):
         """Make sure that event ids are unique"""
         if np.unique(self.event_id).size != self.event_id.size:
