### 2024-07-16 acf3664
made function calls consistent with new structure
