+++ climada/hazard/plot.py
@@ -32,6 +32,7 @@ class HazardPlot():
     Contains all plotting methods of the Hazard class
     """
 
+    #TODO to be deleted or added depreciating warning
     def plot_rp_intensity(self, return_periods=(25, 50, 100, 250),
                           smooth=True, axis=None, figsize=(9, 13), adapt_fontsize=True,
                           **kwargs):
@@ -56,7 +57,7 @@ class HazardPlot():
         axis, inten_stats:  matplotlib.axes._subplots.AxesSubplot, np.ndarray
             intenstats is return_periods.size x num_centroids
         """
-        inten_stats = self.local_exceedance_inten(np.array(return_periods))
+        inten_stats, _, _ = self.local_exceedance_inten(np.array(return_periods))[0].values[:,1:].T
         colbar_name = 'Intensity (' + self.units + ')'
         title = list()
         for ret in return_periods:
