### 2024-07-17 088b7e5
readthedocs: add hazard.plot and hazard.io
