### 2024-07-17 5f23ccf
readthedocs: remove entity.tag and hazard.tag
