### 2024-07-17 700862c
Ensure csr matrices are in "canonical format" before impact calculation (#893)
* Add util function for pruning csr_matrices.
* Add method for checking matrices to Hazard.
* Check matrices when instantiating ImpactCalc.
* Update unit tests.
* Update CHANGELOG.md.

---------

Co-authored-by: Chahan M. Kropf <chahan.kropf@usys.ethz.ch>
