### 2024-07-17 4e4f817
Feature/set geometry bis (#912)
* Add geodataframe to centroids: first commit

* Remove raster methods, obsolete methods, and fix some.

* Remove test for removed methods

* Fix from geodataframe method

* Remove set meta to lat lon

* Update all I/O methods

* Remove area pixel

* Add method to compute the area per pixel (not attribute)

* Make get_area_pixel

* Remove not often used dist_coast and elevation

* Remove unused meta

* Change Centroids call to new signature

* Remove vector - raster hazard methods

* Restore legacy excell reader

* Remove clear method

* Remove clear test

* Upd
[…]
