### 2024-07-17 cb62ba2
ix documentation (#922)
* readthedocs: add hazard.plot and hazard.io
* readthedocs: remove entity.tag and hazard.tag
