### 2024-07-17 642fb8c
Add river flood module citation to guide
