+++ doc/misc/citation.rst
@@ -31,6 +31,8 @@ If you use specific tools and modules of CLIMADA, please cite the appropriate pu
      - Eberenz, S., et al. (2020): Asset exposure data for global physical risk assessment. Earth System Science Data 12, 817–833, https://doi.org/10.3929/ethz-b-000409595
    * - :doc:`Impact function calibration </tutorial/climada_util_calibrate>`
      - Riedel, L., et al. (2024): A Module for Calibrating Impact Functions in the Climate Risk Modeling Platform CLIMADA, Journal of Open Source Software [`under review <https://joss.theoj.org/papers/50845e31c6cb6894baae492bfd853671>`_]
+   * - `GloFAS River Flood Module <https://climada-petals.readthedocs.io/en/latest/tutorial/climada_hazard_glofas_rf.html>`_
+     - Riedel, L. et al. (2024): Fluvial flood inundation and socio-economic impact model based on open data, Geosci. Model Dev., 17, 5291–5308, https://doi.org/10.5194/gmd-17-5291-2024
 
 Please find the code to reprocduce selected CLIMADA-related scientific publications in our `repository of scientific publications <https://github.com/CLIMADA-project/climada_papers>`_.
 
