+++ doc/misc/climada_publications.bib
@@ -71,3 +71,16 @@
   issn = {0951-8320},
   doi = {10.1016/j.ress.2023.109194}
 }
+
+@article{Riedel2024,
+  title = {Fluvial flood inundation and socio-economic impact model based on open data},
+  author = {Riedel, Lukas and R{\"o}{\"o}sli, Thomas and Vogt, Thomas and Bresch, David N.},
+  journal = {Geoscientific Model Development},
+  volume = {17},
+  issn = {1991-959X},
+  doi = {10.5194/gmd-17-5291-2024},
+  number = {13},
+  month = jul,
+  year = {2024},
+  pages = {5291--5308},
+}
