### 2024-07-17 1e0d238
plot NaNs in gray in geo_im_from_array
