### 2024-07-18 241c548
Remove tests that force the use of vmax_10m parameter
