### 2024-07-18 332b0ab
added method options for exceed funcs
