+++ climada/util/test/test_plot.py
@@ -152,7 +152,7 @@ class TestPlots(unittest.TestCase):
         self.assertEqual(cmap, ax.collections[0].cmap.name)
         plt.close()
 
-    def test_subplots_from_gdf(self):
+    def test_plot_from_gdf(self):
         return_periods = gpd.GeoDataFrame(
             data = ((2., 5.), (3., 6.), (None, 2.), (1., 7.)),
             columns = ('10.0', '20.0')
@@ -160,7 +160,7 @@ class TestPlots(unittest.TestCase):
         return_periods['geometry'] = (Point(45., 26.), Point(46., 26.), Point(45., 27.), Point(46., 27.))
         colorbar_name = 'Return Periods (Years)'
         title_subplots = lambda cols: [f'Threshold Intensity: {col} m/s' for col in cols]
-        (axis1, axis2) = u_plot.subplots_from_gdf(
+        (axis1, axis2) = u_plot.plot_from_gdf(
             return_periods, 
             colorbar_name=colorbar_name, 
             title_subplots=title_subplots)
