### 2024-07-18 f0bf633
fixed bug when y_scale is log
