### 2024-07-18 b09a7c6
added some tests for fit function
