### 2024-07-19 0c6377c
group values with same inten/impact
