### 2024-07-22 2790fdd
added util function to group frequencies
