### 2024-07-22 bf63b86
Fixed measure cost discount
