### 2024-07-22 #6755 @peanutfun
@observingClouds We released [Climada v5.0.0](https://github.com/CLIMADA-project/climada_python/releases/tag/v5.0.0), which includes the reviewed module. This release is available from [conda-forge](https://anaconda.org/conda-forge/climada/files). We also created a Zenodo archive: [10.5281/zenodo.12794729](https://doi.org/10.5281/zenodo.12794729)
