### 2024-07-23 7592d65
Remove attribute-style accesses from engine
