+++ climada/engine/impact_data.py
@@ -355,23 +355,23 @@ def create_lookup(emdat_data, start, end, disaster_subtype='Tropical cyclone'):
                                    'Date_start_EM_ordinal', 'Disaster_name',
                                    'EM_ID', 'ibtracsID', 'allocation_level',
                                    'possible_track', 'possible_track_all'])
-    lookup.hit_country = data.ISO
-    lookup.Date_start_EM = data.Date_start_clean
-    lookup.Disaster_name = data.Disaster_name
-    lookup.EM_ID = data.Disaster_No
+    lookup['hit_country'] = data['ISO']
+    lookup['Date_start_EM'] = data['Date_start_clean']
+    lookup['Disaster_name'] = data['Disaster_name']
+    lookup['EM_ID'] = data['Disaster_No']
     lookup = lookup.reset_index(drop=True)
     # create ordinals
-    for i in range(0, len(data.Date_start_clean.values)):
-        lookup.Date_start_EM_ordinal[i] = datetime.toordinal(
-            datetime.strptime(lookup.Date_start_EM.values[i], '%Y-%m-%d'))
+    for i in range(0, len(data['Date_start_clean'].values)):
+        lookup['Date_start_EM_ordinal'][i] = datetime.toordinal(
+            datetime.strptime(lookup['Date_start_EM'].values[i], '%Y-%m-%d'))
         # ordinals to numeric
-    lookup.Date_start_EM_ordinal = pd.to_numeric(lookup.Date_start_EM_ordinal)
+    lookup['Date_start_EM_ordinal'] = pd.to_numeric(lookup['Date_start_EM_ordinal'])
     # select time
     emdat_start = datetime.toordinal(datetime.strptime(start, '%Y-%m-%d'))
     emdat_end = datetime.toordinal(datetime.strptime(end, '%Y-%m-%d'))
 
-    lookup = lookup[lookup.Date_start_EM_ordinal.values > emdat_start]
-    lookup = lookup[lookup.Date_start_EM_ordinal.values < emdat_end]
+    lookup = lookup[lookup['Date_start_EM_ordinal'].values > emdat_start]
+    lookup = lookup[lookup['Date_start_EM_ordinal'].values < emdat_end]
 
     return lookup
 
@@ -397,15 +397,15 @@ def emdat_possible_hit(lookup, hit_countries, delta_t):
     # tracks: processed IBtracks with info which track hit which country
     # delta_t: time difference of start of EMdat and IBrtacks
     possible_hit_all = []
-    for i in range(0, len(lookup.EM_ID.values)):
+    for i in range(0, len(lookup['EM_ID'].values)):
         possible_hit = []
         country_tracks = hit_countries[
-            hit_countries['hit_country'] == lookup.hit_country.values[i]]
-        for j in range(0, len(country_tracks.Date_start.values)):
-            if (lookup.Date_start_EM_ordinal.values[i] - country_tracks.Date_start.values[j]) < \
-                delta_t and (lookup.Date_start_EM_ordinal.values[i] -
-                             country_tracks.Date_start.values[j]) >= 0:
-                possible_hit.append(country_tracks.ibtracsID.values[j])
+            hit_countries['hit_country'] == lookup['hit_country'].values[i]]
+        for j in range(0, len(country_tracks['Date_start'].values)):
+            if (lookup['Date_start_EM_ordinal'].values[i] - country_tracks['Date_start'].values[j]) < \
+                delta_t and (lookup['Date_start_EM_ordinal'].values[i] -
+                             country_tracks['Date_start'].values[j]) >= 0:
+                possible_hit.append(country_tracks['ibtracsID'].values[j])
         possible_hit_all.append(possible_hit)
 
[…]
