### 2024-07-23 5c29173
changed default settings, added old functions
