+++ climada/util/test/test_fit_methods.py
@@ -22,127 +22,110 @@ Test of fit_methods module
 import unittest
 import numpy as np
 
-from climada.util.fit_methods import calc_fit_interp, group_frequency
+from climada.util.fit_methods import interpolate_ev, group_frequency
 
 
 class TestFitMethods(unittest.TestCase):
     """Test different fit configurations"""
 
-    def test_calc_fit_interp_linear_fit(self):
-        """Test linear fit"""
-        x_train = np.array([1., 3., 5.])
-        y_train = np.array([2., 4., 6.])
-        x_test = np.array([0., 3., 4.])
-        np.testing.assert_allclose(
-            calc_fit_interp(x_test, x_train, y_train, method='fit'),
-            np.array([1., 4., 5.])
-        )
-
-    def test_calc_fit_interp_linear_interp(self):
+    def test_interpolate_ev_linear_interp(self):
         """Test linear interpolation"""
         x_train = np.array([1., 3., 5.])
         y_train = np.array([2., 4., 8.])
         x_test = np.array([0., 3., 4.])
         np.testing.assert_allclose(
-            calc_fit_interp(x_test, x_train, y_train, method='interp'),
-            np.array([2., 4., 6.])
+            interpolate_ev(x_test, x_train, y_train, method='interpolate', bounds_error=False),
+            np.array([np.nan, 4., 6.])
         )
         np.testing.assert_allclose(
-            calc_fit_interp(x_test, x_train, y_train, method='interp', left = -1.),
+            interpolate_ev(x_test, x_train, y_train, method='interpolate', fill_value = -1, bounds_error=False),
             np.array([-1., 4., 6.])
         )
 
-    def test_calc_fit_interp_threshold_parameters(self):
+    def test_interpolate_ev_threshold_parameters(self):
         """Test input threshold parameters"""
         x_train = np.array([0., 3., 6.])
         y_train = np.array([4., 1., 4.])
         x_test = np.array([-1., 3., 4.])
         np.testing.assert_allclose(
-            calc_fit_interp(x_test, x_train, y_train, method='fit'),
-            np.array([3., 3., 3.])
+            interpolate_ev(x_test, x_train, y_train, method='interpolate', bounds_error=False),
+            np.array([np.nan, 1., 2.])
         )
         np.testing.assert_allclose(
-            calc_fit_interp(x_test, x_train, y_train, method='fit', x_thres=1.),
-            np.array([-3., 1., 2.])
+            interpolate_ev(x_test, x_train, y_train, method='interpolate', x_threshold=1., bounds_error=False),
+            np.array([np.nan, 1., 2.])
         )
         np.testing.assert_allclose(
-            calc_fit_interp(x_test, x_train, y_train, method='fit', y_thres=2.),
-            np.array([4., 4., 4.])
+            interpolate_ev(x_test, x_train, y_train, method='interpolate', y_threshold=2., bounds_error=False),
+            np.array([np.nan, 4., 4.])
         )
     
-    def test_calc_fit_interp_scale_parameters(self):
+    def test_interpolate_ev_scale_parameters(self):
         """Test log scale parameters"""
         x_train = np.array([1e1, 1e3])
         y_train = np.array([1., 3.])
         x_test = np.array([1e0, 1e2])
         np.testing.assert_allclose(
-            calc_fit_interp(x_test, x_train, y_train, method='fit', x_scale='log'),
+            interpolate_ev(x_test, x_train, y_train, method='interpolate', x_scale='log', fill_value='extrapolate'),
             np.array([0., 2.])
         )
         np.testing.assert_allclose(
-            calc_fit_interp(x_test, x_train, y_train, method='interp', x_scale='log'),
-            np.array([1., 2.])
[…]
