### 2024-07-24 8767319
Updated CB plot and started on risk layer calc
