+++ climada/engine/option_appraisal/impact_metrics.py
@@ -18,6 +18,7 @@ with CLIMADA. If not, see <https://www.gnu.org/licenses/>.
 
 """
 from climada.entity.measures import MeasureSet
+from climada.util.value_representation import value_to_monetary_unit, ABBREV
 
 import pandas as pd
 from typing import Optional
@@ -28,8 +29,6 @@ from tabulate import tabulate
 
 import matplotlib.pyplot as plt
 import seaborn as sns
-
-
 
 
 class ImpactMetrics:
@@ -77,7 +76,33 @@ class ImpactMetrics:
     def print_CB_summary_table(self, metric = 'aai', start_year=None, end_year=None, consider_measure_times=True, risk_disc=None, cost_disc=None):
         # Get the CB analysis
         _, tot_CB_df = self.calc_CB(start_year, end_year, consider_measure_times, risk_disc, cost_disc)
-        print_CB_summary_table(tot_CB_df, self.measure_set, metric=metric, value_unit= self.value_unit)
+        # Get the measure_summary_df and total_summary
+        measure_summary_df, total_summary_dict, unit = generate_CB_plot_print_data(tot_CB_df, self.measure_set, metric, group=np.nan)
+        # Print the measure summary in tabulate format
+        # rename the  the Cost and Benefit headers by adding add ({value_unit} {unit})  to the headers
+        measure_summary_df.columns = ['Measure', f'Cost ({self.value_unit} {unit})', f'Benefit ({self.value_unit} {unit})', 'Benefit/Cost Ratio']
+        print(tabulate(measure_summary_df, headers='keys', tablefmt='fancy_grid'))
+        # Print the total summary
+        print("\n--------------------  ---------")
+        for key, value in total_summary_dict.items():
+            print(f"{key:<22} {value:>11.2f}  ({self.value_unit} {unit})")
+        print("--------------------  ---------")
+        # Create a new line for explanatory text
+        label_Benefit = f"{'Discounted ' if risk_disc else ''} Total averted annual risk metric: {metric}"
+        label_Cost = f"{'Discounted ' if cost_disc else ''} Total annual cost"
+        # Print explanatory text
+        print("Explanatory Notes:")
+        print(f"Benefit: {label_Benefit} – This represents the total benefits in terms of risk reduction.")
+        print(f"Cost: {label_Cost} – This represents the total costs associated with implementing the measures.")
+        return
+    
+    def plot_CB_summary_table(self, metric = 'aai', start_year=None, end_year=None, consider_measure_times=True, risk_disc=None, cost_disc=None):
+        # Get the CB analysis
+        _, tot_CB_df = self.calc_CB(start_year, end_year, consider_measure_times, risk_disc, cost_disc)
+        # Get the measure_summary_df and total_summary
+        measure_summary_df, total_summary_dict, unit = generate_CB_plot_print_data(tot_CB_df, self.measure_set, metric, group=np.nan)
+        # Plot the measure summary
+        plot_CB_summary(measure_summary_df, total_summary_dict, measure_colors=None, y_label=f'Averted Risk {unit}', title=f'Benefit (NPV of total annual {metric}) and Benefit/Cost Ratio by Measure')
         return
     
     # Plot the yearly averted cost for a specific measure
@@ -200,6 +225,85 @@ class ImpactMetrics:
 
 #%% Utility functions
 
+def plot_CB_summary(measure_summary_df, total_summary_dict, measure_colors=None, y_label='Risk (Bn)', title='Benefit and Benefit/Cost Ratio by Measure'):
[…]
