+++ climada/hazard/plot.py
@@ -19,11 +19,13 @@ with CLIMADA. If not, see <https://www.gnu.org/licenses/>.
 Define Hazard Plotting Methods.
 """
 
+import logging
 import numpy as np
 import matplotlib.pyplot as plt
 
 import climada.util.plot as u_plot
 
+LOGGER = logging.getLogger(__name__)
 
 # pylint: disable=no-member
 
@@ -36,6 +38,10 @@ class HazardPlot():
     def plot_rp_intensity(self, return_periods=(25, 50, 100, 250),
                           smooth=True, axis=None, figsize=(9, 13), adapt_fontsize=True,
                           **kwargs):
+        """This function is deprecated, use Impact.local_exceedance_impact and util.plot.plot_from_gdf instead."""
+        LOGGER.warning("The use of Hazard.plot_rp_intensity is deprecated."
+                       "Use Hazard.local_exceedance_intensity and util.plot.plot_from_gdf instead.")
+
         """Compute and plot hazard exceedance intensity maps for different
         return periods. Calls local_exceedance_inten.
 
@@ -57,7 +63,7 @@ class HazardPlot():
         axis, inten_stats:  matplotlib.axes._subplots.AxesSubplot, np.ndarray
             intenstats is return_periods.size x num_centroids
         """
-        inten_stats, _, _ = self.local_exceedance_inten(np.array(return_periods))
+        inten_stats = self.local_exceedance_intensity(return_periods)[0].values[:,1:].T.astype(float)
         colbar_name = 'Intensity (' + self.units + ')'
         title = list()
         for ret in return_periods:
