### 2024-07-24 adfc6e0
new asymptotic and small-input behaviour of funcs
