+++ climada/util/test/test_fit_methods.py
@@ -95,11 +95,15 @@ class TestFitMethods(unittest.TestCase):
     def test_interpolate_ev_stepfunction(self):
         """Test stepfunction method"""
         x_train = np.array([1., 3., 5.])
-        y_train = np.array([2., 4., 8.])
+        y_train = np.array([8., 4., 2.])
         x_test = np.array([0., 3., 4., 6.])
         np.testing.assert_allclose(
             interpolate_ev(x_test, x_train, y_train, method='stepfunction'),
-            np.array([2., 4., 8., np.nan])
+            np.array([8., 4., 2., np.nan])
+        )
+        np.testing.assert_allclose(
+            interpolate_ev(x_test, x_train, y_train, method='stepfunction', y_asymptotic=0.),
+            np.array([8., 4., 2., 0.])
         )
 
     def test_interpolate_ev_degenerate_input(self):
@@ -108,18 +112,33 @@ class TestFitMethods(unittest.TestCase):
         x_test = np.array([0., 2., 4.])
         y_train = np.zeros(3)
         np.testing.assert_allclose(
-            interpolate_ev(x_test, x_train, y_train, method='interpolate', bounds_error=False),
+            interpolate_ev(x_test, x_train, y_train, bounds_error=False),
             np.array([np.nan, 0., 0.])
         )
 
-    def test_interpolate_ev(self):
+    def test_interpolate_ev_small_input(self):
         """Test small input"""
         x_train = np.array([1.])
         y_train = np.array([2.])
-        x_test = np.array([0., 1.])
+        x_test = np.array([0., 1., 2.])
         np.testing.assert_allclose(
-            interpolate_ev(x_test, x_train, y_train, method='interpolate'),
-            np.array([2., 2.])
+            interpolate_ev(x_test, x_train, y_train),
+            np.array([2., 2., np.nan])
+        )
+        np.testing.assert_allclose(
+            interpolate_ev(x_test, x_train, y_train, y_asymptotic=0),
+            np.array([2., 2., 0.])
+        )
+        x_train = np.array([])
+        y_train = np.array([])
+        x_test = np.array([0., 1., 2.])
+        np.testing.assert_allclose(
+            interpolate_ev(x_test, x_train, y_train),
+            np.full(3, np.nan)
+        )
+        np.testing.assert_allclose(
+            interpolate_ev(x_test, x_train, y_train, y_asymptotic=0),
+            np.zeros(3)
         )
     
     def test_frequency_group(self):
