+++ climada/util/fit_methods.py
@@ -37,6 +37,7 @@ def interpolate_ev(
         y_scale: str=None, 
         x_threshold: float = None, 
         y_threshold: float = None,
+        y_asymptotic = np.nan,
         **kwargs
     ):
     """_summary_
@@ -50,6 +51,8 @@ def interpolate_ev(
         y_scale (str, optional): _description_. Defaults to None.
         x_threshold (float, optional): _description_. Defaults to None.
         y_threshold (float, optional): _description_. Defaults to None.
+        y_asymptotic (float, optional): return value if x_test > x_train if 
+            method is stepfunction or x_train.size < 2. Defaults to np.nan.
         **kwargs: further optional parameters for the fit methods
 
     Returns
@@ -66,7 +69,10 @@ def interpolate_ev(
         x_scale, y_scale = None, None
     if x_train.shape != y_train.shape:
         raise ValueError(f'Incompatible shapes of input data, x_train {x_train.shape} and y_train {y_train.shape}. Should be the same')
-    
+
+    # transform input to float arrays
+    x_test, x_train, y_train = np.array(x_test).astype(float), np.array(x_train).astype(float), np.array(y_train).astype(float)
+
     # cut x and y above threshold
     if x_threshold or x_threshold==0:
         x_th = np.asarray(x_train > x_threshold).squeeze()
@@ -78,12 +84,15 @@ def interpolate_ev(
         x_train = x_train[y_th]
         y_train = y_train[y_th]
 
-    # return zeros if x_train and y_train empty
+    # return y_asymptotic if x_train and y_train empty
     if x_train.size == 0:
-        return np.zeros_like(x_test)
-    # return y_train if only one (x_train, y_train) to fit
+        return np.full_like(x_test, y_asymptotic)
+    # if only one (x_train, y_train), return stepfunction with
+    # y_train if x_test < x_train and y_asymtotic if x_test > x_train
     if x_train.size == 1:
-        return np.full_like(x_test, y_train[0])
+        y_test = np.full_like(x_test, y_train[0])
+        y_test[np.squeeze(x_test) > np.squeeze(x_train)] = y_asymptotic
+        return y_test
 
     # adapt x and y scale
     if x_scale == 'log':
@@ -93,10 +102,16 @@ def interpolate_ev(
 
     # calculate interpolation
     if method == 'interpolate':
+        # warn if data is being extrapolated
         if (
             (('fill_value', 'extrapolate') in kwargs.items()) and 
             ((np.min(x_test) < np.min(x_train)) or (np.max(x_test) > np.max(x_train)))):
             LOGGER.warning('Data is being extrapolated.')
+        # calculate fill values 
+        elif 'fill_value' in kwargs.keys():
+            if len(kwargs['fill_value']) == 2:
+                if kwargs['fill_value'][0] == 'maximum':
+                    kwargs['fill_value'] = (np.max(y_train), kwargs['fill_value'][1])
         interpolation = interpolate.interp1d(x_train, y_train, **kwargs)
         y_test = interpolation(x_test)
     
@@ -105,9 +120,9 @@ def interpolate_ev(
         # find indeces of x_test if sorted into x_train
         if not all(sorted(x_train) == x_train):
             raise ValueError(f'Input array x_train must be sorted in ascending order.')
-        indx = np.searchsorted(x_train, x_test, **kwargs)
+        indx = np.searchsorted(x_train, x_test)
         y_test = y_train[indx.clip(max = len(x_train) - 1)]
-        y_test[indx == len(x_train)] = np.nan
+        y_test[indx == len(x_train)] = y_asymptotic
 
     # adapt output scale
     if y_scale == 'log':
