### 2024-07-25 #6755 @observingClouds
Thanks @peanutfun for the archives and releases. I just have one minor comment. Could you put the specific DOI that points to Climada v5.0.0 (which doesn't seem to have been stored on zenodo yet) to the related work section of the JOSS release archive (https://zenodo.org/records/12794730)?
