### 2024-07-25 #6755 @observingClouds
@peanutfun with the merge of this feature branch to main, I think the branch with the paper is gone (calibrate-impact-functions). Could you restore it for the time being or name a different branch (main?)
