### 2024-07-25 #6755 @peanutfun
> Could you put the specific DOI that points to Climada v5.0.0 (which doesn't seem to have been stored on zenodo yet) to the related work section of the JOSS release archive (https://zenodo.org/records/12794730)?

@observingClouds The archive we created ([10.5281/zenodo.12794729](https://doi.org/10.5281/zenodo.12794729)) IS Climada v5.0.0. The original record [10.5281/zenodo.4598943](https://zenodo.org/doi/10.5281/zenodo.4598943) is managed by an automated Zenodo webhook, which unfortunately failed for the v5.0.0 release. We are not sure how to update the record without making another GitHub release after updating the webhook. Since the original record has a different author list than this publication, we decided for a separate record to comply with the author tasks after review. The new archive lists the original Zenodo record of Climada under "Related Works" as "Is version of".

Please let me know how to proceed.
