### 2024-07-25 #6755 @observingClouds
@peanutfun Just to clarify, [10.5281/zenodo.12794729](https://doi.org/10.5281/zenodo.12794729) and the release that should have been done with the GitHub webhook would have been identical from the code perspective and just differ from the metadata?
