### 2024-07-25 ff1bead
Update citation.rst
Add final OSM publication link
