### 2024-07-26 e2f2ead
Update grid filename for icon-eu-eps model
