### 2024-07-26 #6755 @observingClouds
@naingeet congrats on your first review for JOSS :tada: I strongly encourage you to [sign up as a reviewer](https://reviewers.joss.theoj.org/) and join our community.
