### 2024-07-26 #6755 @naingeet
It was a great experience. Thank you @observingClouds
<https://github.com/observingClouds>  for the opportunity! I signed up as a
reviewer , looking forward to reviewing more software and papers in the
geophysical and climate science field. :)
regards,
-Geeta Nain

On Fri, Jul 26, 2024 at 10:11 AM Hauke Schulz ***@***.***>
wrote:

> @naingeet <https://github.com/naingeet> congrats on your first review for
> JOSS 🎉 I strongly encourage you to sign up as a reviewer
> <https://reviewers.joss.theoj.org/> and join our community.
>
